+++ js/app-built.js
@@ -169,5 +169,5 @@ goog.exportProperty(ol.source.OSM.prototype,"changed",ol.source.OSM.prototype.ch
 goog.exportProperty(ol.renderer.webgl.VectorLayer.prototype,"dispatchEvent",ol.renderer.webgl.VectorLayer.prototype.dispatchEvent),goog.exportProperty(ol.renderer.webgl.VectorLayer.prototype,"getRevision",ol.renderer.webgl.VectorLayer.prototype.getRevision),goog.exportProperty(ol.renderer.webgl.VectorLayer.prototype,"on",ol.renderer.webgl.VectorLayer.prototype.on),goog.exportProperty(ol.renderer.webgl.VectorLayer.prototype,"once",ol.renderer.webgl.VectorLayer.prototype.once),goog.exportProperty(ol.renderer.webgl.VectorLayer.prototype,"un",ol.renderer.webgl.VectorLayer.prototype.un),goog.exportProperty(ol.renderer.webgl.VectorLayer.prototype,"unByKey",ol.renderer.webgl.VectorLayer.prototype.unByKey),goog.exportProperty(ol.renderer.canvas.Layer.prototype,"changed",ol.renderer.canvas.Layer.prototype.changed),goog.exportProperty(ol.renderer.canvas.Layer.prototype,"dispatchEvent",ol.renderer.canvas.Layer.prototype.dispatchEvent),goog.exportProperty(ol.renderer.canvas.Layer.prototype,"getRevision",ol.renderer.canvas.Layer.prototype.getRevision),goog.exportProperty(ol.renderer.canvas.Layer.prototype,"on",ol.renderer.canvas.Layer.prototype.on),goog.exportProperty(ol.renderer.canvas.Layer.prototype,"once",ol.renderer.canvas.Layer.prototype.once),goog.exportProperty(ol.renderer.canvas.Layer.prototype,"un",ol.renderer.canvas.Layer.prototype.un),goog.exportProperty(ol.renderer.canvas.Layer.prototype,"unByKey",ol.renderer.canvas.Layer.prototype.unByKey),goog.exportProperty(ol.renderer.canvas.ImageLayer.prototype,"changed",ol.renderer.canvas.ImageLayer.prototype.changed),goog.exportProperty(ol.renderer.canvas.ImageLayer.prototype,"dispatchEvent",ol.renderer.canvas.ImageLayer.prototype.dispatchEvent),goog.exportProperty(ol.renderer.canvas.ImageLayer.prototype,"getRevision",ol.renderer.canvas.ImageLayer.prototype.getRevision),goog.exportProperty(ol.renderer.canvas.ImageLayer.prototype,"on",ol.renderer.canvas.ImageLayer.prototype.on),goog.exportProperty(ol.renderer.canvas.ImageLayer.prototype,"once",ol.renderer.canvas.ImageLayer.prototype.once),goog.exportProperty(ol.renderer.canvas.ImageLayer.prototype,"un",ol.renderer.canvas.ImageLayer.prototype.un),goog.exportProperty(ol.renderer.canvas.ImageLayer.prototype,"unByKey",ol.renderer.canvas.ImageLayer.prototype.unByKey),goog.exportProperty(ol.renderer.canvas.TileLayer.prototype,"changed",ol.renderer.canvas.TileLayer.prototype.changed),goog.exportProperty(ol.renderer.canvas.TileLayer.prototype,"dispatchEvent",ol.renderer.canvas.TileLayer.prototype.dispatchEvent),goog.exportProperty(ol.renderer.canvas.TileLayer.prototype,"getRevision",ol.renderer.canvas.TileLayer.prototype.getRevision),goog.exportProperty(ol.renderer.canvas.TileLayer.prototype,"on",ol.renderer.canvas.TileLayer.prototype.on),goog.exportProperty(ol.renderer.canvas.TileLayer.prototype,"once",ol.renderer.canvas.TileLayer.prototype.once),goog.exportProperty(ol.renderer.canvas.TileLayer.prototype,"un",ol.renderer.canvas.TileLayer.prototype.un),goog.exportProperty(ol.renderer.canvas.TileLayer.prototype,"unByKey",ol.renderer.canvas.TileLayer.prototype.unByKey),goog.exportProperty(ol.renderer.canvas.VectorLayer.prototype,"changed",ol.renderer.canvas.VectorLayer.prototype.changed),goog.exportProperty(ol.renderer.canvas.VectorLayer.prototype,"dispatchEvent",ol.renderer.canvas.VectorLayer.prototype.dispatchEvent),goog.exportProperty(ol.renderer.canvas.VectorLayer.prototype,"getRevision",ol.renderer.canvas.VectorLayer.prototype.getRevision),goog.exportProperty(ol.renderer.canvas.VectorLayer.prototype,"on",ol.renderer.canvas.VectorLayer.prototype.on),goog.exportProperty(ol.renderer.canvas.VectorLayer.prototype,"once",ol.renderer.canvas.VectorLayer.prototype.once),goog.exportProperty(ol.renderer.canvas.VectorLayer.prototype,"un",ol.renderer.canvas.VectorLayer.prototype.un),goog.exportProperty(ol.renderer.canvas.VectorLayer.prototype,"unByKey",ol.renderer.canvas.VectorLayer.prototype.unByKey),goog.exportProperty(ol.renderer.canvas.VectorTileLayer.prototype,"changed",ol.renderer.canvas.VectorTileLayer.prototype.changed),goog.exportProperty(ol.renderer.canvas.VectorTileLayer.prototype,"dispatchEvent",ol.renderer.canvas.VectorTileLayer.prototype.dispatchEvent),goog.exportProperty(ol.renderer.canvas.VectorTileLayer.prototype,"getRevision",ol.renderer.canvas.VectorTileLayer.prototype.getRevision),goog.exportProperty(ol.renderer.canvas.VectorTileLayer.prototype,"on",ol.renderer.canvas.VectorTileLayer.prototype.on),goog.exportProperty(ol.renderer.canvas.VectorTileLayer.prototype,"once",ol.renderer.canvas.VectorTileLayer.prototype.once),goog.exportProperty(ol.renderer.canvas.VectorTileLayer.prototype,"un",ol.renderer.canvas.VectorTileLayer.prototype.un),goog.exportProperty(ol.renderer.canvas.VectorTileLayer.prototype,"unByKey",ol.renderer.canvas.VectorTileLayer.prototype.unByKey),goog.exportProperty(ol.render.Event.prototype,"type",ol.render.Event.prototype.type),goog.exportProperty(ol.render.Event.prototype,"target",ol.render.Event.prototype.target),goog.exportProperty(ol.render.Event.prototype,"preventDefault",ol.render.Event.prototype.preventDefault),goog.exportProperty(ol.render.Event.prototype,"stopPropagation",ol.render.Event.prototype.stopPropagation),goog.exportProperty(ol.pointer.PointerEvent.prototype,"type",ol.pointer.PointerEvent.prototype.type),goog.exportProperty(ol.pointer.PointerEvent.prototype,"target",ol.pointer.PointerEvent.prototype.target),goog.exportProperty(ol.pointer.PointerEvent.prototype,"preventDefault",ol.pointer.PointerEvent.prototype.preventDefault),goog.exportProperty(ol.pointer.PointerEvent.prototype,"stopPropagation",ol.pointer.PointerEvent.prototype.stopPropagation),goog.exportProperty(ol.layer.Base.prototype,"get",ol.layer.Base.prototype.get),goog.exportProperty(ol.layer.Base.prototype,"getKeys",ol.layer.Base.prototype.getKeys),goog.exportProperty(ol.layer.Base.prototype,"getProperties",ol.layer.Base.prototype.getProperties),goog.exportProperty(ol.layer.Base.prototype,"set",ol.layer.Base.prototype.set),goog.exportProperty(ol.layer.Base.prototype,"setProperties",ol.layer.Base.prototype.setProperties),goog.exportProperty(ol.layer.Base.prototype,"unset",ol.layer.Base.prototype.unset),goog.exportProperty(ol.layer.Base.prototype,"changed",ol.layer.Base.prototype.changed),goog.exportProperty(ol.layer.Base.prototype,"dispatchEvent",ol.layer.Base.prototype.dispatchEvent),goog.exportProperty(ol.layer.Base.prototype,"getRevision",ol.layer.Base.prototype.getRevision),goog.exportProperty(ol.layer.Base.prototype,"on",ol.layer.Base.prototype.on),goog.exportProperty(ol.layer.Base.prototype,"once",ol.layer.Base.prototype.once),goog.exportProperty(ol.layer.Base.prototype,"un",ol.layer.Base.prototype.un),goog.exportProperty(ol.layer.Base.prototype,"unByKey",ol.layer.Base.prototype.unByKey),goog.exportProperty(ol.layer.Group.prototype,"getExtent",ol.layer.Group.prototype.getExtent),goog.exportProperty(ol.layer.Group.prototype,"getMaxResolution",ol.layer.Group.prototype.getMaxResolution),goog.exportProperty(ol.layer.Group.prototype,"getMinResolution",ol.layer.Group.prototype.getMinResolution),goog.exportProperty(ol.layer.Group.prototype,"getOpacity",ol.layer.Group.prototype.getOpacity),goog.exportProperty(ol.layer.Group.prototype,"getVisible",ol.layer.Group.prototype.getVisible),goog.exportProperty(ol.layer.Group.prototype,"getZIndex",ol.layer.Group.prototype.getZIndex),goog.exportProperty(ol.layer.Group.prototype,"setExtent",ol.layer.Group.prototype.setExtent),goog.exportProperty(ol.layer.Group.prototype,"setMaxResolution",ol.layer.Group.prototype.setMaxResolution),goog.exportProperty(ol.layer.Group.prototype,"setMinResolution",ol.layer.Group.prototype.setMinResolution),goog.exportProperty(ol.layer.Group.prototype,"setOpacity",ol.layer.Group.prototype.setOpacity),goog.exportProperty(ol.layer.Group.prototype,"setVisible",ol.layer.Group.prototype.setVisible),goog.exportProperty(ol.layer.Group.prototype,"setZIndex",ol.layer.Group.prototype.setZIndex),goog.exportProperty(ol.layer.Group.prototype,"get",ol.layer.Group.prototype.get),goog.exportProperty(ol.layer.Group.prototype,"getKeys",ol.layer.Group.prototype.getKeys),goog.exportProperty(ol.layer.Group.prototype,"getProperties",ol.layer.Group.prototype.getProperties),goog.exportProperty(ol.layer.Group.prototype,"set",ol.layer.Group.prototype.set),goog.exportProperty(ol.layer.Group.prototype,"setProperties",ol.layer.Group.prototype.setProperties),goog.exportProperty(ol.layer.Group.prototype,"unset",ol.layer.Group.prototype.unset),goog.exportProperty(ol.layer.Group.prototype,"changed",ol.layer.Group.prototype.changed),goog.exportProperty(ol.layer.Group.prototype,"dispatchEvent",ol.layer.Group.prototype.dispatchEvent),goog.exportProperty(ol.layer.Group.prototype,"getRevision",ol.layer.Group.prototype.getRevision),goog.exportProperty(ol.layer.Group.prototype,"on",ol.layer.Group.prototype.on),goog.exportProperty(ol.layer.Group.prototype,"once",ol.layer.Group.prototype.once),goog.exportProperty(ol.layer.Group.prototype,"un",ol.layer.Group.prototype.un),goog.exportProperty(ol.layer.Group.prototype,"unByKey",ol.layer.Group.prototype.unByKey),goog.exportProperty(ol.layer.Layer.prototype,"getExtent",ol.layer.Layer.prototype.getExtent),goog.exportProperty(ol.layer.Layer.prototype,"getMaxResolution",ol.layer.Layer.prototype.getMaxResolution),goog.exportProperty(ol.layer.Layer.prototype,"getMinResolution",ol.layer.Layer.prototype.getMinResolution),goog.exportProperty(ol.layer.Layer.prototype,"getOpacity",ol.layer.Layer.prototype.getOpacity),goog.exportProperty(ol.layer.Layer.prototype,"getVisible",ol.layer.Layer.prototype.getVisible),goog.exportProperty(ol.layer.Layer.prototype,"getZIndex",ol.layer.Layer.prototype.getZIndex),goog.exportProperty(ol.layer.Layer.prototype,"setExtent",ol.layer.Layer.prototype.setExtent),goog.exportProperty(ol.layer.Layer.prototype,"setMaxResolution",ol.layer.Layer.prototype.setMaxResolution),goog.exportProperty(ol.layer.Layer.prototype,"setMinResolution",ol.layer.Layer.prototype.setMinResolution),goog.exportProperty(ol.layer.Layer.prototype,"setOpacity",ol.layer.Layer.prototype.setOpacity),goog.exportProperty(ol.layer.Layer.prototype,"setVisible",ol.layer.Layer.prototype.setVisible),goog.exportProperty(ol.layer.Layer.prototype,"setZIndex",ol.layer.Layer.prototype.setZIndex),goog.exportProperty(ol.layer.Layer.prototype,"get",ol.layer.Layer.prototype.get),goog.exportProperty(ol.layer.Layer.prototype,"getKeys",ol.layer.Layer.prototype.getKeys),goog.exportProperty(ol.layer.Layer.prototype,"getProperties",ol.layer.Layer.prototype.getProperties),goog.exportProperty(ol.layer.Layer.prototype,"set",ol.layer.Layer.prototype.set),goog.exportProperty(ol.layer.Layer.prototype,"setProperties",ol.layer.Layer.prototype.setProperties),goog.exportProperty(ol.layer.Layer.prototype,"unset",ol.layer.Layer.prototype.unset),goog.exportProperty(ol.layer.Layer.prototype,"changed",ol.layer.Layer.prototype.changed),goog.exportProperty(ol.layer.Layer.prototype,"dispatchEvent",ol.layer.Layer.prototype.dispatchEvent),goog.exportProperty(ol.layer.Layer.prototype,"getRevision",ol.layer.Layer.prototype.getRevision),goog.exportProperty(ol.layer.Layer.prototype,"on",ol.layer.Layer.prototype.on),goog.exportProperty(ol.layer.Layer.prototype,"once",ol.layer.Layer.prototype.once),goog.exportProperty(ol.layer.Layer.prototype,"un",ol.layer.Layer.prototype.un),goog.exportProperty(ol.layer.Layer.prototype,"unByKey",ol.layer.Layer.prototype.unByKey),goog.exportProperty(ol.layer.Vector.prototype,"setMap",ol.layer.Vector.prototype.setMap),goog.exportProperty(ol.layer.Vector.prototype,"setSource",ol.layer.Vector.prototype.setSource),goog.exportProperty(ol.layer.Vector.prototype,"getExtent",ol.layer.Vector.prototype.getExtent),goog.exportProperty(ol.layer.Vector.prototype,"getMaxResolution",ol.layer.Vector.prototype.getMaxResolution),goog.exportProperty(ol.layer.Vector.prototype,"getMinResolution",ol.layer.Vector.prototype.getMinResolution),goog.exportProperty(ol.layer.Vector.prototype,"getOpacity",ol.layer.Vector.prototype.getOpacity),goog.exportProperty(ol.layer.Vector.prototype,"getVisible",ol.layer.Vector.prototype.getVisible),goog.exportProperty(ol.layer.Vector.prototype,"getZIndex",ol.layer.Vector.prototype.getZIndex),goog.exportProperty(ol.layer.Vector.prototype,"setExtent",ol.layer.Vector.prototype.setExtent),goog.exportProperty(ol.layer.Vector.prototype,"setMaxResolution",ol.layer.Vector.prototype.setMaxResolution),goog.exportProperty(ol.layer.Vector.prototype,"setMinResolution",ol.layer.Vector.prototype.setMinResolution),goog.exportProperty(ol.layer.Vector.prototype,"setOpacity",ol.layer.Vector.prototype.setOpacity),goog.exportProperty(ol.layer.Vector.prototype,"setVisible",ol.layer.Vector.prototype.setVisible),goog.exportProperty(ol.layer.Vector.prototype,"setZIndex",ol.layer.Vector.prototype.setZIndex),goog.exportProperty(ol.layer.Vector.prototype,"get",ol.layer.Vector.prototype.get),goog.exportProperty(ol.layer.Vector.prototype,"getKeys",ol.layer.Vector.prototype.getKeys),goog.exportProperty(ol.layer.Vector.prototype,"getProperties",ol.layer.Vector.prototype.getProperties),goog.exportProperty(ol.layer.Vector.prototype,"set",ol.layer.Vector.prototype.set),goog.exportProperty(ol.layer.Vector.prototype,"setProperties",ol.layer.Vector.prototype.setProperties),goog.exportProperty(ol.layer.Vector.prototype,"unset",ol.layer.Vector.prototype.unset),goog.exportProperty(ol.layer.Vector.prototype,"changed",ol.layer.Vector.prototype.changed),goog.exportProperty(ol.layer.Vector.prototype,"dispatchEvent",ol.layer.Vector.prototype.dispatchEvent),goog.exportProperty(ol.layer.Vector.prototype,"getRevision",ol.layer.Vector.prototype.getRevision),goog.exportProperty(ol.layer.Vector.prototype,"on",ol.layer.Vector.prototype.on),goog.exportProperty(ol.layer.Vector.prototype,"once",ol.layer.Vector.prototype.once),goog.exportProperty(ol.layer.Vector.prototype,"un",ol.layer.Vector.prototype.un),goog.exportProperty(ol.layer.Vector.prototype,"unByKey",ol.layer.Vector.prototype.unByKey),goog.exportProperty(ol.layer.Heatmap.prototype,"getSource",ol.layer.Heatmap.prototype.getSource),goog.exportProperty(ol.layer.Heatmap.prototype,"getStyle",ol.layer.Heatmap.prototype.getStyle),goog.exportProperty(ol.layer.Heatmap.prototype,"getStyleFunction",ol.layer.Heatmap.prototype.getStyleFunction),goog.exportProperty(ol.layer.Heatmap.prototype,"setStyle",ol.layer.Heatmap.prototype.setStyle),goog.exportProperty(ol.layer.Heatmap.prototype,"setMap",ol.layer.Heatmap.prototype.setMap),goog.exportProperty(ol.layer.Heatmap.prototype,"setSource",ol.layer.Heatmap.prototype.setSource),goog.exportProperty(ol.layer.Heatmap.prototype,"getExtent",ol.layer.Heatmap.prototype.getExtent),goog.exportProperty(ol.layer.Heatmap.prototype,"getMaxResolution",ol.layer.Heatmap.prototype.getMaxResolution),goog.exportProperty(ol.layer.Heatmap.prototype,"getMinResolution",ol.layer.Heatmap.prototype.getMinResolution),goog.exportProperty(ol.layer.Heatmap.prototype,"getOpacity",ol.layer.Heatmap.prototype.getOpacity),goog.exportProperty(ol.layer.Heatmap.prototype,"getVisible",ol.layer.Heatmap.prototype.getVisible),goog.exportProperty(ol.layer.Heatmap.prototype,"getZIndex",ol.layer.Heatmap.prototype.getZIndex),goog.exportProperty(ol.layer.Heatmap.prototype,"setExtent",ol.layer.Heatmap.prototype.setExtent),goog.exportProperty(ol.layer.Heatmap.prototype,"setMaxResolution",ol.layer.Heatmap.prototype.setMaxResolution),goog.exportProperty(ol.layer.Heatmap.prototype,"setMinResolution",ol.layer.Heatmap.prototype.setMinResolution),goog.exportProperty(ol.layer.Heatmap.prototype,"setOpacity",ol.layer.Heatmap.prototype.setOpacity),goog.exportProperty(ol.layer.Heatmap.prototype,"setVisible",ol.layer.Heatmap.prototype.setVisible),goog.exportProperty(ol.layer.Heatmap.prototype,"setZIndex",ol.layer.Heatmap.prototype.setZIndex),goog.exportProperty(ol.layer.Heatmap.prototype,"get",ol.layer.Heatmap.prototype.get),goog.exportProperty(ol.layer.Heatmap.prototype,"getKeys",ol.layer.Heatmap.prototype.getKeys),goog.exportProperty(ol.layer.Heatmap.prototype,"getProperties",ol.layer.Heatmap.prototype.getProperties),goog.exportProperty(ol.layer.Heatmap.prototype,"set",ol.layer.Heatmap.prototype.set),goog.exportProperty(ol.layer.Heatmap.prototype,"setProperties",ol.layer.Heatmap.prototype.setProperties),goog.exportProperty(ol.layer.Heatmap.prototype,"unset",ol.layer.Heatmap.prototype.unset),goog.exportProperty(ol.layer.Heatmap.prototype,"changed",ol.layer.Heatmap.prototype.changed),goog.exportProperty(ol.layer.Heatmap.prototype,"dispatchEvent",ol.layer.Heatmap.prototype.dispatchEvent),goog.exportProperty(ol.layer.Heatmap.prototype,"getRevision",ol.layer.Heatmap.prototype.getRevision),goog.exportProperty(ol.layer.Heatmap.prototype,"on",ol.layer.Heatmap.prototype.on),goog.exportProperty(ol.layer.Heatmap.prototype,"once",ol.layer.Heatmap.prototype.once),goog.exportProperty(ol.layer.Heatmap.prototype,"un",ol.layer.Heatmap.prototype.un),goog.exportProperty(ol.layer.Heatmap.prototype,"unByKey",ol.layer.Heatmap.prototype.unByKey),goog.exportProperty(ol.layer.Image.prototype,"setMap",ol.layer.Image.prototype.setMap),goog.exportProperty(ol.layer.Image.prototype,"setSource",ol.layer.Image.prototype.setSource),goog.exportProperty(ol.layer.Image.prototype,"getExtent",ol.layer.Image.prototype.getExtent),goog.exportProperty(ol.layer.Image.prototype,"getMaxResolution",ol.layer.Image.prototype.getMaxResolution),goog.exportProperty(ol.layer.Image.prototype,"getMinResolution",ol.layer.Image.prototype.getMinResolution),goog.exportProperty(ol.layer.Image.prototype,"getOpacity",ol.layer.Image.prototype.getOpacity),goog.exportProperty(ol.layer.Image.prototype,"getVisible",ol.layer.Image.prototype.getVisible),goog.exportProperty(ol.layer.Image.prototype,"getZIndex",ol.layer.Image.prototype.getZIndex),goog.exportProperty(ol.layer.Image.prototype,"setExtent",ol.layer.Image.prototype.setExtent),goog.exportProperty(ol.layer.Image.prototype,"setMaxResolution",ol.layer.Image.prototype.setMaxResolution),goog.exportProperty(ol.layer.Image.prototype,"setMinResolution",ol.layer.Image.prototype.setMinResolution),goog.exportProperty(ol.layer.Image.prototype,"setOpacity",ol.layer.Image.prototype.setOpacity),goog.exportProperty(ol.layer.Image.prototype,"setVisible",ol.layer.Image.prototype.setVisible),goog.exportProperty(ol.layer.Image.prototype,"setZIndex",ol.layer.Image.prototype.setZIndex),goog.exportProperty(ol.layer.Image.prototype,"get",ol.layer.Image.prototype.get),goog.exportProperty(ol.layer.Image.prototype,"getKeys",ol.layer.Image.prototype.getKeys),goog.exportProperty(ol.layer.Image.prototype,"getProperties",ol.layer.Image.prototype.getProperties),goog.exportProperty(ol.layer.Image.prototype,"set",ol.layer.Image.prototype.set),goog.exportProperty(ol.layer.Image.prototype,"setProperties",ol.layer.Image.prototype.setProperties),goog.exportProperty(ol.layer.Image.prototype,"unset",ol.layer.Image.prototype.unset),goog.exportProperty(ol.layer.Image.prototype,"changed",ol.layer.Image.prototype.changed),goog.exportProperty(ol.layer.Image.prototype,"dispatchEvent",ol.layer.Image.prototype.dispatchEvent),goog.exportProperty(ol.layer.Image.prototype,"getRevision",ol.layer.Image.prototype.getRevision),goog.exportProperty(ol.layer.Image.prototype,"on",ol.layer.Image.prototype.on),goog.exportProperty(ol.layer.Image.prototype,"once",ol.layer.Image.prototype.once),goog.exportProperty(ol.layer.Image.prototype,"un",ol.layer.Image.prototype.un),goog.exportProperty(ol.layer.Image.prototype,"unByKey",ol.layer.Image.prototype.unByKey),goog.exportProperty(ol.layer.Tile.prototype,"setMap",ol.layer.Tile.prototype.setMap),goog.exportProperty(ol.layer.Tile.prototype,"setSource",ol.layer.Tile.prototype.setSource),goog.exportProperty(ol.layer.Tile.prototype,"getExtent",ol.layer.Tile.prototype.getExtent),goog.exportProperty(ol.layer.Tile.prototype,"getMaxResolution",ol.layer.Tile.prototype.getMaxResolution),goog.exportProperty(ol.layer.Tile.prototype,"getMinResolution",ol.layer.Tile.prototype.getMinResolution),goog.exportProperty(ol.layer.Tile.prototype,"getOpacity",ol.layer.Tile.prototype.getOpacity),goog.exportProperty(ol.layer.Tile.prototype,"getVisible",ol.layer.Tile.prototype.getVisible),goog.exportProperty(ol.layer.Tile.prototype,"getZIndex",ol.layer.Tile.prototype.getZIndex),goog.exportProperty(ol.layer.Tile.prototype,"setExtent",ol.layer.Tile.prototype.setExtent),goog.exportProperty(ol.layer.Tile.prototype,"setMaxResolution",ol.layer.Tile.prototype.setMaxResolution),goog.exportProperty(ol.layer.Tile.prototype,"setMinResolution",ol.layer.Tile.prototype.setMinResolution),goog.exportProperty(ol.layer.Tile.prototype,"setOpacity",ol.layer.Tile.prototype.setOpacity),goog.exportProperty(ol.layer.Tile.prototype,"setVisible",ol.layer.Tile.prototype.setVisible),goog.exportProperty(ol.layer.Tile.prototype,"setZIndex",ol.layer.Tile.prototype.setZIndex),goog.exportProperty(ol.layer.Tile.prototype,"get",ol.layer.Tile.prototype.get),goog.exportProperty(ol.layer.Tile.prototype,"getKeys",ol.layer.Tile.prototype.getKeys),goog.exportProperty(ol.layer.Tile.prototype,"getProperties",ol.layer.Tile.prototype.getProperties),goog.exportProperty(ol.layer.Tile.prototype,"set",ol.layer.Tile.prototype.set),goog.exportProperty(ol.layer.Tile.prototype,"setProperties",ol.layer.Tile.prototype.setProperties),goog.exportProperty(ol.layer.Tile.prototype,"unset",ol.layer.Tile.prototype.unset),goog.exportProperty(ol.layer.Tile.prototype,"changed",ol.layer.Tile.prototype.changed),goog.exportProperty(ol.layer.Tile.prototype,"dispatchEvent",ol.layer.Tile.prototype.dispatchEvent),goog.exportProperty(ol.layer.Tile.prototype,"getRevision",ol.layer.Tile.prototype.getRevision),goog.exportProperty(ol.layer.Tile.prototype,"on",ol.layer.Tile.prototype.on),goog.exportProperty(ol.layer.Tile.prototype,"once",ol.layer.Tile.prototype.once),goog.exportProperty(ol.layer.Tile.prototype,"un",ol.layer.Tile.prototype.un),goog.exportProperty(ol.layer.Tile.prototype,"unByKey",ol.layer.Tile.prototype.unByKey),goog.exportProperty(ol.layer.VectorTile.prototype,"getSource",ol.layer.VectorTile.prototype.getSource),goog.exportProperty(ol.layer.VectorTile.prototype,"getStyle",ol.layer.VectorTile.prototype.getStyle),goog.exportProperty(ol.layer.VectorTile.prototype,"getStyleFunction",ol.layer.VectorTile.prototype.getStyleFunction),goog.exportProperty(ol.layer.VectorTile.prototype,"setStyle",ol.layer.VectorTile.prototype.setStyle),goog.exportProperty(ol.layer.VectorTile.prototype,"setMap",ol.layer.VectorTile.prototype.setMap),goog.exportProperty(ol.layer.VectorTile.prototype,"setSource",ol.layer.VectorTile.prototype.setSource),goog.exportProperty(ol.layer.VectorTile.prototype,"getExtent",ol.layer.VectorTile.prototype.getExtent),goog.exportProperty(ol.layer.VectorTile.prototype,"getMaxResolution",ol.layer.VectorTile.prototype.getMaxResolution),goog.exportProperty(ol.layer.VectorTile.prototype,"getMinResolution",ol.layer.VectorTile.prototype.getMinResolution),goog.exportProperty(ol.layer.VectorTile.prototype,"getOpacity",ol.layer.VectorTile.prototype.getOpacity),goog.exportProperty(ol.layer.VectorTile.prototype,"getVisible",ol.layer.VectorTile.prototype.getVisible),goog.exportProperty(ol.layer.VectorTile.prototype,"getZIndex",ol.layer.VectorTile.prototype.getZIndex),goog.exportProperty(ol.layer.VectorTile.prototype,"setExtent",ol.layer.VectorTile.prototype.setExtent),goog.exportProperty(ol.layer.VectorTile.prototype,"setMaxResolution",ol.layer.VectorTile.prototype.setMaxResolution),goog.exportProperty(ol.layer.VectorTile.prototype,"setMinResolution",ol.layer.VectorTile.prototype.setMinResolution),goog.exportProperty(ol.layer.VectorTile.prototype,"setOpacity",ol.layer.VectorTile.prototype.setOpacity),goog.exportProperty(ol.layer.VectorTile.prototype,"setVisible",ol.layer.VectorTile.prototype.setVisible),goog.exportProperty(ol.layer.VectorTile.prototype,"setZIndex",ol.layer.VectorTile.prototype.setZIndex),goog.exportProperty(ol.layer.VectorTile.prototype,"get",ol.layer.VectorTile.prototype.get),goog.exportProperty(ol.layer.VectorTile.prototype,"getKeys",ol.layer.VectorTile.prototype.getKeys),goog.exportProperty(ol.layer.VectorTile.prototype,"getProperties",ol.layer.VectorTile.prototype.getProperties),goog.exportProperty(ol.layer.VectorTile.prototype,"set",ol.layer.VectorTile.prototype.set),goog.exportProperty(ol.layer.VectorTile.prototype,"setProperties",ol.layer.VectorTile.prototype.setProperties),goog.exportProperty(ol.layer.VectorTile.prototype,"unset",ol.layer.VectorTile.prototype.unset),goog.exportProperty(ol.layer.VectorTile.prototype,"changed",ol.layer.VectorTile.prototype.changed),goog.exportProperty(ol.layer.VectorTile.prototype,"dispatchEvent",ol.layer.VectorTile.prototype.dispatchEvent),goog.exportProperty(ol.layer.VectorTile.prototype,"getRevision",ol.layer.VectorTile.prototype.getRevision),goog.exportProperty(ol.layer.VectorTile.prototype,"on",ol.layer.VectorTile.prototype.on),goog.exportProperty(ol.layer.VectorTile.prototype,"once",ol.layer.VectorTile.prototype.once),goog.exportProperty(ol.layer.VectorTile.prototype,"un",ol.layer.VectorTile.prototype.un),goog.exportProperty(ol.layer.VectorTile.prototype,"unByKey",ol.layer.VectorTile.prototype.unByKey),goog.exportProperty(ol.interaction.Interaction.prototype,"get",ol.interaction.Interaction.prototype.get),goog.exportProperty(ol.interaction.Interaction.prototype,"getKeys",ol.interaction.Interaction.prototype.getKeys),goog.exportProperty(ol.interaction.Interaction.prototype,"getProperties",ol.interaction.Interaction.prototype.getProperties),goog.exportProperty(ol.interaction.Interaction.prototype,"set",ol.interaction.Interaction.prototype.set),goog.exportProperty(ol.interaction.Interaction.prototype,"setProperties",ol.interaction.Interaction.prototype.setProperties),goog.exportProperty(ol.interaction.Interaction.prototype,"unset",ol.interaction.Interaction.prototype.unset),goog.exportProperty(ol.interaction.Interaction.prototype,"changed",ol.interaction.Interaction.prototype.changed),goog.exportProperty(ol.interaction.Interaction.prototype,"dispatchEvent",ol.interaction.Interaction.prototype.dispatchEvent),goog.exportProperty(ol.interaction.Interaction.prototype,"getRevision",ol.interaction.Interaction.prototype.getRevision),goog.exportProperty(ol.interaction.Interaction.prototype,"on",ol.interaction.Interaction.prototype.on),goog.exportProperty(ol.interaction.Interaction.prototype,"once",ol.interaction.Interaction.prototype.once),goog.exportProperty(ol.interaction.Interaction.prototype,"un",ol.interaction.Interaction.prototype.un),goog.exportProperty(ol.interaction.Interaction.prototype,"unByKey",ol.interaction.Interaction.prototype.unByKey),goog.exportProperty(ol.interaction.DoubleClickZoom.prototype,"getActive",ol.interaction.DoubleClickZoom.prototype.getActive),goog.exportProperty(ol.interaction.DoubleClickZoom.prototype,"getMap",ol.interaction.DoubleClickZoom.prototype.getMap),goog.exportProperty(ol.interaction.DoubleClickZoom.prototype,"setActive",ol.interaction.DoubleClickZoom.prototype.setActive),goog.exportProperty(ol.interaction.DoubleClickZoom.prototype,"get",ol.interaction.DoubleClickZoom.prototype.get),goog.exportProperty(ol.interaction.DoubleClickZoom.prototype,"getKeys",ol.interaction.DoubleClickZoom.prototype.getKeys),goog.exportProperty(ol.interaction.DoubleClickZoom.prototype,"getProperties",ol.interaction.DoubleClickZoom.prototype.getProperties),goog.exportProperty(ol.interaction.DoubleClickZoom.prototype,"set",ol.interaction.DoubleClickZoom.prototype.set),goog.exportProperty(ol.interaction.DoubleClickZoom.prototype,"setProperties",ol.interaction.DoubleClickZoom.prototype.setProperties),goog.exportProperty(ol.interaction.DoubleClickZoom.prototype,"unset",ol.interaction.DoubleClickZoom.prototype.unset),goog.exportProperty(ol.interaction.DoubleClickZoom.prototype,"changed",ol.interaction.DoubleClickZoom.prototype.changed),goog.exportProperty(ol.interaction.DoubleClickZoom.prototype,"dispatchEvent",ol.interaction.DoubleClickZoom.prototype.dispatchEvent),goog.exportProperty(ol.interaction.DoubleClickZoom.prototype,"getRevision",ol.interaction.DoubleClickZoom.prototype.getRevision),goog.exportProperty(ol.interaction.DoubleClickZoom.prototype,"on",ol.interaction.DoubleClickZoom.prototype.on),goog.exportProperty(ol.interaction.DoubleClickZoom.prototype,"once",ol.interaction.DoubleClickZoom.prototype.once),goog.exportProperty(ol.interaction.DoubleClickZoom.prototype,"un",ol.interaction.DoubleClickZoom.prototype.un),goog.exportProperty(ol.interaction.DoubleClickZoom.prototype,"unByKey",ol.interaction.DoubleClickZoom.prototype.unByKey),goog.exportProperty(ol.interaction.DragAndDrop.prototype,"getActive",ol.interaction.DragAndDrop.prototype.getActive),goog.exportProperty(ol.interaction.DragAndDrop.prototype,"getMap",ol.interaction.DragAndDrop.prototype.getMap),goog.exportProperty(ol.interaction.DragAndDrop.prototype,"setActive",ol.interaction.DragAndDrop.prototype.setActive),goog.exportProperty(ol.interaction.DragAndDrop.prototype,"get",ol.interaction.DragAndDrop.prototype.get),goog.exportProperty(ol.interaction.DragAndDrop.prototype,"getKeys",ol.interaction.DragAndDrop.prototype.getKeys),goog.exportProperty(ol.interaction.DragAndDrop.prototype,"getProperties",ol.interaction.DragAndDrop.prototype.getProperties),goog.exportProperty(ol.interaction.DragAndDrop.prototype,"set",ol.interaction.DragAndDrop.prototype.set),goog.exportProperty(ol.interaction.DragAndDrop.prototype,"setProperties",ol.interaction.DragAndDrop.prototype.setProperties),goog.exportProperty(ol.interaction.DragAndDrop.prototype,"unset",ol.interaction.DragAndDrop.prototype.unset),goog.exportProperty(ol.interaction.DragAndDrop.prototype,"changed",ol.interaction.DragAndDrop.prototype.changed),goog.exportProperty(ol.interaction.DragAndDrop.prototype,"dispatchEvent",ol.interaction.DragAndDrop.prototype.dispatchEvent),goog.exportProperty(ol.interaction.DragAndDrop.prototype,"getRevision",ol.interaction.DragAndDrop.prototype.getRevision),goog.exportProperty(ol.interaction.DragAndDrop.prototype,"on",ol.interaction.DragAndDrop.prototype.on),goog.exportProperty(ol.interaction.DragAndDrop.prototype,"once",ol.interaction.DragAndDrop.prototype.once),goog.exportProperty(ol.interaction.DragAndDrop.prototype,"un",ol.interaction.DragAndDrop.prototype.un),goog.exportProperty(ol.interaction.DragAndDrop.prototype,"unByKey",ol.interaction.DragAndDrop.prototype.unByKey),goog.exportProperty(ol.interaction.DragAndDrop.Event.prototype,"type",ol.interaction.DragAndDrop.Event.prototype.type),goog.exportProperty(ol.interaction.DragAndDrop.Event.prototype,"target",ol.interaction.DragAndDrop.Event.prototype.target),goog.exportProperty(ol.interaction.DragAndDrop.Event.prototype,"preventDefault",ol.interaction.DragAndDrop.Event.prototype.preventDefault),goog.exportProperty(ol.interaction.DragAndDrop.Event.prototype,"stopPropagation",ol.interaction.DragAndDrop.Event.prototype.stopPropagation),goog.exportProperty(ol.interaction.Pointer.prototype,"getActive",ol.interaction.Pointer.prototype.getActive),goog.exportProperty(ol.interaction.Pointer.prototype,"getMap",ol.interaction.Pointer.prototype.getMap),goog.exportProperty(ol.interaction.Pointer.prototype,"setActive",ol.interaction.Pointer.prototype.setActive),goog.exportProperty(ol.interaction.Pointer.prototype,"get",ol.interaction.Pointer.prototype.get),goog.exportProperty(ol.interaction.Pointer.prototype,"getKeys",ol.interaction.Pointer.prototype.getKeys),goog.exportProperty(ol.interaction.Pointer.prototype,"getProperties",ol.interaction.Pointer.prototype.getProperties),goog.exportProperty(ol.interaction.Pointer.prototype,"set",ol.interaction.Pointer.prototype.set),goog.exportProperty(ol.interaction.Pointer.prototype,"setProperties",ol.interaction.Pointer.prototype.setProperties),goog.exportProperty(ol.interaction.Pointer.prototype,"unset",ol.interaction.Pointer.prototype.unset),goog.exportProperty(ol.interaction.Pointer.prototype,"changed",ol.interaction.Pointer.prototype.changed),
 goog.exportProperty(ol.interaction.Pointer.prototype,"dispatchEvent",ol.interaction.Pointer.prototype.dispatchEvent),goog.exportProperty(ol.interaction.Pointer.prototype,"getRevision",ol.interaction.Pointer.prototype.getRevision),goog.exportProperty(ol.interaction.Pointer.prototype,"on",ol.interaction.Pointer.prototype.on),goog.exportProperty(ol.interaction.Pointer.prototype,"once",ol.interaction.Pointer.prototype.once),goog.exportProperty(ol.interaction.Pointer.prototype,"un",ol.interaction.Pointer.prototype.un),goog.exportProperty(ol.interaction.Pointer.prototype,"unByKey",ol.interaction.Pointer.prototype.unByKey),goog.exportProperty(ol.interaction.DragBox.prototype,"getActive",ol.interaction.DragBox.prototype.getActive),goog.exportProperty(ol.interaction.DragBox.prototype,"getMap",ol.interaction.DragBox.prototype.getMap),goog.exportProperty(ol.interaction.DragBox.prototype,"setActive",ol.interaction.DragBox.prototype.setActive),goog.exportProperty(ol.interaction.DragBox.prototype,"get",ol.interaction.DragBox.prototype.get),goog.exportProperty(ol.interaction.DragBox.prototype,"getKeys",ol.interaction.DragBox.prototype.getKeys),goog.exportProperty(ol.interaction.DragBox.prototype,"getProperties",ol.interaction.DragBox.prototype.getProperties),goog.exportProperty(ol.interaction.DragBox.prototype,"set",ol.interaction.DragBox.prototype.set),goog.exportProperty(ol.interaction.DragBox.prototype,"setProperties",ol.interaction.DragBox.prototype.setProperties),goog.exportProperty(ol.interaction.DragBox.prototype,"unset",ol.interaction.DragBox.prototype.unset),goog.exportProperty(ol.interaction.DragBox.prototype,"changed",ol.interaction.DragBox.prototype.changed),goog.exportProperty(ol.interaction.DragBox.prototype,"dispatchEvent",ol.interaction.DragBox.prototype.dispatchEvent),goog.exportProperty(ol.interaction.DragBox.prototype,"getRevision",ol.interaction.DragBox.prototype.getRevision),goog.exportProperty(ol.interaction.DragBox.prototype,"on",ol.interaction.DragBox.prototype.on),goog.exportProperty(ol.interaction.DragBox.prototype,"once",ol.interaction.DragBox.prototype.once),goog.exportProperty(ol.interaction.DragBox.prototype,"un",ol.interaction.DragBox.prototype.un),goog.exportProperty(ol.interaction.DragBox.prototype,"unByKey",ol.interaction.DragBox.prototype.unByKey),goog.exportProperty(ol.interaction.DragBox.Event.prototype,"type",ol.interaction.DragBox.Event.prototype.type),goog.exportProperty(ol.interaction.DragBox.Event.prototype,"target",ol.interaction.DragBox.Event.prototype.target),goog.exportProperty(ol.interaction.DragBox.Event.prototype,"preventDefault",ol.interaction.DragBox.Event.prototype.preventDefault),goog.exportProperty(ol.interaction.DragBox.Event.prototype,"stopPropagation",ol.interaction.DragBox.Event.prototype.stopPropagation),goog.exportProperty(ol.interaction.DragPan.prototype,"getActive",ol.interaction.DragPan.prototype.getActive),goog.exportProperty(ol.interaction.DragPan.prototype,"getMap",ol.interaction.DragPan.prototype.getMap),goog.exportProperty(ol.interaction.DragPan.prototype,"setActive",ol.interaction.DragPan.prototype.setActive),goog.exportProperty(ol.interaction.DragPan.prototype,"get",ol.interaction.DragPan.prototype.get),goog.exportProperty(ol.interaction.DragPan.prototype,"getKeys",ol.interaction.DragPan.prototype.getKeys),goog.exportProperty(ol.interaction.DragPan.prototype,"getProperties",ol.interaction.DragPan.prototype.getProperties),goog.exportProperty(ol.interaction.DragPan.prototype,"set",ol.interaction.DragPan.prototype.set),goog.exportProperty(ol.interaction.DragPan.prototype,"setProperties",ol.interaction.DragPan.prototype.setProperties),goog.exportProperty(ol.interaction.DragPan.prototype,"unset",ol.interaction.DragPan.prototype.unset),goog.exportProperty(ol.interaction.DragPan.prototype,"changed",ol.interaction.DragPan.prototype.changed),goog.exportProperty(ol.interaction.DragPan.prototype,"dispatchEvent",ol.interaction.DragPan.prototype.dispatchEvent),goog.exportProperty(ol.interaction.DragPan.prototype,"getRevision",ol.interaction.DragPan.prototype.getRevision),goog.exportProperty(ol.interaction.DragPan.prototype,"on",ol.interaction.DragPan.prototype.on),goog.exportProperty(ol.interaction.DragPan.prototype,"once",ol.interaction.DragPan.prototype.once),goog.exportProperty(ol.interaction.DragPan.prototype,"un",ol.interaction.DragPan.prototype.un),goog.exportProperty(ol.interaction.DragPan.prototype,"unByKey",ol.interaction.DragPan.prototype.unByKey),goog.exportProperty(ol.interaction.DragRotate.prototype,"getActive",ol.interaction.DragRotate.prototype.getActive),goog.exportProperty(ol.interaction.DragRotate.prototype,"getMap",ol.interaction.DragRotate.prototype.getMap),goog.exportProperty(ol.interaction.DragRotate.prototype,"setActive",ol.interaction.DragRotate.prototype.setActive),goog.exportProperty(ol.interaction.DragRotate.prototype,"get",ol.interaction.DragRotate.prototype.get),goog.exportProperty(ol.interaction.DragRotate.prototype,"getKeys",ol.interaction.DragRotate.prototype.getKeys),goog.exportProperty(ol.interaction.DragRotate.prototype,"getProperties",ol.interaction.DragRotate.prototype.getProperties),goog.exportProperty(ol.interaction.DragRotate.prototype,"set",ol.interaction.DragRotate.prototype.set),goog.exportProperty(ol.interaction.DragRotate.prototype,"setProperties",ol.interaction.DragRotate.prototype.setProperties),goog.exportProperty(ol.interaction.DragRotate.prototype,"unset",ol.interaction.DragRotate.prototype.unset),goog.exportProperty(ol.interaction.DragRotate.prototype,"changed",ol.interaction.DragRotate.prototype.changed),goog.exportProperty(ol.interaction.DragRotate.prototype,"dispatchEvent",ol.interaction.DragRotate.prototype.dispatchEvent),goog.exportProperty(ol.interaction.DragRotate.prototype,"getRevision",ol.interaction.DragRotate.prototype.getRevision),goog.exportProperty(ol.interaction.DragRotate.prototype,"on",ol.interaction.DragRotate.prototype.on),goog.exportProperty(ol.interaction.DragRotate.prototype,"once",ol.interaction.DragRotate.prototype.once),goog.exportProperty(ol.interaction.DragRotate.prototype,"un",ol.interaction.DragRotate.prototype.un),goog.exportProperty(ol.interaction.DragRotate.prototype,"unByKey",ol.interaction.DragRotate.prototype.unByKey),goog.exportProperty(ol.interaction.DragRotateAndZoom.prototype,"getActive",ol.interaction.DragRotateAndZoom.prototype.getActive),goog.exportProperty(ol.interaction.DragRotateAndZoom.prototype,"getMap",ol.interaction.DragRotateAndZoom.prototype.getMap),goog.exportProperty(ol.interaction.DragRotateAndZoom.prototype,"setActive",ol.interaction.DragRotateAndZoom.prototype.setActive),goog.exportProperty(ol.interaction.DragRotateAndZoom.prototype,"get",ol.interaction.DragRotateAndZoom.prototype.get),goog.exportProperty(ol.interaction.DragRotateAndZoom.prototype,"getKeys",ol.interaction.DragRotateAndZoom.prototype.getKeys),goog.exportProperty(ol.interaction.DragRotateAndZoom.prototype,"getProperties",ol.interaction.DragRotateAndZoom.prototype.getProperties),goog.exportProperty(ol.interaction.DragRotateAndZoom.prototype,"set",ol.interaction.DragRotateAndZoom.prototype.set),goog.exportProperty(ol.interaction.DragRotateAndZoom.prototype,"setProperties",ol.interaction.DragRotateAndZoom.prototype.setProperties),goog.exportProperty(ol.interaction.DragRotateAndZoom.prototype,"unset",ol.interaction.DragRotateAndZoom.prototype.unset),goog.exportProperty(ol.interaction.DragRotateAndZoom.prototype,"changed",ol.interaction.DragRotateAndZoom.prototype.changed),goog.exportProperty(ol.interaction.DragRotateAndZoom.prototype,"dispatchEvent",ol.interaction.DragRotateAndZoom.prototype.dispatchEvent),goog.exportProperty(ol.interaction.DragRotateAndZoom.prototype,"getRevision",ol.interaction.DragRotateAndZoom.prototype.getRevision),goog.exportProperty(ol.interaction.DragRotateAndZoom.prototype,"on",ol.interaction.DragRotateAndZoom.prototype.on),goog.exportProperty(ol.interaction.DragRotateAndZoom.prototype,"once",ol.interaction.DragRotateAndZoom.prototype.once),goog.exportProperty(ol.interaction.DragRotateAndZoom.prototype,"un",ol.interaction.DragRotateAndZoom.prototype.un),goog.exportProperty(ol.interaction.DragRotateAndZoom.prototype,"unByKey",ol.interaction.DragRotateAndZoom.prototype.unByKey),goog.exportProperty(ol.interaction.DragZoom.prototype,"getGeometry",ol.interaction.DragZoom.prototype.getGeometry),goog.exportProperty(ol.interaction.DragZoom.prototype,"getActive",ol.interaction.DragZoom.prototype.getActive),goog.exportProperty(ol.interaction.DragZoom.prototype,"getMap",ol.interaction.DragZoom.prototype.getMap),goog.exportProperty(ol.interaction.DragZoom.prototype,"setActive",ol.interaction.DragZoom.prototype.setActive),goog.exportProperty(ol.interaction.DragZoom.prototype,"get",ol.interaction.DragZoom.prototype.get),goog.exportProperty(ol.interaction.DragZoom.prototype,"getKeys",ol.interaction.DragZoom.prototype.getKeys),goog.exportProperty(ol.interaction.DragZoom.prototype,"getProperties",ol.interaction.DragZoom.prototype.getProperties),goog.exportProperty(ol.interaction.DragZoom.prototype,"set",ol.interaction.DragZoom.prototype.set),goog.exportProperty(ol.interaction.DragZoom.prototype,"setProperties",ol.interaction.DragZoom.prototype.setProperties),goog.exportProperty(ol.interaction.DragZoom.prototype,"unset",ol.interaction.DragZoom.prototype.unset),goog.exportProperty(ol.interaction.DragZoom.prototype,"changed",ol.interaction.DragZoom.prototype.changed),goog.exportProperty(ol.interaction.DragZoom.prototype,"dispatchEvent",ol.interaction.DragZoom.prototype.dispatchEvent),goog.exportProperty(ol.interaction.DragZoom.prototype,"getRevision",ol.interaction.DragZoom.prototype.getRevision),goog.exportProperty(ol.interaction.DragZoom.prototype,"on",ol.interaction.DragZoom.prototype.on),goog.exportProperty(ol.interaction.DragZoom.prototype,"once",ol.interaction.DragZoom.prototype.once),goog.exportProperty(ol.interaction.DragZoom.prototype,"un",ol.interaction.DragZoom.prototype.un),goog.exportProperty(ol.interaction.DragZoom.prototype,"unByKey",ol.interaction.DragZoom.prototype.unByKey),goog.exportProperty(ol.interaction.Draw.prototype,"getActive",ol.interaction.Draw.prototype.getActive),goog.exportProperty(ol.interaction.Draw.prototype,"getMap",ol.interaction.Draw.prototype.getMap),goog.exportProperty(ol.interaction.Draw.prototype,"setActive",ol.interaction.Draw.prototype.setActive),goog.exportProperty(ol.interaction.Draw.prototype,"get",ol.interaction.Draw.prototype.get),goog.exportProperty(ol.interaction.Draw.prototype,"getKeys",ol.interaction.Draw.prototype.getKeys),goog.exportProperty(ol.interaction.Draw.prototype,"getProperties",ol.interaction.Draw.prototype.getProperties),goog.exportProperty(ol.interaction.Draw.prototype,"set",ol.interaction.Draw.prototype.set),goog.exportProperty(ol.interaction.Draw.prototype,"setProperties",ol.interaction.Draw.prototype.setProperties),goog.exportProperty(ol.interaction.Draw.prototype,"unset",ol.interaction.Draw.prototype.unset),goog.exportProperty(ol.interaction.Draw.prototype,"changed",ol.interaction.Draw.prototype.changed),goog.exportProperty(ol.interaction.Draw.prototype,"dispatchEvent",ol.interaction.Draw.prototype.dispatchEvent),goog.exportProperty(ol.interaction.Draw.prototype,"getRevision",ol.interaction.Draw.prototype.getRevision),goog.exportProperty(ol.interaction.Draw.prototype,"on",ol.interaction.Draw.prototype.on),goog.exportProperty(ol.interaction.Draw.prototype,"once",ol.interaction.Draw.prototype.once),goog.exportProperty(ol.interaction.Draw.prototype,"un",ol.interaction.Draw.prototype.un),goog.exportProperty(ol.interaction.Draw.prototype,"unByKey",ol.interaction.Draw.prototype.unByKey),goog.exportProperty(ol.interaction.Draw.Event.prototype,"type",ol.interaction.Draw.Event.prototype.type),goog.exportProperty(ol.interaction.Draw.Event.prototype,"target",ol.interaction.Draw.Event.prototype.target),goog.exportProperty(ol.interaction.Draw.Event.prototype,"preventDefault",ol.interaction.Draw.Event.prototype.preventDefault),goog.exportProperty(ol.interaction.Draw.Event.prototype,"stopPropagation",ol.interaction.Draw.Event.prototype.stopPropagation),goog.exportProperty(ol.interaction.Extent.prototype,"getActive",ol.interaction.Extent.prototype.getActive),goog.exportProperty(ol.interaction.Extent.prototype,"getMap",ol.interaction.Extent.prototype.getMap),goog.exportProperty(ol.interaction.Extent.prototype,"setActive",ol.interaction.Extent.prototype.setActive),goog.exportProperty(ol.interaction.Extent.prototype,"get",ol.interaction.Extent.prototype.get),goog.exportProperty(ol.interaction.Extent.prototype,"getKeys",ol.interaction.Extent.prototype.getKeys),goog.exportProperty(ol.interaction.Extent.prototype,"getProperties",ol.interaction.Extent.prototype.getProperties),goog.exportProperty(ol.interaction.Extent.prototype,"set",ol.interaction.Extent.prototype.set),goog.exportProperty(ol.interaction.Extent.prototype,"setProperties",ol.interaction.Extent.prototype.setProperties),goog.exportProperty(ol.interaction.Extent.prototype,"unset",ol.interaction.Extent.prototype.unset),goog.exportProperty(ol.interaction.Extent.prototype,"changed",ol.interaction.Extent.prototype.changed),goog.exportProperty(ol.interaction.Extent.prototype,"dispatchEvent",ol.interaction.Extent.prototype.dispatchEvent),goog.exportProperty(ol.interaction.Extent.prototype,"getRevision",ol.interaction.Extent.prototype.getRevision),goog.exportProperty(ol.interaction.Extent.prototype,"on",ol.interaction.Extent.prototype.on),goog.exportProperty(ol.interaction.Extent.prototype,"once",ol.interaction.Extent.prototype.once),goog.exportProperty(ol.interaction.Extent.prototype,"un",ol.interaction.Extent.prototype.un),goog.exportProperty(ol.interaction.Extent.prototype,"unByKey",ol.interaction.Extent.prototype.unByKey),goog.exportProperty(ol.interaction.Extent.Event.prototype,"type",ol.interaction.Extent.Event.prototype.type),goog.exportProperty(ol.interaction.Extent.Event.prototype,"target",ol.interaction.Extent.Event.prototype.target),goog.exportProperty(ol.interaction.Extent.Event.prototype,"preventDefault",ol.interaction.Extent.Event.prototype.preventDefault),goog.exportProperty(ol.interaction.Extent.Event.prototype,"stopPropagation",ol.interaction.Extent.Event.prototype.stopPropagation),goog.exportProperty(ol.interaction.KeyboardPan.prototype,"getActive",ol.interaction.KeyboardPan.prototype.getActive),goog.exportProperty(ol.interaction.KeyboardPan.prototype,"getMap",ol.interaction.KeyboardPan.prototype.getMap),goog.exportProperty(ol.interaction.KeyboardPan.prototype,"setActive",ol.interaction.KeyboardPan.prototype.setActive),goog.exportProperty(ol.interaction.KeyboardPan.prototype,"get",ol.interaction.KeyboardPan.prototype.get),goog.exportProperty(ol.interaction.KeyboardPan.prototype,"getKeys",ol.interaction.KeyboardPan.prototype.getKeys),goog.exportProperty(ol.interaction.KeyboardPan.prototype,"getProperties",ol.interaction.KeyboardPan.prototype.getProperties),goog.exportProperty(ol.interaction.KeyboardPan.prototype,"set",ol.interaction.KeyboardPan.prototype.set),goog.exportProperty(ol.interaction.KeyboardPan.prototype,"setProperties",ol.interaction.KeyboardPan.prototype.setProperties),goog.exportProperty(ol.interaction.KeyboardPan.prototype,"unset",ol.interaction.KeyboardPan.prototype.unset),goog.exportProperty(ol.interaction.KeyboardPan.prototype,"changed",ol.interaction.KeyboardPan.prototype.changed),goog.exportProperty(ol.interaction.KeyboardPan.prototype,"dispatchEvent",ol.interaction.KeyboardPan.prototype.dispatchEvent),goog.exportProperty(ol.interaction.KeyboardPan.prototype,"getRevision",ol.interaction.KeyboardPan.prototype.getRevision),goog.exportProperty(ol.interaction.KeyboardPan.prototype,"on",ol.interaction.KeyboardPan.prototype.on),goog.exportProperty(ol.interaction.KeyboardPan.prototype,"once",ol.interaction.KeyboardPan.prototype.once),goog.exportProperty(ol.interaction.KeyboardPan.prototype,"un",ol.interaction.KeyboardPan.prototype.un),goog.exportProperty(ol.interaction.KeyboardPan.prototype,"unByKey",ol.interaction.KeyboardPan.prototype.unByKey),goog.exportProperty(ol.interaction.KeyboardZoom.prototype,"getActive",ol.interaction.KeyboardZoom.prototype.getActive),goog.exportProperty(ol.interaction.KeyboardZoom.prototype,"getMap",ol.interaction.KeyboardZoom.prototype.getMap),goog.exportProperty(ol.interaction.KeyboardZoom.prototype,"setActive",ol.interaction.KeyboardZoom.prototype.setActive),goog.exportProperty(ol.interaction.KeyboardZoom.prototype,"get",ol.interaction.KeyboardZoom.prototype.get),goog.exportProperty(ol.interaction.KeyboardZoom.prototype,"getKeys",ol.interaction.KeyboardZoom.prototype.getKeys),goog.exportProperty(ol.interaction.KeyboardZoom.prototype,"getProperties",ol.interaction.KeyboardZoom.prototype.getProperties),goog.exportProperty(ol.interaction.KeyboardZoom.prototype,"set",ol.interaction.KeyboardZoom.prototype.set),goog.exportProperty(ol.interaction.KeyboardZoom.prototype,"setProperties",ol.interaction.KeyboardZoom.prototype.setProperties),goog.exportProperty(ol.interaction.KeyboardZoom.prototype,"unset",ol.interaction.KeyboardZoom.prototype.unset),goog.exportProperty(ol.interaction.KeyboardZoom.prototype,"changed",ol.interaction.KeyboardZoom.prototype.changed),goog.exportProperty(ol.interaction.KeyboardZoom.prototype,"dispatchEvent",ol.interaction.KeyboardZoom.prototype.dispatchEvent),goog.exportProperty(ol.interaction.KeyboardZoom.prototype,"getRevision",ol.interaction.KeyboardZoom.prototype.getRevision),goog.exportProperty(ol.interaction.KeyboardZoom.prototype,"on",ol.interaction.KeyboardZoom.prototype.on),goog.exportProperty(ol.interaction.KeyboardZoom.prototype,"once",ol.interaction.KeyboardZoom.prototype.once),goog.exportProperty(ol.interaction.KeyboardZoom.prototype,"un",ol.interaction.KeyboardZoom.prototype.un),goog.exportProperty(ol.interaction.KeyboardZoom.prototype,"unByKey",ol.interaction.KeyboardZoom.prototype.unByKey),goog.exportProperty(ol.interaction.Modify.prototype,"getActive",ol.interaction.Modify.prototype.getActive),goog.exportProperty(ol.interaction.Modify.prototype,"getMap",ol.interaction.Modify.prototype.getMap),goog.exportProperty(ol.interaction.Modify.prototype,"setActive",ol.interaction.Modify.prototype.setActive),goog.exportProperty(ol.interaction.Modify.prototype,"get",ol.interaction.Modify.prototype.get),goog.exportProperty(ol.interaction.Modify.prototype,"getKeys",ol.interaction.Modify.prototype.getKeys),goog.exportProperty(ol.interaction.Modify.prototype,"getProperties",ol.interaction.Modify.prototype.getProperties),goog.exportProperty(ol.interaction.Modify.prototype,"set",ol.interaction.Modify.prototype.set),goog.exportProperty(ol.interaction.Modify.prototype,"setProperties",ol.interaction.Modify.prototype.setProperties),goog.exportProperty(ol.interaction.Modify.prototype,"unset",ol.interaction.Modify.prototype.unset),goog.exportProperty(ol.interaction.Modify.prototype,"changed",ol.interaction.Modify.prototype.changed),goog.exportProperty(ol.interaction.Modify.prototype,"dispatchEvent",ol.interaction.Modify.prototype.dispatchEvent),goog.exportProperty(ol.interaction.Modify.prototype,"getRevision",ol.interaction.Modify.prototype.getRevision),goog.exportProperty(ol.interaction.Modify.prototype,"on",ol.interaction.Modify.prototype.on),goog.exportProperty(ol.interaction.Modify.prototype,"once",ol.interaction.Modify.prototype.once),goog.exportProperty(ol.interaction.Modify.prototype,"un",ol.interaction.Modify.prototype.un),goog.exportProperty(ol.interaction.Modify.prototype,"unByKey",ol.interaction.Modify.prototype.unByKey),goog.exportProperty(ol.interaction.Modify.Event.prototype,"type",ol.interaction.Modify.Event.prototype.type),goog.exportProperty(ol.interaction.Modify.Event.prototype,"target",ol.interaction.Modify.Event.prototype.target),goog.exportProperty(ol.interaction.Modify.Event.prototype,"preventDefault",ol.interaction.Modify.Event.prototype.preventDefault),goog.exportProperty(ol.interaction.Modify.Event.prototype,"stopPropagation",ol.interaction.Modify.Event.prototype.stopPropagation),goog.exportProperty(ol.interaction.MouseWheelZoom.prototype,"getActive",ol.interaction.MouseWheelZoom.prototype.getActive),goog.exportProperty(ol.interaction.MouseWheelZoom.prototype,"getMap",ol.interaction.MouseWheelZoom.prototype.getMap),goog.exportProperty(ol.interaction.MouseWheelZoom.prototype,"setActive",ol.interaction.MouseWheelZoom.prototype.setActive),goog.exportProperty(ol.interaction.MouseWheelZoom.prototype,"get",ol.interaction.MouseWheelZoom.prototype.get),goog.exportProperty(ol.interaction.MouseWheelZoom.prototype,"getKeys",ol.interaction.MouseWheelZoom.prototype.getKeys),goog.exportProperty(ol.interaction.MouseWheelZoom.prototype,"getProperties",ol.interaction.MouseWheelZoom.prototype.getProperties),goog.exportProperty(ol.interaction.MouseWheelZoom.prototype,"set",ol.interaction.MouseWheelZoom.prototype.set),goog.exportProperty(ol.interaction.MouseWheelZoom.prototype,"setProperties",ol.interaction.MouseWheelZoom.prototype.setProperties),goog.exportProperty(ol.interaction.MouseWheelZoom.prototype,"unset",ol.interaction.MouseWheelZoom.prototype.unset),goog.exportProperty(ol.interaction.MouseWheelZoom.prototype,"changed",ol.interaction.MouseWheelZoom.prototype.changed),goog.exportProperty(ol.interaction.MouseWheelZoom.prototype,"dispatchEvent",ol.interaction.MouseWheelZoom.prototype.dispatchEvent),goog.exportProperty(ol.interaction.MouseWheelZoom.prototype,"getRevision",ol.interaction.MouseWheelZoom.prototype.getRevision),goog.exportProperty(ol.interaction.MouseWheelZoom.prototype,"on",ol.interaction.MouseWheelZoom.prototype.on),goog.exportProperty(ol.interaction.MouseWheelZoom.prototype,"once",ol.interaction.MouseWheelZoom.prototype.once),goog.exportProperty(ol.interaction.MouseWheelZoom.prototype,"un",ol.interaction.MouseWheelZoom.prototype.un),goog.exportProperty(ol.interaction.MouseWheelZoom.prototype,"unByKey",ol.interaction.MouseWheelZoom.prototype.unByKey),goog.exportProperty(ol.interaction.PinchRotate.prototype,"getActive",ol.interaction.PinchRotate.prototype.getActive),goog.exportProperty(ol.interaction.PinchRotate.prototype,"getMap",ol.interaction.PinchRotate.prototype.getMap),goog.exportProperty(ol.interaction.PinchRotate.prototype,"setActive",ol.interaction.PinchRotate.prototype.setActive),goog.exportProperty(ol.interaction.PinchRotate.prototype,"get",ol.interaction.PinchRotate.prototype.get),goog.exportProperty(ol.interaction.PinchRotate.prototype,"getKeys",ol.interaction.PinchRotate.prototype.getKeys),goog.exportProperty(ol.interaction.PinchRotate.prototype,"getProperties",ol.interaction.PinchRotate.prototype.getProperties),goog.exportProperty(ol.interaction.PinchRotate.prototype,"set",ol.interaction.PinchRotate.prototype.set),goog.exportProperty(ol.interaction.PinchRotate.prototype,"setProperties",ol.interaction.PinchRotate.prototype.setProperties),goog.exportProperty(ol.interaction.PinchRotate.prototype,"unset",ol.interaction.PinchRotate.prototype.unset),goog.exportProperty(ol.interaction.PinchRotate.prototype,"changed",ol.interaction.PinchRotate.prototype.changed),goog.exportProperty(ol.interaction.PinchRotate.prototype,"dispatchEvent",ol.interaction.PinchRotate.prototype.dispatchEvent),goog.exportProperty(ol.interaction.PinchRotate.prototype,"getRevision",ol.interaction.PinchRotate.prototype.getRevision),goog.exportProperty(ol.interaction.PinchRotate.prototype,"on",ol.interaction.PinchRotate.prototype.on),goog.exportProperty(ol.interaction.PinchRotate.prototype,"once",ol.interaction.PinchRotate.prototype.once),goog.exportProperty(ol.interaction.PinchRotate.prototype,"un",ol.interaction.PinchRotate.prototype.un),goog.exportProperty(ol.interaction.PinchRotate.prototype,"unByKey",ol.interaction.PinchRotate.prototype.unByKey),goog.exportProperty(ol.interaction.PinchZoom.prototype,"getActive",ol.interaction.PinchZoom.prototype.getActive),goog.exportProperty(ol.interaction.PinchZoom.prototype,"getMap",ol.interaction.PinchZoom.prototype.getMap),goog.exportProperty(ol.interaction.PinchZoom.prototype,"setActive",ol.interaction.PinchZoom.prototype.setActive),goog.exportProperty(ol.interaction.PinchZoom.prototype,"get",ol.interaction.PinchZoom.prototype.get),goog.exportProperty(ol.interaction.PinchZoom.prototype,"getKeys",ol.interaction.PinchZoom.prototype.getKeys),goog.exportProperty(ol.interaction.PinchZoom.prototype,"getProperties",ol.interaction.PinchZoom.prototype.getProperties),goog.exportProperty(ol.interaction.PinchZoom.prototype,"set",ol.interaction.PinchZoom.prototype.set),goog.exportProperty(ol.interaction.PinchZoom.prototype,"setProperties",ol.interaction.PinchZoom.prototype.setProperties),goog.exportProperty(ol.interaction.PinchZoom.prototype,"unset",ol.interaction.PinchZoom.prototype.unset),goog.exportProperty(ol.interaction.PinchZoom.prototype,"changed",ol.interaction.PinchZoom.prototype.changed),goog.exportProperty(ol.interaction.PinchZoom.prototype,"dispatchEvent",ol.interaction.PinchZoom.prototype.dispatchEvent),goog.exportProperty(ol.interaction.PinchZoom.prototype,"getRevision",ol.interaction.PinchZoom.prototype.getRevision),goog.exportProperty(ol.interaction.PinchZoom.prototype,"on",ol.interaction.PinchZoom.prototype.on),goog.exportProperty(ol.interaction.PinchZoom.prototype,"once",ol.interaction.PinchZoom.prototype.once),goog.exportProperty(ol.interaction.PinchZoom.prototype,"un",ol.interaction.PinchZoom.prototype.un),goog.exportProperty(ol.interaction.PinchZoom.prototype,"unByKey",ol.interaction.PinchZoom.prototype.unByKey),goog.exportProperty(ol.interaction.Select.prototype,"getActive",ol.interaction.Select.prototype.getActive),goog.exportProperty(ol.interaction.Select.prototype,"getMap",ol.interaction.Select.prototype.getMap),goog.exportProperty(ol.interaction.Select.prototype,"setActive",ol.interaction.Select.prototype.setActive),goog.exportProperty(ol.interaction.Select.prototype,"get",ol.interaction.Select.prototype.get),goog.exportProperty(ol.interaction.Select.prototype,"getKeys",ol.interaction.Select.prototype.getKeys),goog.exportProperty(ol.interaction.Select.prototype,"getProperties",ol.interaction.Select.prototype.getProperties),goog.exportProperty(ol.interaction.Select.prototype,"set",ol.interaction.Select.prototype.set),goog.exportProperty(ol.interaction.Select.prototype,"setProperties",ol.interaction.Select.prototype.setProperties),goog.exportProperty(ol.interaction.Select.prototype,"unset",ol.interaction.Select.prototype.unset),goog.exportProperty(ol.interaction.Select.prototype,"changed",ol.interaction.Select.prototype.changed),goog.exportProperty(ol.interaction.Select.prototype,"dispatchEvent",ol.interaction.Select.prototype.dispatchEvent),goog.exportProperty(ol.interaction.Select.prototype,"getRevision",ol.interaction.Select.prototype.getRevision),goog.exportProperty(ol.interaction.Select.prototype,"on",ol.interaction.Select.prototype.on),goog.exportProperty(ol.interaction.Select.prototype,"once",ol.interaction.Select.prototype.once),goog.exportProperty(ol.interaction.Select.prototype,"un",ol.interaction.Select.prototype.un),goog.exportProperty(ol.interaction.Select.prototype,"unByKey",ol.interaction.Select.prototype.unByKey),goog.exportProperty(ol.interaction.Select.Event.prototype,"type",ol.interaction.Select.Event.prototype.type),goog.exportProperty(ol.interaction.Select.Event.prototype,"target",ol.interaction.Select.Event.prototype.target),goog.exportProperty(ol.interaction.Select.Event.prototype,"preventDefault",ol.interaction.Select.Event.prototype.preventDefault),goog.exportProperty(ol.interaction.Select.Event.prototype,"stopPropagation",ol.interaction.Select.Event.prototype.stopPropagation),goog.exportProperty(ol.interaction.Snap.prototype,"getActive",ol.interaction.Snap.prototype.getActive),goog.exportProperty(ol.interaction.Snap.prototype,"getMap",ol.interaction.Snap.prototype.getMap),goog.exportProperty(ol.interaction.Snap.prototype,"setActive",ol.interaction.Snap.prototype.setActive),goog.exportProperty(ol.interaction.Snap.prototype,"get",ol.interaction.Snap.prototype.get),goog.exportProperty(ol.interaction.Snap.prototype,"getKeys",ol.interaction.Snap.prototype.getKeys),goog.exportProperty(ol.interaction.Snap.prototype,"getProperties",ol.interaction.Snap.prototype.getProperties),goog.exportProperty(ol.interaction.Snap.prototype,"set",ol.interaction.Snap.prototype.set),goog.exportProperty(ol.interaction.Snap.prototype,"setProperties",ol.interaction.Snap.prototype.setProperties),goog.exportProperty(ol.interaction.Snap.prototype,"unset",ol.interaction.Snap.prototype.unset),goog.exportProperty(ol.interaction.Snap.prototype,"changed",ol.interaction.Snap.prototype.changed),goog.exportProperty(ol.interaction.Snap.prototype,"dispatchEvent",ol.interaction.Snap.prototype.dispatchEvent),goog.exportProperty(ol.interaction.Snap.prototype,"getRevision",ol.interaction.Snap.prototype.getRevision),goog.exportProperty(ol.interaction.Snap.prototype,"on",ol.interaction.Snap.prototype.on),goog.exportProperty(ol.interaction.Snap.prototype,"once",ol.interaction.Snap.prototype.once),goog.exportProperty(ol.interaction.Snap.prototype,"un",ol.interaction.Snap.prototype.un),goog.exportProperty(ol.interaction.Snap.prototype,"unByKey",ol.interaction.Snap.prototype.unByKey),goog.exportProperty(ol.interaction.Translate.prototype,"getActive",ol.interaction.Translate.prototype.getActive),goog.exportProperty(ol.interaction.Translate.prototype,"getMap",ol.interaction.Translate.prototype.getMap),goog.exportProperty(ol.interaction.Translate.prototype,"setActive",ol.interaction.Translate.prototype.setActive),goog.exportProperty(ol.interaction.Translate.prototype,"get",ol.interaction.Translate.prototype.get),goog.exportProperty(ol.interaction.Translate.prototype,"getKeys",ol.interaction.Translate.prototype.getKeys),goog.exportProperty(ol.interaction.Translate.prototype,"getProperties",ol.interaction.Translate.prototype.getProperties),goog.exportProperty(ol.interaction.Translate.prototype,"set",ol.interaction.Translate.prototype.set),goog.exportProperty(ol.interaction.Translate.prototype,"setProperties",ol.interaction.Translate.prototype.setProperties),goog.exportProperty(ol.interaction.Translate.prototype,"unset",ol.interaction.Translate.prototype.unset),goog.exportProperty(ol.interaction.Translate.prototype,"changed",ol.interaction.Translate.prototype.changed),goog.exportProperty(ol.interaction.Translate.prototype,"dispatchEvent",ol.interaction.Translate.prototype.dispatchEvent),goog.exportProperty(ol.interaction.Translate.prototype,"getRevision",ol.interaction.Translate.prototype.getRevision),goog.exportProperty(ol.interaction.Translate.prototype,"on",ol.interaction.Translate.prototype.on),goog.exportProperty(ol.interaction.Translate.prototype,"once",ol.interaction.Translate.prototype.once),goog.exportProperty(ol.interaction.Translate.prototype,"un",ol.interaction.Translate.prototype.un),goog.exportProperty(ol.interaction.Translate.prototype,"unByKey",ol.interaction.Translate.prototype.unByKey),goog.exportProperty(ol.interaction.Translate.Event.prototype,"type",ol.interaction.Translate.Event.prototype.type),goog.exportProperty(ol.interaction.Translate.Event.prototype,"target",ol.interaction.Translate.Event.prototype.target),goog.exportProperty(ol.interaction.Translate.Event.prototype,"preventDefault",ol.interaction.Translate.Event.prototype.preventDefault),goog.exportProperty(ol.interaction.Translate.Event.prototype,"stopPropagation",ol.interaction.Translate.Event.prototype.stopPropagation),goog.exportProperty(ol.geom.Geometry.prototype,"get",ol.geom.Geometry.prototype.get),goog.exportProperty(ol.geom.Geometry.prototype,"getKeys",ol.geom.Geometry.prototype.getKeys),goog.exportProperty(ol.geom.Geometry.prototype,"getProperties",ol.geom.Geometry.prototype.getProperties),goog.exportProperty(ol.geom.Geometry.prototype,"set",ol.geom.Geometry.prototype.set),goog.exportProperty(ol.geom.Geometry.prototype,"setProperties",ol.geom.Geometry.prototype.setProperties),goog.exportProperty(ol.geom.Geometry.prototype,"unset",ol.geom.Geometry.prototype.unset),
 goog.exportProperty(ol.geom.Geometry.prototype,"changed",ol.geom.Geometry.prototype.changed),goog.exportProperty(ol.geom.Geometry.prototype,"dispatchEvent",ol.geom.Geometry.prototype.dispatchEvent),goog.exportProperty(ol.geom.Geometry.prototype,"getRevision",ol.geom.Geometry.prototype.getRevision),goog.exportProperty(ol.geom.Geometry.prototype,"on",ol.geom.Geometry.prototype.on),goog.exportProperty(ol.geom.Geometry.prototype,"once",ol.geom.Geometry.prototype.once),goog.exportProperty(ol.geom.Geometry.prototype,"un",ol.geom.Geometry.prototype.un),goog.exportProperty(ol.geom.Geometry.prototype,"unByKey",ol.geom.Geometry.prototype.unByKey),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"getClosestPoint",ol.geom.SimpleGeometry.prototype.getClosestPoint),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"intersectsCoordinate",ol.geom.SimpleGeometry.prototype.intersectsCoordinate),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"getExtent",ol.geom.SimpleGeometry.prototype.getExtent),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"rotate",ol.geom.SimpleGeometry.prototype.rotate),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"scale",ol.geom.SimpleGeometry.prototype.scale),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"simplify",ol.geom.SimpleGeometry.prototype.simplify),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"transform",ol.geom.SimpleGeometry.prototype.transform),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"get",ol.geom.SimpleGeometry.prototype.get),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"getKeys",ol.geom.SimpleGeometry.prototype.getKeys),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"getProperties",ol.geom.SimpleGeometry.prototype.getProperties),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"set",ol.geom.SimpleGeometry.prototype.set),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"setProperties",ol.geom.SimpleGeometry.prototype.setProperties),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"unset",ol.geom.SimpleGeometry.prototype.unset),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"changed",ol.geom.SimpleGeometry.prototype.changed),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"dispatchEvent",ol.geom.SimpleGeometry.prototype.dispatchEvent),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"getRevision",ol.geom.SimpleGeometry.prototype.getRevision),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"on",ol.geom.SimpleGeometry.prototype.on),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"once",ol.geom.SimpleGeometry.prototype.once),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"un",ol.geom.SimpleGeometry.prototype.un),goog.exportProperty(ol.geom.SimpleGeometry.prototype,"unByKey",ol.geom.SimpleGeometry.prototype.unByKey),goog.exportProperty(ol.geom.Circle.prototype,"getFirstCoordinate",ol.geom.Circle.prototype.getFirstCoordinate),goog.exportProperty(ol.geom.Circle.prototype,"getLastCoordinate",ol.geom.Circle.prototype.getLastCoordinate),goog.exportProperty(ol.geom.Circle.prototype,"getLayout",ol.geom.Circle.prototype.getLayout),goog.exportProperty(ol.geom.Circle.prototype,"rotate",ol.geom.Circle.prototype.rotate),goog.exportProperty(ol.geom.Circle.prototype,"scale",ol.geom.Circle.prototype.scale),goog.exportProperty(ol.geom.Circle.prototype,"getClosestPoint",ol.geom.Circle.prototype.getClosestPoint),goog.exportProperty(ol.geom.Circle.prototype,"intersectsCoordinate",ol.geom.Circle.prototype.intersectsCoordinate),goog.exportProperty(ol.geom.Circle.prototype,"getExtent",ol.geom.Circle.prototype.getExtent),goog.exportProperty(ol.geom.Circle.prototype,"simplify",ol.geom.Circle.prototype.simplify),goog.exportProperty(ol.geom.Circle.prototype,"get",ol.geom.Circle.prototype.get),goog.exportProperty(ol.geom.Circle.prototype,"getKeys",ol.geom.Circle.prototype.getKeys),goog.exportProperty(ol.geom.Circle.prototype,"getProperties",ol.geom.Circle.prototype.getProperties),goog.exportProperty(ol.geom.Circle.prototype,"set",ol.geom.Circle.prototype.set),goog.exportProperty(ol.geom.Circle.prototype,"setProperties",ol.geom.Circle.prototype.setProperties),goog.exportProperty(ol.geom.Circle.prototype,"unset",ol.geom.Circle.prototype.unset),goog.exportProperty(ol.geom.Circle.prototype,"changed",ol.geom.Circle.prototype.changed),goog.exportProperty(ol.geom.Circle.prototype,"dispatchEvent",ol.geom.Circle.prototype.dispatchEvent),goog.exportProperty(ol.geom.Circle.prototype,"getRevision",ol.geom.Circle.prototype.getRevision),goog.exportProperty(ol.geom.Circle.prototype,"on",ol.geom.Circle.prototype.on),goog.exportProperty(ol.geom.Circle.prototype,"once",ol.geom.Circle.prototype.once),goog.exportProperty(ol.geom.Circle.prototype,"un",ol.geom.Circle.prototype.un),goog.exportProperty(ol.geom.Circle.prototype,"unByKey",ol.geom.Circle.prototype.unByKey),goog.exportProperty(ol.geom.GeometryCollection.prototype,"getClosestPoint",ol.geom.GeometryCollection.prototype.getClosestPoint),goog.exportProperty(ol.geom.GeometryCollection.prototype,"intersectsCoordinate",ol.geom.GeometryCollection.prototype.intersectsCoordinate),goog.exportProperty(ol.geom.GeometryCollection.prototype,"getExtent",ol.geom.GeometryCollection.prototype.getExtent),goog.exportProperty(ol.geom.GeometryCollection.prototype,"rotate",ol.geom.GeometryCollection.prototype.rotate),goog.exportProperty(ol.geom.GeometryCollection.prototype,"scale",ol.geom.GeometryCollection.prototype.scale),goog.exportProperty(ol.geom.GeometryCollection.prototype,"simplify",ol.geom.GeometryCollection.prototype.simplify),goog.exportProperty(ol.geom.GeometryCollection.prototype,"transform",ol.geom.GeometryCollection.prototype.transform),goog.exportProperty(ol.geom.GeometryCollection.prototype,"get",ol.geom.GeometryCollection.prototype.get),goog.exportProperty(ol.geom.GeometryCollection.prototype,"getKeys",ol.geom.GeometryCollection.prototype.getKeys),goog.exportProperty(ol.geom.GeometryCollection.prototype,"getProperties",ol.geom.GeometryCollection.prototype.getProperties),goog.exportProperty(ol.geom.GeometryCollection.prototype,"set",ol.geom.GeometryCollection.prototype.set),goog.exportProperty(ol.geom.GeometryCollection.prototype,"setProperties",ol.geom.GeometryCollection.prototype.setProperties),goog.exportProperty(ol.geom.GeometryCollection.prototype,"unset",ol.geom.GeometryCollection.prototype.unset),goog.exportProperty(ol.geom.GeometryCollection.prototype,"changed",ol.geom.GeometryCollection.prototype.changed),goog.exportProperty(ol.geom.GeometryCollection.prototype,"dispatchEvent",ol.geom.GeometryCollection.prototype.dispatchEvent),goog.exportProperty(ol.geom.GeometryCollection.prototype,"getRevision",ol.geom.GeometryCollection.prototype.getRevision),goog.exportProperty(ol.geom.GeometryCollection.prototype,"on",ol.geom.GeometryCollection.prototype.on),goog.exportProperty(ol.geom.GeometryCollection.prototype,"once",ol.geom.GeometryCollection.prototype.once),goog.exportProperty(ol.geom.GeometryCollection.prototype,"un",ol.geom.GeometryCollection.prototype.un),goog.exportProperty(ol.geom.GeometryCollection.prototype,"unByKey",ol.geom.GeometryCollection.prototype.unByKey),goog.exportProperty(ol.geom.LinearRing.prototype,"getFirstCoordinate",ol.geom.LinearRing.prototype.getFirstCoordinate),goog.exportProperty(ol.geom.LinearRing.prototype,"getLastCoordinate",ol.geom.LinearRing.prototype.getLastCoordinate),goog.exportProperty(ol.geom.LinearRing.prototype,"getLayout",ol.geom.LinearRing.prototype.getLayout),goog.exportProperty(ol.geom.LinearRing.prototype,"rotate",ol.geom.LinearRing.prototype.rotate),goog.exportProperty(ol.geom.LinearRing.prototype,"scale",ol.geom.LinearRing.prototype.scale),goog.exportProperty(ol.geom.LinearRing.prototype,"getClosestPoint",ol.geom.LinearRing.prototype.getClosestPoint),goog.exportProperty(ol.geom.LinearRing.prototype,"intersectsCoordinate",ol.geom.LinearRing.prototype.intersectsCoordinate),goog.exportProperty(ol.geom.LinearRing.prototype,"getExtent",ol.geom.LinearRing.prototype.getExtent),goog.exportProperty(ol.geom.LinearRing.prototype,"simplify",ol.geom.LinearRing.prototype.simplify),goog.exportProperty(ol.geom.LinearRing.prototype,"transform",ol.geom.LinearRing.prototype.transform),goog.exportProperty(ol.geom.LinearRing.prototype,"get",ol.geom.LinearRing.prototype.get),goog.exportProperty(ol.geom.LinearRing.prototype,"getKeys",ol.geom.LinearRing.prototype.getKeys),goog.exportProperty(ol.geom.LinearRing.prototype,"getProperties",ol.geom.LinearRing.prototype.getProperties),goog.exportProperty(ol.geom.LinearRing.prototype,"set",ol.geom.LinearRing.prototype.set),goog.exportProperty(ol.geom.LinearRing.prototype,"setProperties",ol.geom.LinearRing.prototype.setProperties),goog.exportProperty(ol.geom.LinearRing.prototype,"unset",ol.geom.LinearRing.prototype.unset),goog.exportProperty(ol.geom.LinearRing.prototype,"changed",ol.geom.LinearRing.prototype.changed),goog.exportProperty(ol.geom.LinearRing.prototype,"dispatchEvent",ol.geom.LinearRing.prototype.dispatchEvent),goog.exportProperty(ol.geom.LinearRing.prototype,"getRevision",ol.geom.LinearRing.prototype.getRevision),goog.exportProperty(ol.geom.LinearRing.prototype,"on",ol.geom.LinearRing.prototype.on),goog.exportProperty(ol.geom.LinearRing.prototype,"once",ol.geom.LinearRing.prototype.once),goog.exportProperty(ol.geom.LinearRing.prototype,"un",ol.geom.LinearRing.prototype.un),goog.exportProperty(ol.geom.LinearRing.prototype,"unByKey",ol.geom.LinearRing.prototype.unByKey),goog.exportProperty(ol.geom.LineString.prototype,"getFirstCoordinate",ol.geom.LineString.prototype.getFirstCoordinate),goog.exportProperty(ol.geom.LineString.prototype,"getLastCoordinate",ol.geom.LineString.prototype.getLastCoordinate),goog.exportProperty(ol.geom.LineString.prototype,"getLayout",ol.geom.LineString.prototype.getLayout),goog.exportProperty(ol.geom.LineString.prototype,"rotate",ol.geom.LineString.prototype.rotate),goog.exportProperty(ol.geom.LineString.prototype,"scale",ol.geom.LineString.prototype.scale),goog.exportProperty(ol.geom.LineString.prototype,"getClosestPoint",ol.geom.LineString.prototype.getClosestPoint),goog.exportProperty(ol.geom.LineString.prototype,"intersectsCoordinate",ol.geom.LineString.prototype.intersectsCoordinate),goog.exportProperty(ol.geom.LineString.prototype,"getExtent",ol.geom.LineString.prototype.getExtent),goog.exportProperty(ol.geom.LineString.prototype,"simplify",ol.geom.LineString.prototype.simplify),goog.exportProperty(ol.geom.LineString.prototype,"transform",ol.geom.LineString.prototype.transform),goog.exportProperty(ol.geom.LineString.prototype,"get",ol.geom.LineString.prototype.get),goog.exportProperty(ol.geom.LineString.prototype,"getKeys",ol.geom.LineString.prototype.getKeys),goog.exportProperty(ol.geom.LineString.prototype,"getProperties",ol.geom.LineString.prototype.getProperties),goog.exportProperty(ol.geom.LineString.prototype,"set",ol.geom.LineString.prototype.set),goog.exportProperty(ol.geom.LineString.prototype,"setProperties",ol.geom.LineString.prototype.setProperties),goog.exportProperty(ol.geom.LineString.prototype,"unset",ol.geom.LineString.prototype.unset),goog.exportProperty(ol.geom.LineString.prototype,"changed",ol.geom.LineString.prototype.changed),goog.exportProperty(ol.geom.LineString.prototype,"dispatchEvent",ol.geom.LineString.prototype.dispatchEvent),goog.exportProperty(ol.geom.LineString.prototype,"getRevision",ol.geom.LineString.prototype.getRevision),goog.exportProperty(ol.geom.LineString.prototype,"on",ol.geom.LineString.prototype.on),goog.exportProperty(ol.geom.LineString.prototype,"once",ol.geom.LineString.prototype.once),goog.exportProperty(ol.geom.LineString.prototype,"un",ol.geom.LineString.prototype.un),goog.exportProperty(ol.geom.LineString.prototype,"unByKey",ol.geom.LineString.prototype.unByKey),goog.exportProperty(ol.geom.MultiLineString.prototype,"getFirstCoordinate",ol.geom.MultiLineString.prototype.getFirstCoordinate),goog.exportProperty(ol.geom.MultiLineString.prototype,"getLastCoordinate",ol.geom.MultiLineString.prototype.getLastCoordinate),goog.exportProperty(ol.geom.MultiLineString.prototype,"getLayout",ol.geom.MultiLineString.prototype.getLayout),goog.exportProperty(ol.geom.MultiLineString.prototype,"rotate",ol.geom.MultiLineString.prototype.rotate),goog.exportProperty(ol.geom.MultiLineString.prototype,"scale",ol.geom.MultiLineString.prototype.scale),goog.exportProperty(ol.geom.MultiLineString.prototype,"getClosestPoint",ol.geom.MultiLineString.prototype.getClosestPoint),goog.exportProperty(ol.geom.MultiLineString.prototype,"intersectsCoordinate",ol.geom.MultiLineString.prototype.intersectsCoordinate),goog.exportProperty(ol.geom.MultiLineString.prototype,"getExtent",ol.geom.MultiLineString.prototype.getExtent),goog.exportProperty(ol.geom.MultiLineString.prototype,"simplify",ol.geom.MultiLineString.prototype.simplify),goog.exportProperty(ol.geom.MultiLineString.prototype,"transform",ol.geom.MultiLineString.prototype.transform),goog.exportProperty(ol.geom.MultiLineString.prototype,"get",ol.geom.MultiLineString.prototype.get),goog.exportProperty(ol.geom.MultiLineString.prototype,"getKeys",ol.geom.MultiLineString.prototype.getKeys),goog.exportProperty(ol.geom.MultiLineString.prototype,"getProperties",ol.geom.MultiLineString.prototype.getProperties),goog.exportProperty(ol.geom.MultiLineString.prototype,"set",ol.geom.MultiLineString.prototype.set),goog.exportProperty(ol.geom.MultiLineString.prototype,"setProperties",ol.geom.MultiLineString.prototype.setProperties),goog.exportProperty(ol.geom.MultiLineString.prototype,"unset",ol.geom.MultiLineString.prototype.unset),goog.exportProperty(ol.geom.MultiLineString.prototype,"changed",ol.geom.MultiLineString.prototype.changed),goog.exportProperty(ol.geom.MultiLineString.prototype,"dispatchEvent",ol.geom.MultiLineString.prototype.dispatchEvent),goog.exportProperty(ol.geom.MultiLineString.prototype,"getRevision",ol.geom.MultiLineString.prototype.getRevision),goog.exportProperty(ol.geom.MultiLineString.prototype,"on",ol.geom.MultiLineString.prototype.on),goog.exportProperty(ol.geom.MultiLineString.prototype,"once",ol.geom.MultiLineString.prototype.once),goog.exportProperty(ol.geom.MultiLineString.prototype,"un",ol.geom.MultiLineString.prototype.un),goog.exportProperty(ol.geom.MultiLineString.prototype,"unByKey",ol.geom.MultiLineString.prototype.unByKey),goog.exportProperty(ol.geom.MultiPoint.prototype,"getFirstCoordinate",ol.geom.MultiPoint.prototype.getFirstCoordinate),goog.exportProperty(ol.geom.MultiPoint.prototype,"getLastCoordinate",ol.geom.MultiPoint.prototype.getLastCoordinate),goog.exportProperty(ol.geom.MultiPoint.prototype,"getLayout",ol.geom.MultiPoint.prototype.getLayout),goog.exportProperty(ol.geom.MultiPoint.prototype,"rotate",ol.geom.MultiPoint.prototype.rotate),goog.exportProperty(ol.geom.MultiPoint.prototype,"scale",ol.geom.MultiPoint.prototype.scale),goog.exportProperty(ol.geom.MultiPoint.prototype,"getClosestPoint",ol.geom.MultiPoint.prototype.getClosestPoint),goog.exportProperty(ol.geom.MultiPoint.prototype,"intersectsCoordinate",ol.geom.MultiPoint.prototype.intersectsCoordinate),goog.exportProperty(ol.geom.MultiPoint.prototype,"getExtent",ol.geom.MultiPoint.prototype.getExtent),goog.exportProperty(ol.geom.MultiPoint.prototype,"simplify",ol.geom.MultiPoint.prototype.simplify),goog.exportProperty(ol.geom.MultiPoint.prototype,"transform",ol.geom.MultiPoint.prototype.transform),goog.exportProperty(ol.geom.MultiPoint.prototype,"get",ol.geom.MultiPoint.prototype.get),goog.exportProperty(ol.geom.MultiPoint.prototype,"getKeys",ol.geom.MultiPoint.prototype.getKeys),goog.exportProperty(ol.geom.MultiPoint.prototype,"getProperties",ol.geom.MultiPoint.prototype.getProperties),goog.exportProperty(ol.geom.MultiPoint.prototype,"set",ol.geom.MultiPoint.prototype.set),goog.exportProperty(ol.geom.MultiPoint.prototype,"setProperties",ol.geom.MultiPoint.prototype.setProperties),goog.exportProperty(ol.geom.MultiPoint.prototype,"unset",ol.geom.MultiPoint.prototype.unset),goog.exportProperty(ol.geom.MultiPoint.prototype,"changed",ol.geom.MultiPoint.prototype.changed),goog.exportProperty(ol.geom.MultiPoint.prototype,"dispatchEvent",ol.geom.MultiPoint.prototype.dispatchEvent),goog.exportProperty(ol.geom.MultiPoint.prototype,"getRevision",ol.geom.MultiPoint.prototype.getRevision),goog.exportProperty(ol.geom.MultiPoint.prototype,"on",ol.geom.MultiPoint.prototype.on),goog.exportProperty(ol.geom.MultiPoint.prototype,"once",ol.geom.MultiPoint.prototype.once),goog.exportProperty(ol.geom.MultiPoint.prototype,"un",ol.geom.MultiPoint.prototype.un),goog.exportProperty(ol.geom.MultiPoint.prototype,"unByKey",ol.geom.MultiPoint.prototype.unByKey),goog.exportProperty(ol.geom.MultiPolygon.prototype,"getFirstCoordinate",ol.geom.MultiPolygon.prototype.getFirstCoordinate),goog.exportProperty(ol.geom.MultiPolygon.prototype,"getLastCoordinate",ol.geom.MultiPolygon.prototype.getLastCoordinate),goog.exportProperty(ol.geom.MultiPolygon.prototype,"getLayout",ol.geom.MultiPolygon.prototype.getLayout),goog.exportProperty(ol.geom.MultiPolygon.prototype,"rotate",ol.geom.MultiPolygon.prototype.rotate),goog.exportProperty(ol.geom.MultiPolygon.prototype,"scale",ol.geom.MultiPolygon.prototype.scale),goog.exportProperty(ol.geom.MultiPolygon.prototype,"getClosestPoint",ol.geom.MultiPolygon.prototype.getClosestPoint),goog.exportProperty(ol.geom.MultiPolygon.prototype,"intersectsCoordinate",ol.geom.MultiPolygon.prototype.intersectsCoordinate),goog.exportProperty(ol.geom.MultiPolygon.prototype,"getExtent",ol.geom.MultiPolygon.prototype.getExtent),goog.exportProperty(ol.geom.MultiPolygon.prototype,"simplify",ol.geom.MultiPolygon.prototype.simplify),goog.exportProperty(ol.geom.MultiPolygon.prototype,"transform",ol.geom.MultiPolygon.prototype.transform),goog.exportProperty(ol.geom.MultiPolygon.prototype,"get",ol.geom.MultiPolygon.prototype.get),goog.exportProperty(ol.geom.MultiPolygon.prototype,"getKeys",ol.geom.MultiPolygon.prototype.getKeys),goog.exportProperty(ol.geom.MultiPolygon.prototype,"getProperties",ol.geom.MultiPolygon.prototype.getProperties),goog.exportProperty(ol.geom.MultiPolygon.prototype,"set",ol.geom.MultiPolygon.prototype.set),goog.exportProperty(ol.geom.MultiPolygon.prototype,"setProperties",ol.geom.MultiPolygon.prototype.setProperties),goog.exportProperty(ol.geom.MultiPolygon.prototype,"unset",ol.geom.MultiPolygon.prototype.unset),goog.exportProperty(ol.geom.MultiPolygon.prototype,"changed",ol.geom.MultiPolygon.prototype.changed),goog.exportProperty(ol.geom.MultiPolygon.prototype,"dispatchEvent",ol.geom.MultiPolygon.prototype.dispatchEvent),goog.exportProperty(ol.geom.MultiPolygon.prototype,"getRevision",ol.geom.MultiPolygon.prototype.getRevision),goog.exportProperty(ol.geom.MultiPolygon.prototype,"on",ol.geom.MultiPolygon.prototype.on),goog.exportProperty(ol.geom.MultiPolygon.prototype,"once",ol.geom.MultiPolygon.prototype.once),goog.exportProperty(ol.geom.MultiPolygon.prototype,"un",ol.geom.MultiPolygon.prototype.un),goog.exportProperty(ol.geom.MultiPolygon.prototype,"unByKey",ol.geom.MultiPolygon.prototype.unByKey),goog.exportProperty(ol.geom.Point.prototype,"getFirstCoordinate",ol.geom.Point.prototype.getFirstCoordinate),goog.exportProperty(ol.geom.Point.prototype,"getLastCoordinate",ol.geom.Point.prototype.getLastCoordinate),goog.exportProperty(ol.geom.Point.prototype,"getLayout",ol.geom.Point.prototype.getLayout),goog.exportProperty(ol.geom.Point.prototype,"rotate",ol.geom.Point.prototype.rotate),goog.exportProperty(ol.geom.Point.prototype,"scale",ol.geom.Point.prototype.scale),goog.exportProperty(ol.geom.Point.prototype,"getClosestPoint",ol.geom.Point.prototype.getClosestPoint),goog.exportProperty(ol.geom.Point.prototype,"intersectsCoordinate",ol.geom.Point.prototype.intersectsCoordinate),goog.exportProperty(ol.geom.Point.prototype,"getExtent",ol.geom.Point.prototype.getExtent),goog.exportProperty(ol.geom.Point.prototype,"simplify",ol.geom.Point.prototype.simplify),goog.exportProperty(ol.geom.Point.prototype,"transform",ol.geom.Point.prototype.transform),goog.exportProperty(ol.geom.Point.prototype,"get",ol.geom.Point.prototype.get),goog.exportProperty(ol.geom.Point.prototype,"getKeys",ol.geom.Point.prototype.getKeys),goog.exportProperty(ol.geom.Point.prototype,"getProperties",ol.geom.Point.prototype.getProperties),goog.exportProperty(ol.geom.Point.prototype,"set",ol.geom.Point.prototype.set),goog.exportProperty(ol.geom.Point.prototype,"setProperties",ol.geom.Point.prototype.setProperties),goog.exportProperty(ol.geom.Point.prototype,"unset",ol.geom.Point.prototype.unset),goog.exportProperty(ol.geom.Point.prototype,"changed",ol.geom.Point.prototype.changed),goog.exportProperty(ol.geom.Point.prototype,"dispatchEvent",ol.geom.Point.prototype.dispatchEvent),goog.exportProperty(ol.geom.Point.prototype,"getRevision",ol.geom.Point.prototype.getRevision),goog.exportProperty(ol.geom.Point.prototype,"on",ol.geom.Point.prototype.on),goog.exportProperty(ol.geom.Point.prototype,"once",ol.geom.Point.prototype.once),goog.exportProperty(ol.geom.Point.prototype,"un",ol.geom.Point.prototype.un),goog.exportProperty(ol.geom.Point.prototype,"unByKey",ol.geom.Point.prototype.unByKey),goog.exportProperty(ol.geom.Polygon.prototype,"getFirstCoordinate",ol.geom.Polygon.prototype.getFirstCoordinate),goog.exportProperty(ol.geom.Polygon.prototype,"getLastCoordinate",ol.geom.Polygon.prototype.getLastCoordinate),goog.exportProperty(ol.geom.Polygon.prototype,"getLayout",ol.geom.Polygon.prototype.getLayout),goog.exportProperty(ol.geom.Polygon.prototype,"rotate",ol.geom.Polygon.prototype.rotate),goog.exportProperty(ol.geom.Polygon.prototype,"scale",ol.geom.Polygon.prototype.scale),goog.exportProperty(ol.geom.Polygon.prototype,"getClosestPoint",ol.geom.Polygon.prototype.getClosestPoint),goog.exportProperty(ol.geom.Polygon.prototype,"intersectsCoordinate",ol.geom.Polygon.prototype.intersectsCoordinate),goog.exportProperty(ol.geom.Polygon.prototype,"getExtent",ol.geom.Polygon.prototype.getExtent),goog.exportProperty(ol.geom.Polygon.prototype,"simplify",ol.geom.Polygon.prototype.simplify),goog.exportProperty(ol.geom.Polygon.prototype,"transform",ol.geom.Polygon.prototype.transform),goog.exportProperty(ol.geom.Polygon.prototype,"get",ol.geom.Polygon.prototype.get),goog.exportProperty(ol.geom.Polygon.prototype,"getKeys",ol.geom.Polygon.prototype.getKeys),goog.exportProperty(ol.geom.Polygon.prototype,"getProperties",ol.geom.Polygon.prototype.getProperties),goog.exportProperty(ol.geom.Polygon.prototype,"set",ol.geom.Polygon.prototype.set),goog.exportProperty(ol.geom.Polygon.prototype,"setProperties",ol.geom.Polygon.prototype.setProperties),goog.exportProperty(ol.geom.Polygon.prototype,"unset",ol.geom.Polygon.prototype.unset),goog.exportProperty(ol.geom.Polygon.prototype,"changed",ol.geom.Polygon.prototype.changed),goog.exportProperty(ol.geom.Polygon.prototype,"dispatchEvent",ol.geom.Polygon.prototype.dispatchEvent),goog.exportProperty(ol.geom.Polygon.prototype,"getRevision",ol.geom.Polygon.prototype.getRevision),goog.exportProperty(ol.geom.Polygon.prototype,"on",ol.geom.Polygon.prototype.on),goog.exportProperty(ol.geom.Polygon.prototype,"once",ol.geom.Polygon.prototype.once),goog.exportProperty(ol.geom.Polygon.prototype,"un",ol.geom.Polygon.prototype.un),goog.exportProperty(ol.geom.Polygon.prototype,"unByKey",ol.geom.Polygon.prototype.unByKey),goog.exportProperty(ol.format.GML.prototype,"readFeatures",ol.format.GML.prototype.readFeatures),goog.exportProperty(ol.format.GML2.prototype,"readFeatures",ol.format.GML2.prototype.readFeatures),goog.exportProperty(ol.format.GML3.prototype,"readFeatures",ol.format.GML3.prototype.readFeatures),goog.exportProperty(ol.control.Control.prototype,"get",ol.control.Control.prototype.get),goog.exportProperty(ol.control.Control.prototype,"getKeys",ol.control.Control.prototype.getKeys),goog.exportProperty(ol.control.Control.prototype,"getProperties",ol.control.Control.prototype.getProperties),goog.exportProperty(ol.control.Control.prototype,"set",ol.control.Control.prototype.set),goog.exportProperty(ol.control.Control.prototype,"setProperties",ol.control.Control.prototype.setProperties),goog.exportProperty(ol.control.Control.prototype,"unset",ol.control.Control.prototype.unset),goog.exportProperty(ol.control.Control.prototype,"changed",ol.control.Control.prototype.changed),goog.exportProperty(ol.control.Control.prototype,"dispatchEvent",ol.control.Control.prototype.dispatchEvent),goog.exportProperty(ol.control.Control.prototype,"getRevision",ol.control.Control.prototype.getRevision),goog.exportProperty(ol.control.Control.prototype,"on",ol.control.Control.prototype.on),goog.exportProperty(ol.control.Control.prototype,"once",ol.control.Control.prototype.once),goog.exportProperty(ol.control.Control.prototype,"un",ol.control.Control.prototype.un),goog.exportProperty(ol.control.Control.prototype,"unByKey",ol.control.Control.prototype.unByKey),goog.exportProperty(ol.control.Attribution.prototype,"getMap",ol.control.Attribution.prototype.getMap),goog.exportProperty(ol.control.Attribution.prototype,"setMap",ol.control.Attribution.prototype.setMap),goog.exportProperty(ol.control.Attribution.prototype,"setTarget",ol.control.Attribution.prototype.setTarget),goog.exportProperty(ol.control.Attribution.prototype,"get",ol.control.Attribution.prototype.get),goog.exportProperty(ol.control.Attribution.prototype,"getKeys",ol.control.Attribution.prototype.getKeys),goog.exportProperty(ol.control.Attribution.prototype,"getProperties",ol.control.Attribution.prototype.getProperties),goog.exportProperty(ol.control.Attribution.prototype,"set",ol.control.Attribution.prototype.set),goog.exportProperty(ol.control.Attribution.prototype,"setProperties",ol.control.Attribution.prototype.setProperties),goog.exportProperty(ol.control.Attribution.prototype,"unset",ol.control.Attribution.prototype.unset),goog.exportProperty(ol.control.Attribution.prototype,"changed",ol.control.Attribution.prototype.changed),goog.exportProperty(ol.control.Attribution.prototype,"dispatchEvent",ol.control.Attribution.prototype.dispatchEvent),goog.exportProperty(ol.control.Attribution.prototype,"getRevision",ol.control.Attribution.prototype.getRevision),goog.exportProperty(ol.control.Attribution.prototype,"on",ol.control.Attribution.prototype.on),goog.exportProperty(ol.control.Attribution.prototype,"once",ol.control.Attribution.prototype.once),goog.exportProperty(ol.control.Attribution.prototype,"un",ol.control.Attribution.prototype.un),goog.exportProperty(ol.control.Attribution.prototype,"unByKey",ol.control.Attribution.prototype.unByKey),goog.exportProperty(ol.control.FullScreen.prototype,"getMap",ol.control.FullScreen.prototype.getMap),goog.exportProperty(ol.control.FullScreen.prototype,"setMap",ol.control.FullScreen.prototype.setMap),goog.exportProperty(ol.control.FullScreen.prototype,"setTarget",ol.control.FullScreen.prototype.setTarget),goog.exportProperty(ol.control.FullScreen.prototype,"get",ol.control.FullScreen.prototype.get),goog.exportProperty(ol.control.FullScreen.prototype,"getKeys",ol.control.FullScreen.prototype.getKeys),goog.exportProperty(ol.control.FullScreen.prototype,"getProperties",ol.control.FullScreen.prototype.getProperties),goog.exportProperty(ol.control.FullScreen.prototype,"set",ol.control.FullScreen.prototype.set),goog.exportProperty(ol.control.FullScreen.prototype,"setProperties",ol.control.FullScreen.prototype.setProperties),goog.exportProperty(ol.control.FullScreen.prototype,"unset",ol.control.FullScreen.prototype.unset),goog.exportProperty(ol.control.FullScreen.prototype,"changed",ol.control.FullScreen.prototype.changed),goog.exportProperty(ol.control.FullScreen.prototype,"dispatchEvent",ol.control.FullScreen.prototype.dispatchEvent),goog.exportProperty(ol.control.FullScreen.prototype,"getRevision",ol.control.FullScreen.prototype.getRevision),goog.exportProperty(ol.control.FullScreen.prototype,"on",ol.control.FullScreen.prototype.on),goog.exportProperty(ol.control.FullScreen.prototype,"once",ol.control.FullScreen.prototype.once),goog.exportProperty(ol.control.FullScreen.prototype,"un",ol.control.FullScreen.prototype.un),goog.exportProperty(ol.control.FullScreen.prototype,"unByKey",ol.control.FullScreen.prototype.unByKey),goog.exportProperty(ol.control.MousePosition.prototype,"getMap",ol.control.MousePosition.prototype.getMap),goog.exportProperty(ol.control.MousePosition.prototype,"setMap",ol.control.MousePosition.prototype.setMap),goog.exportProperty(ol.control.MousePosition.prototype,"setTarget",ol.control.MousePosition.prototype.setTarget),goog.exportProperty(ol.control.MousePosition.prototype,"get",ol.control.MousePosition.prototype.get),goog.exportProperty(ol.control.MousePosition.prototype,"getKeys",ol.control.MousePosition.prototype.getKeys),goog.exportProperty(ol.control.MousePosition.prototype,"getProperties",ol.control.MousePosition.prototype.getProperties),goog.exportProperty(ol.control.MousePosition.prototype,"set",ol.control.MousePosition.prototype.set),goog.exportProperty(ol.control.MousePosition.prototype,"setProperties",ol.control.MousePosition.prototype.setProperties),goog.exportProperty(ol.control.MousePosition.prototype,"unset",ol.control.MousePosition.prototype.unset),goog.exportProperty(ol.control.MousePosition.prototype,"changed",ol.control.MousePosition.prototype.changed),goog.exportProperty(ol.control.MousePosition.prototype,"dispatchEvent",ol.control.MousePosition.prototype.dispatchEvent),goog.exportProperty(ol.control.MousePosition.prototype,"getRevision",ol.control.MousePosition.prototype.getRevision),goog.exportProperty(ol.control.MousePosition.prototype,"on",ol.control.MousePosition.prototype.on),goog.exportProperty(ol.control.MousePosition.prototype,"once",ol.control.MousePosition.prototype.once),goog.exportProperty(ol.control.MousePosition.prototype,"un",ol.control.MousePosition.prototype.un),goog.exportProperty(ol.control.MousePosition.prototype,"unByKey",ol.control.MousePosition.prototype.unByKey),goog.exportProperty(ol.control.OverviewMap.prototype,"getMap",ol.control.OverviewMap.prototype.getMap),goog.exportProperty(ol.control.OverviewMap.prototype,"setMap",ol.control.OverviewMap.prototype.setMap),goog.exportProperty(ol.control.OverviewMap.prototype,"setTarget",ol.control.OverviewMap.prototype.setTarget),goog.exportProperty(ol.control.OverviewMap.prototype,"get",ol.control.OverviewMap.prototype.get),goog.exportProperty(ol.control.OverviewMap.prototype,"getKeys",ol.control.OverviewMap.prototype.getKeys),goog.exportProperty(ol.control.OverviewMap.prototype,"getProperties",ol.control.OverviewMap.prototype.getProperties),goog.exportProperty(ol.control.OverviewMap.prototype,"set",ol.control.OverviewMap.prototype.set),goog.exportProperty(ol.control.OverviewMap.prototype,"setProperties",ol.control.OverviewMap.prototype.setProperties),goog.exportProperty(ol.control.OverviewMap.prototype,"unset",ol.control.OverviewMap.prototype.unset),goog.exportProperty(ol.control.OverviewMap.prototype,"changed",ol.control.OverviewMap.prototype.changed),goog.exportProperty(ol.control.OverviewMap.prototype,"dispatchEvent",ol.control.OverviewMap.prototype.dispatchEvent),goog.exportProperty(ol.control.OverviewMap.prototype,"getRevision",ol.control.OverviewMap.prototype.getRevision),goog.exportProperty(ol.control.OverviewMap.prototype,"on",ol.control.OverviewMap.prototype.on),goog.exportProperty(ol.control.OverviewMap.prototype,"once",ol.control.OverviewMap.prototype.once),goog.exportProperty(ol.control.OverviewMap.prototype,"un",ol.control.OverviewMap.prototype.un),goog.exportProperty(ol.control.OverviewMap.prototype,"unByKey",ol.control.OverviewMap.prototype.unByKey),goog.exportProperty(ol.control.Rotate.prototype,"getMap",ol.control.Rotate.prototype.getMap),goog.exportProperty(ol.control.Rotate.prototype,"setMap",ol.control.Rotate.prototype.setMap),goog.exportProperty(ol.control.Rotate.prototype,"setTarget",ol.control.Rotate.prototype.setTarget),goog.exportProperty(ol.control.Rotate.prototype,"get",ol.control.Rotate.prototype.get),
-goog.exportProperty(ol.control.Rotate.prototype,"getKeys",ol.control.Rotate.prototype.getKeys),goog.exportProperty(ol.control.Rotate.prototype,"getProperties",ol.control.Rotate.prototype.getProperties),goog.exportProperty(ol.control.Rotate.prototype,"set",ol.control.Rotate.prototype.set),goog.exportProperty(ol.control.Rotate.prototype,"setProperties",ol.control.Rotate.prototype.setProperties),goog.exportProperty(ol.control.Rotate.prototype,"unset",ol.control.Rotate.prototype.unset),goog.exportProperty(ol.control.Rotate.prototype,"changed",ol.control.Rotate.prototype.changed),goog.exportProperty(ol.control.Rotate.prototype,"dispatchEvent",ol.control.Rotate.prototype.dispatchEvent),goog.exportProperty(ol.control.Rotate.prototype,"getRevision",ol.control.Rotate.prototype.getRevision),goog.exportProperty(ol.control.Rotate.prototype,"on",ol.control.Rotate.prototype.on),goog.exportProperty(ol.control.Rotate.prototype,"once",ol.control.Rotate.prototype.once),goog.exportProperty(ol.control.Rotate.prototype,"un",ol.control.Rotate.prototype.un),goog.exportProperty(ol.control.Rotate.prototype,"unByKey",ol.control.Rotate.prototype.unByKey),goog.exportProperty(ol.control.ScaleLine.prototype,"getMap",ol.control.ScaleLine.prototype.getMap),goog.exportProperty(ol.control.ScaleLine.prototype,"setMap",ol.control.ScaleLine.prototype.setMap),goog.exportProperty(ol.control.ScaleLine.prototype,"setTarget",ol.control.ScaleLine.prototype.setTarget),goog.exportProperty(ol.control.ScaleLine.prototype,"get",ol.control.ScaleLine.prototype.get),goog.exportProperty(ol.control.ScaleLine.prototype,"getKeys",ol.control.ScaleLine.prototype.getKeys),goog.exportProperty(ol.control.ScaleLine.prototype,"getProperties",ol.control.ScaleLine.prototype.getProperties),goog.exportProperty(ol.control.ScaleLine.prototype,"set",ol.control.ScaleLine.prototype.set),goog.exportProperty(ol.control.ScaleLine.prototype,"setProperties",ol.control.ScaleLine.prototype.setProperties),goog.exportProperty(ol.control.ScaleLine.prototype,"unset",ol.control.ScaleLine.prototype.unset),goog.exportProperty(ol.control.ScaleLine.prototype,"changed",ol.control.ScaleLine.prototype.changed),goog.exportProperty(ol.control.ScaleLine.prototype,"dispatchEvent",ol.control.ScaleLine.prototype.dispatchEvent),goog.exportProperty(ol.control.ScaleLine.prototype,"getRevision",ol.control.ScaleLine.prototype.getRevision),goog.exportProperty(ol.control.ScaleLine.prototype,"on",ol.control.ScaleLine.prototype.on),goog.exportProperty(ol.control.ScaleLine.prototype,"once",ol.control.ScaleLine.prototype.once),goog.exportProperty(ol.control.ScaleLine.prototype,"un",ol.control.ScaleLine.prototype.un),goog.exportProperty(ol.control.ScaleLine.prototype,"unByKey",ol.control.ScaleLine.prototype.unByKey),goog.exportProperty(ol.control.Zoom.prototype,"getMap",ol.control.Zoom.prototype.getMap),goog.exportProperty(ol.control.Zoom.prototype,"setMap",ol.control.Zoom.prototype.setMap),goog.exportProperty(ol.control.Zoom.prototype,"setTarget",ol.control.Zoom.prototype.setTarget),goog.exportProperty(ol.control.Zoom.prototype,"get",ol.control.Zoom.prototype.get),goog.exportProperty(ol.control.Zoom.prototype,"getKeys",ol.control.Zoom.prototype.getKeys),goog.exportProperty(ol.control.Zoom.prototype,"getProperties",ol.control.Zoom.prototype.getProperties),goog.exportProperty(ol.control.Zoom.prototype,"set",ol.control.Zoom.prototype.set),goog.exportProperty(ol.control.Zoom.prototype,"setProperties",ol.control.Zoom.prototype.setProperties),goog.exportProperty(ol.control.Zoom.prototype,"unset",ol.control.Zoom.prototype.unset),goog.exportProperty(ol.control.Zoom.prototype,"changed",ol.control.Zoom.prototype.changed),goog.exportProperty(ol.control.Zoom.prototype,"dispatchEvent",ol.control.Zoom.prototype.dispatchEvent),goog.exportProperty(ol.control.Zoom.prototype,"getRevision",ol.control.Zoom.prototype.getRevision),goog.exportProperty(ol.control.Zoom.prototype,"on",ol.control.Zoom.prototype.on),goog.exportProperty(ol.control.Zoom.prototype,"once",ol.control.Zoom.prototype.once),goog.exportProperty(ol.control.Zoom.prototype,"un",ol.control.Zoom.prototype.un),goog.exportProperty(ol.control.Zoom.prototype,"unByKey",ol.control.Zoom.prototype.unByKey),goog.exportProperty(ol.control.ZoomSlider.prototype,"getMap",ol.control.ZoomSlider.prototype.getMap),goog.exportProperty(ol.control.ZoomSlider.prototype,"setMap",ol.control.ZoomSlider.prototype.setMap),goog.exportProperty(ol.control.ZoomSlider.prototype,"setTarget",ol.control.ZoomSlider.prototype.setTarget),goog.exportProperty(ol.control.ZoomSlider.prototype,"get",ol.control.ZoomSlider.prototype.get),goog.exportProperty(ol.control.ZoomSlider.prototype,"getKeys",ol.control.ZoomSlider.prototype.getKeys),goog.exportProperty(ol.control.ZoomSlider.prototype,"getProperties",ol.control.ZoomSlider.prototype.getProperties),goog.exportProperty(ol.control.ZoomSlider.prototype,"set",ol.control.ZoomSlider.prototype.set),goog.exportProperty(ol.control.ZoomSlider.prototype,"setProperties",ol.control.ZoomSlider.prototype.setProperties),goog.exportProperty(ol.control.ZoomSlider.prototype,"unset",ol.control.ZoomSlider.prototype.unset),goog.exportProperty(ol.control.ZoomSlider.prototype,"changed",ol.control.ZoomSlider.prototype.changed),goog.exportProperty(ol.control.ZoomSlider.prototype,"dispatchEvent",ol.control.ZoomSlider.prototype.dispatchEvent),goog.exportProperty(ol.control.ZoomSlider.prototype,"getRevision",ol.control.ZoomSlider.prototype.getRevision),goog.exportProperty(ol.control.ZoomSlider.prototype,"on",ol.control.ZoomSlider.prototype.on),goog.exportProperty(ol.control.ZoomSlider.prototype,"once",ol.control.ZoomSlider.prototype.once),goog.exportProperty(ol.control.ZoomSlider.prototype,"un",ol.control.ZoomSlider.prototype.un),goog.exportProperty(ol.control.ZoomSlider.prototype,"unByKey",ol.control.ZoomSlider.prototype.unByKey),goog.exportProperty(ol.control.ZoomToExtent.prototype,"getMap",ol.control.ZoomToExtent.prototype.getMap),goog.exportProperty(ol.control.ZoomToExtent.prototype,"setMap",ol.control.ZoomToExtent.prototype.setMap),goog.exportProperty(ol.control.ZoomToExtent.prototype,"setTarget",ol.control.ZoomToExtent.prototype.setTarget),goog.exportProperty(ol.control.ZoomToExtent.prototype,"get",ol.control.ZoomToExtent.prototype.get),goog.exportProperty(ol.control.ZoomToExtent.prototype,"getKeys",ol.control.ZoomToExtent.prototype.getKeys),goog.exportProperty(ol.control.ZoomToExtent.prototype,"getProperties",ol.control.ZoomToExtent.prototype.getProperties),goog.exportProperty(ol.control.ZoomToExtent.prototype,"set",ol.control.ZoomToExtent.prototype.set),goog.exportProperty(ol.control.ZoomToExtent.prototype,"setProperties",ol.control.ZoomToExtent.prototype.setProperties),goog.exportProperty(ol.control.ZoomToExtent.prototype,"unset",ol.control.ZoomToExtent.prototype.unset),goog.exportProperty(ol.control.ZoomToExtent.prototype,"changed",ol.control.ZoomToExtent.prototype.changed),goog.exportProperty(ol.control.ZoomToExtent.prototype,"dispatchEvent",ol.control.ZoomToExtent.prototype.dispatchEvent),goog.exportProperty(ol.control.ZoomToExtent.prototype,"getRevision",ol.control.ZoomToExtent.prototype.getRevision),goog.exportProperty(ol.control.ZoomToExtent.prototype,"on",ol.control.ZoomToExtent.prototype.on),goog.exportProperty(ol.control.ZoomToExtent.prototype,"once",ol.control.ZoomToExtent.prototype.once),goog.exportProperty(ol.control.ZoomToExtent.prototype,"un",ol.control.ZoomToExtent.prototype.un),goog.exportProperty(ol.control.ZoomToExtent.prototype,"unByKey",ol.control.ZoomToExtent.prototype.unByKey),ol.VERSION="v3.19.1",OPENLAYERS.ol=ol,OPENLAYERS.ol}),define("ol-custom",["ol3"],function(e){e.interaction.PinchZoom.handleUpEvent_=function(o){if(this.targetPointers.length<2){var t=o.map,r=t.getView();return r.setHint(e.View.Hint.INTERACTING,-1),!1}return!0},e.View.prototype.getDecimalZoom=function(){var e,o=this.getResolution();return e=Math.log(this.maxResolution_/o)/Math.log(2),void 0!==e?this.minZoom_+e:e},e.control.CustomControl=function(o){var t=o||{},r=document.createElement("button");r.setAttribute("type","button");var i=document.createElement("span");i.innerHTML=t.character,r.appendChild(i),e.events.listen(r,e.events.EventType.CLICK,t.callback,this);var n=document.createElement("div");n.className=t.cls+" ol-unselectable ol-control",n.appendChild(r),e.control.Control.call(this,{element:n,target:t.target,render:t.render})},e.inherits(e.control.CustomControl,e.control.Control),e.control.GoHome=function(o){var t=o||{};t.character='<i class="fa fa-home fa-lg"></i>',t.cls="home";var r=this;t.callback=function(){var e=r.getMap().getLayers().item(0).getSource();e.goHome()},e.control.CustomControl.call(this,t)},e.inherits(e.control.GoHome,e.control.CustomControl),e.control.SetGPS=function(o){var t=o||{};t.character='<i class="fa fa-crosshairs fa-lg"></i>',t.cls="gps";var r=this;t.render=function(e){var o=e.frameState;if(o){var t=this.getMap();if(t.geolocation){var i=t.geolocation.getTracking(),n=r.element.classList.contains("receiving_gps");n&&!i?this.element.classList.remove("receiving_gps"):!n&&i&&this.element.classList.add("receiving_gps")}}},t.callback=function(){var e=r.element.classList.contains("receiving_gps"),o=r.getMap();o.handleGPS(!e),e?r.element.classList.remove("receiving_gps"):r.element.classList.add("receiving_gps")},e.control.CustomControl.call(this,t)},e.inherits(e.control.SetGPS,e.control.CustomControl),e.control.CompassRotate=function(o){var t=o||{};t.autoHide=!1;var r=document.createElement("span");r.innerHTML='<i class="fa fa-compass fa-lg ol-compass-fa"></i>',t.label=r,t.render=function(e){var o=e.frameState;if(o){var t=this.getMap().getView(),r=o.viewState.rotation,i=t.getCenter(),n=t.getDecimalZoom();if(r!=this.rotation_||i[0]!=this.center_[0]||i[1]!=this.center_[1]||n!=this.zoom_){var l=this.element.classList.contains("no_rotation");l||0!==r?l&&0!==r&&this.element.classList.remove("no_rotation"):this.element.classList.add("no_rotation");var a=this,s=this.getMap().getLayers().item(0).getSource();s.size2MercsAsync().then(function(e){var o=s.mercs2MercRotation(e),t="rotate("+o+"rad)";a.label_.style.msTransform=t,a.label_.style.webkitTransform=t,a.label_.style.transform=t})}this.rotation_=r,this.center_=i,this.zoom_=n}},e.control.Rotate.call(this,t),this.center_=[],this.zoom_=void 0},e.inherits(e.control.CompassRotate,e.control.Rotate),e.const=e.const?e.const:{},e.const.MERC_MAX=20037508.342789244,e.const.MERC_CROSSMATRIX=[[0,0],[0,1],[1,0],[0,-1],[-1,0]];var o=new e.style.Style({image:new e.style.Icon({anchor:[.5,.5],anchorXUnits:"fraction",anchorYUnits:"fraction",src:"parts/bluedot.png"})}),t=new e.style.Style({fill:new e.style.Fill({color:[128,128,256,.2]}),stroke:new e.style.Stroke({color:[128,128,256,1],width:3})}),r=new e.style.Style({image:new e.style.Icon({anchor:[.5,1],anchorXUnits:"fraction",anchorYUnits:"fraction",src:"parts/defaultpin.png"})});return e.source.setCustomFunction=function(o){o.prototype.getMap=function(){return this._map?this._map:(this._map=new e.MaplatMap({source:this,div:this.map_option.div,center:this.map_option.default_center,zoom:this.map_option.default_zoom,rotation:this.map_option.default_rotation}),this._map)},o.prototype.goHome=function(){var o=e.proj.transform(this.home_position,"EPSG:4326","EPSG:3857"),t=this._map,r=t.getView(),i=this.mercsFromGivenZoom(o,this.merc_zoom);this.mercs2SizeAsync(i).then(function(e){r.setCenter(e[0]),r.setZoom(e[1]),r.setRotation(0)})},o.prototype.setGPSMarkerAsync=function(e,o){var t=this,r=t.getMap(),i=r.getView(),n=e?t.mercsFromGPSValue(e.lnglat,e.acc):["dummy"];return Promise.all(n.map(function(e,o){return 5==o||"dummy"==e?e:t.merc2XyAsync(e)})).then(function(e){var n=null;if("dummy"!=e[0]){if(n={xy:e[0]},!t.insideCheckHistMapCoords(e[0]))return r.handleGPS(!1,!0),!1;var l=e.slice(1);n.rad=l.reduce(function(e,o,t){var r=e+Math.sqrt(Math.pow(o[0]-n.xy[0],2)+Math.pow(o[1]-n.xy[1],2));return 3==t?r/4:r},0),o||i.setCenter(n.xy)}return r.setGPSPosition(n),!0})},o.prototype.setGPSMarker=function(e,o){this.setGPSMarkerAsync(e,o).then(function(){})},o.prototype.getRadius=function(o){var t=Math.floor(Math.min(o[0],o[1])/4),r=this._map.getView().getDecimalZoom();return t*e.const.MERC_MAX/128/Math.pow(2,r)},o.prototype.mercsFromGivenZoom=function(o,t){var r=this._map.getSize(),i=Math.floor(Math.min(r[0],r[1])/4),n=i*e.const.MERC_MAX/128/Math.pow(2,t);return e.const.MERC_CROSSMATRIX.map(function(e){return[e[0]*n+o[0],e[1]*n+o[1]]})},o.prototype.mercsFromGPSValue=function(o,t){var r=e.proj.transform(o,"EPSG:4326","EPSG:3857"),i=o[1]*Math.PI/180,n=t/Math.cos(i);return e.const.MERC_CROSSMATRIX.map(function(e){return[e[0]*n+r[0],e[1]*n+r[1]]})},o.prototype.rotateMatrix=function(e){for(var o=1*this._map.getView().getRotation(),t=[],r=0;r<e.length;r++){var i=e[r],n=i[0]*Math.cos(o)-i[1]*Math.sin(o),l=i[0]*Math.sin(o)+i[1]*Math.cos(o);t.push([n,l])}return t},o.prototype.size2Xys=function(o){o||(o=this._map.getView().getCenter());var t=this._map.getSize(),r=this.getRadius(t),i=this.rotateMatrix(e.const.MERC_CROSSMATRIX),n=i.map(function(e){return[e[0]*r+o[0],e[1]*r+o[1]]});return n.push(t),n},o.prototype.size2MercsAsync=function(e){var o=this.size2Xys(e),t=this,r=o.map(function(e,o){return 5==o?e:t.xy2MercAsync(e)});return Promise.all(r)},o.prototype.mercs2SizeAsync=function(e){var o=this,t=e.map(function(e,t){return 5==t?e:o.merc2XyAsync(e)});return Promise.all(t).then(function(e){return o.xys2Size(e)})},o.prototype.xys2Size=function(o){for(var t=o[0],r=o[5],i=o.slice(1,5),n=i.map(function(e){return[e[0]-t[0],e[1]-t[1]]}),l=[[0,1],[1,0],[0,-1],[-1,0]],a=0,s=0,p=0,g=0;g<4;g++){var c=n[g],u=l[g],y=Math.sqrt(Math.pow(c[0],2)+Math.pow(c[1],2));a+=y;var h=c[0]*u[1]-c[1]*u[0],d=Math.acos((c[0]*u[0]+c[1]*u[1])/y),m=h>0?-1*d:d;s+=Math.cos(m),p+=Math.sin(m)}var f=a/4,_=Math.atan2(p,s);r||(r=this._map.getSize());var E=Math.floor(Math.min(r[0],r[1])/4),v=Math.log(E*e.const.MERC_MAX/128/f)/Math.log(2);return[t,v,_]},o.prototype.mercs2MercRotation=function(e){for(var o=e[0],t=e.slice(1,5),r=t.map(function(e){return[e[0]-o[0],e[1]-o[1]]}),i=[[0,1],[1,0],[0,-1],[-1,0]],n=0,l=0,a=0;a<4;a++){var s=r[a],p=i[a],g=Math.sqrt(Math.pow(s[0],2)+Math.pow(s[1],2)),c=s[0]*p[1]-s[1]*p[0],u=Math.acos((s[0]*p[0]+s[1]*p[1])/g),y=c>0?-1*u:u;n+=Math.cos(y),l+=Math.sin(y)}return Math.atan2(l,n)}},e.source.setCustomInitialize=function(e,o){e.sourceID=o.sourceID,e.map_option=o.map_option||{},e.home_position=o.home_position,e.merc_zoom=o.merc_zoom,e.fake_gps=o.fake_gps||!1},e.source.NowMap=function(o){var t=o||{};e.source.OSM.call(this,t),e.source.setCustomInitialize(this,t)},e.inherits(e.source.NowMap,e.source.OSM),e.source.NowMap.createAsync=function(o){return new Promise(function(t,r){var i=new e.source.NowMap(o);t(i)})},e.source.setCustomFunction(e.source.NowMap),e.source.NowMap.prototype.xy2MercAsync=function(e){return new Promise(function(o,t){o(e)})},e.source.NowMap.prototype.merc2XyAsync=function(e){return new Promise(function(o,t){o(e)})},e.source.NowMap.prototype.insideCheckXy=function(e){return!0},e.source.NowMap.prototype.insideCheckHistMapCoords=function(e){return!0},e.source.TmsMap=function(o){var t=o||{};e.source.NowMap.call(this,t)},e.inherits(e.source.TmsMap,e.source.NowMap),e.source.TmsMap.createAsync=function(o){var t=new Promise(function(t,r){var i=new e.source.TmsMap(o);t(i)});return t},e.MaplatMap=function(o){o=o||{},this._gps_source=new e.source.Vector({wrapX:!1});var t=new e.layer.Vector({source:this._gps_source});this._marker_source=new e.source.Vector({wrapX:!1});var r=new e.layer.Vector({source:this._marker_source}),i=this._overlay_group=new e.layer.Group,n=o.off_control?[]:[new e.control.Attribution,new e.control.CompassRotate,new e.control.Zoom,new e.control.SetGPS,new e.control.GoHome],l={controls:n,layers:[new e.layer.Tile({source:o.source}),i,t,r],target:o.div,view:new e.View({center:o.default_center||[0,0],zoom:o.default_zoom||2,rotation:o.default_rotation||0})};e.Map.call(this,l);var a=this.getView(),s=this;s.__AvoidFirstMoveStart=!0;var p=function(){s.__AvoidFirstMoveStart||s.dispatchEvent("movestart"),s.__AvoidFirstMoveStart=!1,a.un("propertychange",p)};a.on("propertychange",p),s.on("moveend",function(){a.on("propertychange",p)})},e.inherits(e.MaplatMap,e.Map),e.MaplatMap.prototype.setGPSPosition=function(r){var i=this._gps_source;if(i.clear(),r){var n=new e.Feature({geometry:new e.geom.Point(r.xy)});n.setStyle(o);var l=new e.Feature({geometry:new e.geom.Circle(r.xy,r.rad)});l.setStyle(t),i.addFeature(n),i.addFeature(l)}},e.MaplatMap.prototype.resetMarker=function(){var e=this._marker_source;e.clear()},e.MaplatMap.prototype.setMarker=function(o,t,i){var n=this._marker_source;t.geometry=new e.geom.Point(o);var l=new e.Feature(t);i||(i=r),l.setStyle(i),n.addFeature(l)},e.MaplatMap.prototype.exchangeSource=function(e){var o=this.getLayers(),t=o.item(0);t.setSource(e),e&&(e._map=this)},e.MaplatMap.prototype.setLayer=function(o){var t=this._overlay_group.getLayers();if(t.clear(),o){var r=new e.layer.Tile({source:o});t.push(r)}},e.MaplatMap.prototype.getSource=function(){return this.getLayers().item(0).getSource()},e.MaplatMap.prototype.setOpacity=function(o){var t=(100-o)/100,r=this.getSource();r instanceof e.source.NowMap?(this.getLayers().item(0).setOpacity(1),this.getLayers().item(1).setOpacity(t)):this.getLayers().item(0).setOpacity(t)},e.MaplatMap.prototype.handleGPS=function(o,t){if(o)if(this.dispatchEvent("gps_request"),this._first_gps_request=!0,this.geolocation)this.geolocation.setTracking(!0);else{var r=this.geolocation=new e.Geolocation({tracking:!0}),i=this;r.on("change",function(o){var t=i.getLayers().item(0).getSource(),n=r.getPosition(),l=r.getAccuracy();t.fake_gps&&e.MathEx.getDistance(t.home_position,n)>t.fake_gps&&(n=[e.MathEx.randomFromCenter(t.home_position[0],.001),e.MathEx.randomFromCenter(t.home_position[1],.001)],l=e.MathEx.randomFromCenter(15,10));var a={lnglat:n,acc:l};t.setGPSMarkerAsync(a,!i._first_gps_request).then(function(o){o||(a={error:"gps_out"}),i._first_gps_request=!1,i.dispatchEvent(new e.MapEvent("gps_result",i,a))})}),r.on("error",function(o){var t=i.getLayers().item(0).getSource(),r=null;if(t.fake_gps){var n=[e.MathEx.randomFromCenter(t.home_position[0],.001),e.MathEx.randomFromCenter(t.home_position[1],.001)],l=e.MathEx.randomFromCenter(15,10);r={lnglat:n,acc:l}}t.setGPSMarkerAsync(r,!i._first_gps_request).then(function(o){o||(r={error:"gps_out"}),i._first_gps_request=!1,i.dispatchEvent(new e.MapEvent("gps_result",i,r))})})}else{this.geolocation&&this.geolocation.setTracking(!1);var n=this.getLayers().item(0).getSource();n.setGPSMarker(),t||this.dispatchEvent(new e.MapEvent("gps_result",i,{error:"gps_off"}))}},e.MathEx={},e.MathEx.randomFromCenter=function(e,o){return e+(Math.random()-.5)*o},e.MathEx.recursiveRound=function(o,t){if(o instanceof Array)return o.map(function(o){return e.MathEx.recursiveRound(o,t)});var r=Math.pow(10,t);return Math.round(o*r)/r},e.MathEx.getDistance=function(e,o){function t(e){return e*Math.PI/180}return 6378.14*Math.acos(Math.cos(t(e[1]))*Math.cos(t(o[1]))*Math.cos(t(o[0])-t(e[0]))+Math.sin(t(e[1]))*Math.sin(t(o[1])))},e}),define("histmap",["ol-custom"],function(e){var o="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAQAAAAEACAMAAABrrFhUAAAAB3RJTUUH3QgIBToaSbAjlwAAABd0RVh0U29mdHdhcmUAR0xEUE5HIHZlciAzLjRxhaThAAAACHRwTkdHTEQzAAAAAEqAKR8AAAAEZ0FNQQAAsY8L/GEFAAAAA1BMVEX///+nxBvIAAAAAXRSTlMAQObYZgAAAFRJREFUeNrtwQEBAAAAgJD+r+4ICgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABgBDwABHHIJwwAAAABJRU5ErkJggg==",t=256,r='<canvas width="'+t+'" height="'+t+'" src="'+o+'"></canvas>';return e.source.HistMap=function(i){var n=this,l=i||{};l.wrapX=!1,l.mapID&&(this.mapID=l.mapID,l.url="tiles/"+l.mapID+"/{z}/{x}/{y}.jpg"),l.urls?this._tileUrlFunction=e.TileUrlFunction.createFromTemplates(l.urls):l.url&&(this._tileUrlFunction=e.TileUrlFunction.createFromTemplates(e.TileUrlFunction.expandUrl(l.url))),this.width=l.width,this.height=l.height;var a=Math.log2(this.width/t),s=Math.log2(this.height/t);this.maxZoom=l.maxZoom=Math.ceil(Math.max(a,s)),this._maxxy=Math.pow(2,this.maxZoom)*t,l.tileUrlFunction=function(e){var r=e[0],i=e[1],n=-1*e[2]-1;return i*t*Math.pow(2,this.maxZoom-r)>=this.width||n*t*Math.pow(2,this.maxZoom-r)>=this.height||i<0||n<0?o:this._tileUrlFunction(e)},e.source.XYZ.call(this,l),e.source.setCustomInitialize(this,l),this.setTileLoadFunction(function(){var e=0,o=n.getTileLoadFunction();return function(i,n){++e;var l=i.getImage(),a=i.tImage;a||(a=$("<img>").get(0),a.crossOrigin="Anonymous",i.tImage=a),a.onload=a.onerror=function(){if(a.width&&a.height)if(a.width!=t||a.height!=t){var s=$(r).get(0),p=s.getContext("2d");p.drawImage(a,0,0);var g=s.toDataURL();l.crossOrigin=null,o(i,g),s=a=p=null}else l.crossOrigin="Anonymous",o(i,n);--e},a.src=n}}())},e.inherits(e.source.HistMap,e.source.XYZ),e.source.HistMap.createAsync=function(o){var t=o.algorythm||"tin";return e.source["HistMap_"+t].createAsync(o)},e.source.setCustomFunction(e.source.HistMap),e.source.HistMap.prototype.xy2MercAsync=function(e){return this.xy2MercAsync_(e)},e.source.HistMap.prototype.merc2XyAsync=function(e){return this.merc2XyAsync_(e)},e.source.HistMap.prototype.histMapCoords2Xy=function(o){var t=(o[0]+e.const.MERC_MAX)*this._maxxy/(2*e.const.MERC_MAX),r=(-o[1]+e.const.MERC_MAX)*this._maxxy/(2*e.const.MERC_MAX);return[t,r]},e.source.HistMap.prototype.xy2HistMapCoords=function(o){var t=o[0]*(2*e.const.MERC_MAX)/this._maxxy-e.const.MERC_MAX,r=-1*(o[1]*(2*e.const.MERC_MAX)/this._maxxy-e.const.MERC_MAX);return[t,r]},e.source.HistMap.prototype.insideCheckXy=function(e){return!(e[0]<0||e[0]>this.width||e[1]<0||e[1]>this.height)},e.source.HistMap.prototype.insideCheckHistMapCoords=function(e){return this.insideCheckXy(this.histMapCoords2Xy(e))},e}),define("histmap_tps",["histmap","tps"],function(e,o){return e.source.HistMap_tps=function(t){var r=t||{};if(e.source.HistMap.call(this,r),r.tps_points||r.tps_serial){var i={use_worker:!0,transform_callback:r.transform_callback,error_callback:r.error_callback,web_fallback:r.web_fallback,on_solved:r.on_solved,on_serialized:r.on_serialized};this.tps=new o(i),r.tps_points?this.tps.load_points(r.tps_points):this.tps.load_serial(r.tps_serial)}},e.inherits(e.source.HistMap_tps,e.source.HistMap),e.source.HistMap_tps.createAsync=function(o){return new Promise(function(t,r){var i;o.on_serialized=function(){if(o.tps_points){var e=document.createElement("a");document.body.appendChild(e),e.style="display: none";var r=new Blob([i.tps.serialize()]),n=window.URL.createObjectURL(r);e.href=n,e.download=o.tps_serial,e.click(),window.URL.revokeObjectURL(n)}t(i)},o.transform_callback=function(e,o,t){t.callback&&t.callback(e)},i=new e.source.HistMap_tps(o)})},e.source.HistMap_tps.prototype.xy2MercAsync_=function(e){var o=this;return new Promise(function(t,r){var i=o.histMapCoords2Xy(e);o.tps.transform(i,!1,{callback:function(e){t(e)}})})},e.source.HistMap_tps.prototype.merc2XyAsync_=function(e){var o=this;return new Promise(function(t,r){o.tps.transform(e,!0,{callback:function(e){var r=o.xy2HistMapCoords(e);t(r)}})})},e}),function(e){e(function(e){function o(o,t){return e.point(o,{target:t})}function t(e,o){var t=o.geometry.coordinates[0][0],r=o.geometry.coordinates[0][1],i=o.geometry.coordinates[0][2],n=e.geometry.coordinates,l=o.properties.a,a=o.properties.b,s=o.properties.c,p=[r[0]-t[0],r[1]-t[1]],g=[i[0]-t[0],i[1]-t[1]],c=[n[0]-t[0],n[1]-t[1]],u=[a[0]-l[0],a[1]-l[1]],y=[s[0]-l[0],s[1]-l[1]],h=(g[1]*c[0]-g[0]*c[1])/(p[0]*g[1]-p[1]*g[0]),d=(p[0]*c[1]-p[1]*c[0])/(p[0]*g[1]-p[1]*g[0]),m=[h*u[0]+d*y[0]+l[0],h*u[1]+d*y[1]+l[1]];return m}function r(o,r,i){var n=r,l=[];i||(i=7);for(var a=0;a<i;a++){var s=e.nearest(o,n);if(l.push(s),n=e.featureCollection(n.features.filter(function(e){return e!=s})),0==n.features.length)break}var p=e.featureCollection(l),g=e.tin(p,"target"),c=g.features.map(function(e){return t(o,e)}).reduce(function(e,o,t,r){return[e[0]+o[0],e[1]+o[1]].map(function(e){return t==r.length-1?e/r.length:e})},[0,0]);return c}function i(o,t){for(var r=0;r<t.features.length;r++){var i=e.inside(o,t.features[r]);if(i)return t.features[r]}}function n(e,o,n){var l=i(e,o);return l?t(e,l):r(e,n,11)}var l=function(e){this.points=e.points,this.wh=e.wh};return l.setTurf=function(o){e=o},l.prototype.setPoints=function(e){this.points=e,this.for_tins=void 0,this.bak_tins=void 0},l.prototype.setWh=function(e){this.wh=e,this.for_tins=void 0,this.bak_tins=void 0},l.prototype.updateTin=function(){var t=[],i=this.points;this.wh&&(t=[[0,0],[.5*this.wh[0],0],[this.wh[0],0],[0,.5*this.wh[1]],[this.wh[0],.5*this.wh[1]],[0,this.wh[1]],[.5*this.wh[0],this.wh[1]],[this.wh[0],this.wh[1]]]);for(var n=[],l=[],a=0;a<i.length;a++){var s=i[a][0],p=i[a][1];n.push(o(s,p)),l.push(o(p,s))}var g=e.featureCollection(n),c=e.featureCollection(l);t.map(function(t){var i=e.point(t),n=r(i,g,11);return[o(t,n),o(n,t)]}).map(function(e){g.features.push(e[0]),c.features.push(e[1])}),this.for_points=g,this.bak_points=c,this.for_tins=e.tin(g,"target"),this.bak_tins=e.tin(c,"target")},l.prototype.transform=function(o,t){this.bak_tins&&this.for_tins||this.updateTin();var r=e.point(o),i=t?this.bak_tins:this.for_tins,l=t?this.bak_points:this.for_points;return n(r,i,l)},l})}("function"==typeof turf?function(e){this.Tin=e(this.turf)}:"function"==typeof define&&define.amd?function(e){define("tin",["turf"],function(o){return e(o)})}:"undefined"!=typeof module&&module.exports&&"function"==typeof require?function(e){var o;try{o=require("turf")}catch(e){}module.exports=e(o)}:function(e){if(!this.turf)throw new Error('"turf" not found');this.Tin=e(this.turf)}),define("histmap_tin",["histmap","tin"],function(e,o){e.source.HistMap_tin=function(t){var r=t||{};e.source.HistMap.call(this,r),this.tin=new o({wh:[this.width,this.height]})},e.inherits(e.source.HistMap_tin,e.source.HistMap),e.source.HistMap_tin.createAsync=function(o){return new Promise(function(t,r){var i,n=o.tin_points_url,l=new XMLHttpRequest;l.open("GET",n,!0),l.responseType="json",l.onload=function(e){if(200==this.status){var o=this.response;i.tin.setPoints(o),i.tin.updateTin(),t(i)}},l.send(),i=new e.source.HistMap_tin(o)})},e.source.HistMap_tin.prototype.xy2MercAsync_=function(e){var o=this;return new Promise(function(t,r){var i=o.histMapCoords2Xy(e),n=o.tin.transform(i,!1);t(n)})},e.source.HistMap_tin.prototype.merc2XyAsync_=function(e){var o=this;return new Promise(function(t,r){var i=o.tin.transform(e,!0),n=o.xy2HistMapCoords(i);t(n)})}}),function(e,o,t){"function"==typeof define&&"undefined"!=typeof define.amd?define(o,function(){return t(e)}):"object"==typeof module&&"object"==typeof module.exports?module.exports=t(e):e[o]=t(e)}(this,"sprintf",function(){var e=function(){if(!arguments||arguments.length<1)throw new Error("sprintf: ERROR: not enough arguments");for(var e,o=0,t="",r=arguments[o++];e=/^([^%]*)%(?:(\d+)\$|\((.*?)\))?([#0 +'-]+)?(\*|\d+)?(\.\*|\.\d+)?([%diouxXfFeEcs])((?:.|[\r\n])*)$/.exec(r);){var i,n=e[1],l=e[2],a=e[3],s=e[4],p=e[5],g=e[6],c=e[7],u=e[8];if("%"===c)i="%";else{var y=" ",h=!0;s?(s.indexOf("0")>=0&&(y="0"),s.indexOf("-")>=0&&(y=" ",h=!1)):s="";var d,m=-1;if(p)if("*"===p){if(d=o++,d>=arguments.length)throw new Error("sprintf: ERROR: not enough arguments");m=arguments[d]}else m=parseInt(p,10);var f=-1;if(g)if(".*"===g){if(d=o++,d>=arguments.length)throw new Error("sprintf: ERROR: not enough arguments");f=arguments[d]}else f=parseInt(g.substring(1),10);if(d=o++,l){if(d=parseInt(l,10),d>=arguments.length)throw new Error("sprintf: ERROR: not enough arguments");i=arguments[d]}else if(a){if("object"!=typeof arguments[1])throw new Error("sprintf: ERROR: invalid non-object arguments for named argument");if(i=arguments[1][a],"undefined"==typeof i)throw new Error("sprintf: ERROR: invalid undefined value for named argument")}else{if(d>=arguments.length)throw new Error("sprintf: ERROR: not enough arguments");i=arguments[d]}var _="";switch(c){case"d":case"i":"number"!=typeof i&&(i=0),i=i.toString(10),s.indexOf("#")>=0&&i>=0&&(i="+"+i),s.indexOf(" ")>=0&&i>=0&&(i=" "+i);break;case"b":"number"!=typeof i&&(i=0),i=i.toString(2);break;case"o":"number"!=typeof i&&(i=0),i=i.toString(8);break;case"u":"number"!=typeof i&&(i=0),i=Math.abs(i),i=i.toString(10);break;case"x":"number"!=typeof i&&(i=0),i=i.toString(16).toLowerCase(),s.indexOf("#")>=0&&(_="0x");break;case"X":"number"!=typeof i&&(i=0),i=i.toString(16).toUpperCase(),s.indexOf("#")>=0&&(_="0X");break;case"f":case"F":if("number"!=typeof i&&(i=0),i=0+i,f>-1&&(i.toFixed?i=i.toFixed(f):(i=Math.round(i*Math.pow(10,f))/Math.pow(10,f),i+="0000000000",i=i.substr(0,i.indexOf(".")+f+1))),i=""+i,s.indexOf("'")>=0)for(var E=0,v=i.length-1-3;v>=0;v-=3)i=i.substring(0,v)+(0===E?".":",")+i.substring(v),E=(E+1)%2;break;case"e":case"E":if("number"!=typeof i&&(i=0),i=0+i,f>-1){if(!i.toExponential)throw new Error("sprintf: ERROR: toExponential() method not supported");i=i.toExponential(f)}i=""+i,"E"===c&&(i=i.replace(/e\+/,"E+"));break;case"c":"number"!=typeof i&&(i=0),i=String.fromCharCode(i);break;case"s":"string"!=typeof i&&(i=String(i)),f>-1&&(i=i.substr(0,f));break;default:throw new Error('sprintf: ERROR: invalid conversion character "'+c+'"')}var S=m-i.toString().length-_.toString().length;if(S>0){var P=new Array(S+1);h?i=P.join(y)+i:i+=P.join(y)}i=_+i}t=t+n+i,r=u}return t+r};return e.version={text:"1.2.3",major:1,minor:2,micro:3},e}),define("sprintf",function(){}),!function(e,o){"object"==typeof exports&&"undefined"!=typeof module?module.exports=o():"function"==typeof define&&define.amd?define("i18nxhr",o):e.i18nextXHRBackend=o()}(this,function(){"use strict";function e(e){return i.call(n.call(arguments,1),function(o){if(o)for(var t in o)void 0===e[t]&&(e[t]=o[t])}),e}function o(e,o,t,r,i){if(r&&"object"===("undefined"==typeof r?"undefined":l(r))){var n="",a=encodeURIComponent;for(var s in r)n+="&"+a(s)+"="+a(r[s]);r=n.slice(1)+(i?"":"&_t="+new Date)}try{var p=new(XMLHttpRequest||ActiveXObject)("MSXML2.XMLHTTP.3.0");p.open(r?"POST":"GET",e,1),o.crossDomain||p.setRequestHeader("X-Requested-With","XMLHttpRequest"),p.withCredentials=!!o.withCredentials,r&&p.setRequestHeader("Content-type","application/x-www-form-urlencoded"),p.onreadystatechange=function(){p.readyState>3&&t&&t(p.responseText,p)},p.send(r)}catch(e){window.console&&console.log(e)}}function t(){return{loadPath:"/locales/{{lng}}/{{ns}}.json",addPath:"locales/add/{{lng}}/{{ns}}",allowMultiLoading:!1,parse:JSON.parse,crossDomain:!1,ajax:o}}var r=[],i=r.forEach,n=r.slice,l="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e},a=(function(){function e(e){this.value=e}function o(o){function t(e,o){return new Promise(function(t,i){var a={key:e,arg:o,resolve:t,reject:i,next:null};l?l=l.next=a:(n=l=a,r(e,o))})}function r(t,n){try{var l=o[t](n),a=l.value;a instanceof e?Promise.resolve(a.value).then(function(e){r("next",e)},function(e){r("throw",e)}):i(l.done?"return":"normal",l.value)}catch(e){i("throw",e)}}function i(e,o){switch(e){case"return":n.resolve({value:o,done:!0});break;case"throw":n.reject(o);break;default:n.resolve({value:o,done:!1})}n=n.next,n?r(n.key,n.arg):l=null}var n,l;this._invoke=t,"function"!=typeof o.return&&(this.return=void 0)}return"function"==typeof Symbol&&Symbol.asyncIterator&&(o.prototype[Symbol.asyncIterator]=function(){return this}),o.prototype.next=function(e){return this._invoke("next",e)},o.prototype.throw=function(e){return this._invoke("throw",e)},o.prototype.return=function(e){return this._invoke("return",e)},{wrap:function(e){return function(){return new o(e.apply(this,arguments))}},await:function(o){return new e(o)}}}(),function(e,o){if(!(e instanceof o))throw new TypeError("Cannot call a class as a function")}),s=function(){function e(e,o){for(var t=0;t<o.length;t++){var r=o[t];r.enumerable=r.enumerable||!1,
-r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}return function(o,t,r){return t&&e(o.prototype,t),r&&e(o,r),o}}(),p=function(){function o(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{};a(this,o),this.init(e,t),this.type="backend"}return s(o,[{key:"init",value:function(o){var r=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{};this.services=o,this.options=e(r,this.options||{},t())}},{key:"readMulti",value:function(e,o,t){var r=this.options.loadPath;"function"==typeof this.options.loadPath&&(r=this.options.loadPath(e,o));var i=this.services.interpolator.interpolate(r,{lng:e.join("+"),ns:o.join("+")});this.loadUrl(i,t)}},{key:"read",value:function(e,o,t){var r=this.options.loadPath;"function"==typeof this.options.loadPath&&(r=this.options.loadPath([e],[o]));var i=this.services.interpolator.interpolate(r,{lng:e,ns:o});this.loadUrl(i,t)}},{key:"loadUrl",value:function(e,o){var t=this;this.options.ajax(e,this.options,function(r,i){if(i.status>=500&&i.status<600)return o("failed loading "+e,!0);if(i.status>=400&&i.status<500)return o("failed loading "+e,!1);var n=void 0,l=void 0;try{n=t.options.parse(r,e)}catch(o){l="failed parsing "+e+" to json"}return l?o(l,!1):void o(null,n)})}},{key:"create",value:function(e,o,t,r){var i=this;"string"==typeof e&&(e=[e]);var n={};n[t]=r||"",e.forEach(function(e){var t=i.services.interpolator.interpolate(i.options.addPath,{lng:e,ns:o});i.options.ajax(t,i.options,function(e,o){},n)})}}]),o}();return p.type="backend",p}),define("app",["jquery","ol-custom","sprintf","i18n","i18nxhr","ji18n","bootstrap","slick"],function(e,o,t,r,i,n){return e.fn.nodoubletapzoom=function(){e(this).bind("touchstart",function(o){var t=o.timeStamp,r=e(this).data("lastTouch")||t,i=t-r,n=o.originalEvent.touches.length;e(this).data("lastTouch",t),!i||i>500||n>1||(o.preventDefault(),e(o.target).trigger("click"))})},e("body").nodoubletapzoom(),function(l){var a=l.appid||"sample",s="json/"+a+".json",p=l.debug?function(e){console.log(e)}:function(){},g=l.lang||"ja",c=l.overlay||!1;c&&(e("body").addClass("with-opacity"),e(".opacity-slider").removeClass("hide"),e(".opacity-slider input").prop("disabled",!0)),r.use(i).init({lng:g,backend:{loadPath:"locales/{{lng}}/{{ns}}.json"}},function(i,a){n.init(r,e),e("body").localize(),e.get(s,function(r){e("#all").show(),e("#loadWait").modal(),e(".slick-class").slick({prevArrow:"",nextArrow:"",centerMode:!0,focusOnSelect:!0,slidesToScroll:3,centerPadding:"40px"});var i,n=null,s=r.home_position,g=r.default_zoom,u=r.app_name,y=!!l.fake&&r.fake_gps,h=!!l.fake&&r.fake_center,d=!!l.fake&&r.fake_radius,m=r.make_binary,f=null,_=null,E=null,v="map_div",S=null;c&&(e('<div id="map_div_back" class="map h100p w100p" style="top:0px; left:0px; width: 100%; position:absolute;"></div>').insertBefore("#center_circle"),e('<div id="map_div_front" class="map h100p w100p" style="top:0px; left:0px; width: 100%; position:absolute;"></div>').insertBefore("#center_circle"),v="map_div_front",S="map_div_back",E=new o.MaplatMap({off_control:!0,div:S})),y?e("#gps_etc").append(t(a("app.fake_explanation"),h,d)):e("#gps_etc").append(a("app.acquiring_gps_desc"));var P=r.pois;e("title").html(u);for(var x=r.sources,T={},R=[],M={map_option:{div:v},home_position:s,merc_zoom:g,fake_gps:!!y&&d},L=0;L<=x.length;L++)if(L==x.length)R.push(o.source.NowMap.createAsync(Object.assign({sourceID:"osm"},M))),e(".slick-class").slick("slickAdd",'<div class="slick-item" data="osm"><img src="./tmbs/osm_menu.jpg"><div>'+a("app.osm_now")+"</div></div>"),e(".slick-class").slick("slickGoTo",x.length);else{var C=x[L];C.maptype||(C.maptype="maplat"),C.algorythm||(C.algorythm=l.algorythm||"tin"),function(t){if("base"==t.maptype||"overlay"==t.maptype){t.sourceID=t.mapID;var r="base"==t.maptype?o.source.NowMap:o.source.TmsMap;R.push(r.createAsync(Object.assign({attributions:[new o.Attribution({html:t.attr})],url:t.url,sourceID:t.sourceID},M)))}else t.sourceID=t.mapID+":"+t.maptype+":"+t.algorythm,R.push(new Promise(function(e,r){var i=function(){T[t.sourceID]=t;var r=Object.assign({attributions:[new o.Attribution({html:t.attr})],mapID:t.mapID,width:t.width,height:t.height,maptype:t.maptype,algorythm:t.algorythm,sourceID:t.sourceID},M);"tin"==t.algorythm?r.tin_points_url="json/"+t.mapID+"_points.json":m?(r.tps_serial=t.mapID+".bin",r.tps_points="../json/"+t.mapID+"_points.json"):r.tps_serial="../bin/"+t.mapID+".bin",e(o.source.HistMap.createAsync(r))};"maplat"==t.maptype?require(["histmap_"+t.algorythm],i):i()}));e(".slick-class").slick("slickAdd",'<div class="slick-item" data="'+t.sourceID+'"><img src="./tmbs/'+t.mapID+'_menu.jpg"><div>'+(t.label||t.year)+"</div></div>")}(C)}Promise.all(R).then(function(t){function r(e,t){var r=_.getView(),l=i.size2MercsAsync();if(n&&n.mercs&&n.buffer[i.sourceID]){var a=n.buffer[i.sourceID],s=o.MathEx.recursiveRound([r.getCenter(),r.getZoom(),r.getRotation()],10);a[0][0]==s[0][0]&&a[0][1]==s[0][1]&&a[1]==s[1]&&a[2]==s[2]?(p("From: Use buffer"),l=new Promise(function(e,o){e(n.mercs)})):n={buffer:{}}}else n={buffer:{}};l.then(function(r){n.mercs=r;var l=_.getView();p("From: Center: "+l.getCenter()+" Zoom: "+l.getZoom()+" Rotation: "+l.getRotation()),n.buffer[i.sourceID]=o.MathEx.recursiveRound([l.getCenter(),l.getZoom(),l.getRotation()],10),p("Mercs: "+r);var a=e.mercs2SizeAsync(r),s=e.sourceID;n.buffer[s]&&(p("To: Use buffer"),a=new Promise(function(e,o){e(n.buffer[s])})),a.then(function(r){p("To: Center: "+[r[0][0],r[0][1]]+" Zoom: "+r[1]+" Rotation: "+r[2]),n.buffer[e.sourceID]=o.MathEx.recursiveRound(r,10),t(r)})})}function l(t,n){var l=c.osm,s=c[n];s==i&&s!=l||s!=i&&r(s,function(n){var p=null,g=null;E&&(p=E.getSource(),s instanceof o.source.NowMap?s instanceof o.source.NowMap&&E.exchangeSource():p?g=p:(g=l,i instanceof o.source.NowMap&&(g=i instanceof o.source.TmsMap?_.getSource():i),E.exchangeSource(g)),!(s instanceof o.source.NowMap)||s instanceof o.source.TmsMap?e(".opacity-slider input").removeProp("disabled"):(e(".opacity-slider input").val(0),e(".opacity-slider input").prop("disabled",!0))),s instanceof o.source.TmsMap?(_.setLayer(s),i instanceof o.source.NowMap||_.exchangeSource(p||l),e(".opacity-slider input").removeProp("disabled")):(_.setLayer(),_.exchangeSource(s)),_.setOpacity(e(".opacity-slider input").val());var c=_.getView();s.insideCheckHistMapCoords(n[0])?(c.setCenter(n[0]),c.setZoom(n[1]),c.setRotation(n[2])):(e("#gpsDialogTitle").text(a("app.out_of_map")),e("#gpsDialogBody").text(a("app.out_of_map_area")),e("#gpsDialog").modal(),s.goHome()),s.setGPSMarker(f,!0),_.resetMarker();for(var u=0;u<P.length;u++)!function(e){var t=[e.lng,e.lat],r=o.proj.transform(t,"EPSG:4326","EPSG:3857");s.merc2XyAsync(r).then(function(o){s.insideCheckHistMapCoords(o)&&_.setMarker(o,{datum:e})})}(P[u]);_.updateSize(),_.renderSync(),i=s,1==t&&s.goHome(),E&&g&&r(g,function(e){var o=E.getView();o.setCenter(e[0]),o.setZoom(e[1]),o.setRotation(e[2]),E.updateSize(),E.renderSync()})})}function s(o){e("#poi_name").text(o.name),e("#poi_img").attr("src","img/"+o.image),e("#poi_address").text(o.address),e("#poi_desc").html(o.desc.replace(/\n/g,"<br>")),e("#poi_info").modal()}e("#loadWait").modal("hide");for(var g=[],c={},u=!1,y=0;y<t.length;y++){var h=t[y];_||h instanceof o.source.TmsMap||(_=h.getMap(),_.on("gps_request",function(){e("#gpsWait").modal()}),_.on("gps_result",function(o){var t=(e("#gpsWait").data("bs.modal")||{isShown:!1}).isShown,r=o.frameState;r&&r.error?(f=null,"gps_out"==r.error&&t&&(e("#gpsWait").modal("hide"),e("#gpsDialogTitle").text(a("app.out_of_map")),e("#gpsDialogBody").text(a("app.out_of_map_desc")),e("#gpsDialog").modal())):f=r,t&&e("#gpsWait").modal("hide")})),h._map=_,g.push(h),c[h.sourceID]=h}e(".slick-item").on("click",function(){u||l(!1,e(this).attr("data"))}),e(".slick-class").on("beforeChange",function(e,o,t,r){u=t!=r}),e(".slick-class").on("afterChange",function(e,o,t){u=!1}),i=g.reduce(function(e,t){return e?e:t instanceof o.source.HistMap?t:e},null),l(!0,"osm");var d=function(e){return function(o){var t=e.forEachFeatureAtPixel(o.pixel,function(e){if(e.get("datum"))return e});t&&s(t.get("datum"))}}(_);_.on("click",d);var m=function(o){return function(t){var r=o.getEventPixel(t.originalEvent),i=o.hasFeatureAtPixel(r),n=o.getTarget();if(i){var l=o.forEachFeatureAtPixel(t.pixel,function(e){if(e.get("datum"))return e});return void e("#"+n).css("cursor",l?"pointer":"")}e("#"+n).css("cursor","")}}(_);_.on("pointermove",m);var v=function(e){return function(o){var t=o.frameState.viewState.center,r=e.getSource();if(!r.insideCheckHistMapCoords(t)){var i=r.histMapCoords2Xy(t);p(i);var n=i[0]/(r.width/2)-1,l=i[1]/(r.height/2)-1,a=Math.max(Math.abs(n),Math.abs(l));i=[(n/a+1)*r.width/2,(l/a+1)*r.height/2],p(i),t=r.xy2HistMapCoords(i),e.getView().setCenter(t)}}}(_);_.on("moveend",v);var S=function(e){return function(o){if(E){if(e._backMapMoving)return void p("Backmap moving skipped");var t=E.getSource();t&&(e._backMapMoving=!0,p("Backmap moving started"),r(t,function(o){var t=E.getView();t.setCenter(o[0]),t.setZoom(o[1]),t.setRotation(o[2]),p("Backmap moving ended"),e._backMapMoving=!1}))}}}(_);_.on("postrender",S);var x=function(){_.setOpacity(e(".opacity-slider input").val())};e(".opacity-slider input").on("input",x),e(".opacity-slider input").on("change",x)})},"json")})}});
+goog.exportProperty(ol.control.Rotate.prototype,"getKeys",ol.control.Rotate.prototype.getKeys),goog.exportProperty(ol.control.Rotate.prototype,"getProperties",ol.control.Rotate.prototype.getProperties),goog.exportProperty(ol.control.Rotate.prototype,"set",ol.control.Rotate.prototype.set),goog.exportProperty(ol.control.Rotate.prototype,"setProperties",ol.control.Rotate.prototype.setProperties),goog.exportProperty(ol.control.Rotate.prototype,"unset",ol.control.Rotate.prototype.unset),goog.exportProperty(ol.control.Rotate.prototype,"changed",ol.control.Rotate.prototype.changed),goog.exportProperty(ol.control.Rotate.prototype,"dispatchEvent",ol.control.Rotate.prototype.dispatchEvent),goog.exportProperty(ol.control.Rotate.prototype,"getRevision",ol.control.Rotate.prototype.getRevision),goog.exportProperty(ol.control.Rotate.prototype,"on",ol.control.Rotate.prototype.on),goog.exportProperty(ol.control.Rotate.prototype,"once",ol.control.Rotate.prototype.once),goog.exportProperty(ol.control.Rotate.prototype,"un",ol.control.Rotate.prototype.un),goog.exportProperty(ol.control.Rotate.prototype,"unByKey",ol.control.Rotate.prototype.unByKey),goog.exportProperty(ol.control.ScaleLine.prototype,"getMap",ol.control.ScaleLine.prototype.getMap),goog.exportProperty(ol.control.ScaleLine.prototype,"setMap",ol.control.ScaleLine.prototype.setMap),goog.exportProperty(ol.control.ScaleLine.prototype,"setTarget",ol.control.ScaleLine.prototype.setTarget),goog.exportProperty(ol.control.ScaleLine.prototype,"get",ol.control.ScaleLine.prototype.get),goog.exportProperty(ol.control.ScaleLine.prototype,"getKeys",ol.control.ScaleLine.prototype.getKeys),goog.exportProperty(ol.control.ScaleLine.prototype,"getProperties",ol.control.ScaleLine.prototype.getProperties),goog.exportProperty(ol.control.ScaleLine.prototype,"set",ol.control.ScaleLine.prototype.set),goog.exportProperty(ol.control.ScaleLine.prototype,"setProperties",ol.control.ScaleLine.prototype.setProperties),goog.exportProperty(ol.control.ScaleLine.prototype,"unset",ol.control.ScaleLine.prototype.unset),goog.exportProperty(ol.control.ScaleLine.prototype,"changed",ol.control.ScaleLine.prototype.changed),goog.exportProperty(ol.control.ScaleLine.prototype,"dispatchEvent",ol.control.ScaleLine.prototype.dispatchEvent),goog.exportProperty(ol.control.ScaleLine.prototype,"getRevision",ol.control.ScaleLine.prototype.getRevision),goog.exportProperty(ol.control.ScaleLine.prototype,"on",ol.control.ScaleLine.prototype.on),goog.exportProperty(ol.control.ScaleLine.prototype,"once",ol.control.ScaleLine.prototype.once),goog.exportProperty(ol.control.ScaleLine.prototype,"un",ol.control.ScaleLine.prototype.un),goog.exportProperty(ol.control.ScaleLine.prototype,"unByKey",ol.control.ScaleLine.prototype.unByKey),goog.exportProperty(ol.control.Zoom.prototype,"getMap",ol.control.Zoom.prototype.getMap),goog.exportProperty(ol.control.Zoom.prototype,"setMap",ol.control.Zoom.prototype.setMap),goog.exportProperty(ol.control.Zoom.prototype,"setTarget",ol.control.Zoom.prototype.setTarget),goog.exportProperty(ol.control.Zoom.prototype,"get",ol.control.Zoom.prototype.get),goog.exportProperty(ol.control.Zoom.prototype,"getKeys",ol.control.Zoom.prototype.getKeys),goog.exportProperty(ol.control.Zoom.prototype,"getProperties",ol.control.Zoom.prototype.getProperties),goog.exportProperty(ol.control.Zoom.prototype,"set",ol.control.Zoom.prototype.set),goog.exportProperty(ol.control.Zoom.prototype,"setProperties",ol.control.Zoom.prototype.setProperties),goog.exportProperty(ol.control.Zoom.prototype,"unset",ol.control.Zoom.prototype.unset),goog.exportProperty(ol.control.Zoom.prototype,"changed",ol.control.Zoom.prototype.changed),goog.exportProperty(ol.control.Zoom.prototype,"dispatchEvent",ol.control.Zoom.prototype.dispatchEvent),goog.exportProperty(ol.control.Zoom.prototype,"getRevision",ol.control.Zoom.prototype.getRevision),goog.exportProperty(ol.control.Zoom.prototype,"on",ol.control.Zoom.prototype.on),goog.exportProperty(ol.control.Zoom.prototype,"once",ol.control.Zoom.prototype.once),goog.exportProperty(ol.control.Zoom.prototype,"un",ol.control.Zoom.prototype.un),goog.exportProperty(ol.control.Zoom.prototype,"unByKey",ol.control.Zoom.prototype.unByKey),goog.exportProperty(ol.control.ZoomSlider.prototype,"getMap",ol.control.ZoomSlider.prototype.getMap),goog.exportProperty(ol.control.ZoomSlider.prototype,"setMap",ol.control.ZoomSlider.prototype.setMap),goog.exportProperty(ol.control.ZoomSlider.prototype,"setTarget",ol.control.ZoomSlider.prototype.setTarget),goog.exportProperty(ol.control.ZoomSlider.prototype,"get",ol.control.ZoomSlider.prototype.get),goog.exportProperty(ol.control.ZoomSlider.prototype,"getKeys",ol.control.ZoomSlider.prototype.getKeys),goog.exportProperty(ol.control.ZoomSlider.prototype,"getProperties",ol.control.ZoomSlider.prototype.getProperties),goog.exportProperty(ol.control.ZoomSlider.prototype,"set",ol.control.ZoomSlider.prototype.set),goog.exportProperty(ol.control.ZoomSlider.prototype,"setProperties",ol.control.ZoomSlider.prototype.setProperties),goog.exportProperty(ol.control.ZoomSlider.prototype,"unset",ol.control.ZoomSlider.prototype.unset),goog.exportProperty(ol.control.ZoomSlider.prototype,"changed",ol.control.ZoomSlider.prototype.changed),goog.exportProperty(ol.control.ZoomSlider.prototype,"dispatchEvent",ol.control.ZoomSlider.prototype.dispatchEvent),goog.exportProperty(ol.control.ZoomSlider.prototype,"getRevision",ol.control.ZoomSlider.prototype.getRevision),goog.exportProperty(ol.control.ZoomSlider.prototype,"on",ol.control.ZoomSlider.prototype.on),goog.exportProperty(ol.control.ZoomSlider.prototype,"once",ol.control.ZoomSlider.prototype.once),goog.exportProperty(ol.control.ZoomSlider.prototype,"un",ol.control.ZoomSlider.prototype.un),goog.exportProperty(ol.control.ZoomSlider.prototype,"unByKey",ol.control.ZoomSlider.prototype.unByKey),goog.exportProperty(ol.control.ZoomToExtent.prototype,"getMap",ol.control.ZoomToExtent.prototype.getMap),goog.exportProperty(ol.control.ZoomToExtent.prototype,"setMap",ol.control.ZoomToExtent.prototype.setMap),goog.exportProperty(ol.control.ZoomToExtent.prototype,"setTarget",ol.control.ZoomToExtent.prototype.setTarget),goog.exportProperty(ol.control.ZoomToExtent.prototype,"get",ol.control.ZoomToExtent.prototype.get),goog.exportProperty(ol.control.ZoomToExtent.prototype,"getKeys",ol.control.ZoomToExtent.prototype.getKeys),goog.exportProperty(ol.control.ZoomToExtent.prototype,"getProperties",ol.control.ZoomToExtent.prototype.getProperties),goog.exportProperty(ol.control.ZoomToExtent.prototype,"set",ol.control.ZoomToExtent.prototype.set),goog.exportProperty(ol.control.ZoomToExtent.prototype,"setProperties",ol.control.ZoomToExtent.prototype.setProperties),goog.exportProperty(ol.control.ZoomToExtent.prototype,"unset",ol.control.ZoomToExtent.prototype.unset),goog.exportProperty(ol.control.ZoomToExtent.prototype,"changed",ol.control.ZoomToExtent.prototype.changed),goog.exportProperty(ol.control.ZoomToExtent.prototype,"dispatchEvent",ol.control.ZoomToExtent.prototype.dispatchEvent),goog.exportProperty(ol.control.ZoomToExtent.prototype,"getRevision",ol.control.ZoomToExtent.prototype.getRevision),goog.exportProperty(ol.control.ZoomToExtent.prototype,"on",ol.control.ZoomToExtent.prototype.on),goog.exportProperty(ol.control.ZoomToExtent.prototype,"once",ol.control.ZoomToExtent.prototype.once),goog.exportProperty(ol.control.ZoomToExtent.prototype,"un",ol.control.ZoomToExtent.prototype.un),goog.exportProperty(ol.control.ZoomToExtent.prototype,"unByKey",ol.control.ZoomToExtent.prototype.unByKey),ol.VERSION="v3.19.1",OPENLAYERS.ol=ol,OPENLAYERS.ol}),define("ol-custom",["ol3"],function(e){e.interaction.PinchZoom.handleUpEvent_=function(o){if(this.targetPointers.length<2){var t=o.map,r=t.getView();return r.setHint(e.View.Hint.INTERACTING,-1),!1}return!0},e.View.prototype.getDecimalZoom=function(){var e,o=this.getResolution();return e=Math.log(this.maxResolution_/o)/Math.log(2),void 0!==e?this.minZoom_+e:e},e.control.CustomControl=function(o){var t=o||{},r=document.createElement("button");r.setAttribute("type","button");var i=document.createElement("span");i.innerHTML=t.character,r.appendChild(i),e.events.listen(r,e.events.EventType.CLICK,t.callback,this);var n=document.createElement("div");n.className=t.cls+" ol-unselectable ol-control",n.appendChild(r),e.control.Control.call(this,{element:n,target:t.target,render:t.render})},e.inherits(e.control.CustomControl,e.control.Control),e.control.GoHome=function(o){var t=o||{};t.character='<i class="fa fa-home fa-lg"></i>',t.cls="home";var r=this;t.callback=function(){var e=r.getMap().getLayers().item(0).getSource();e.goHome()},e.control.CustomControl.call(this,t)},e.inherits(e.control.GoHome,e.control.CustomControl),e.control.SetGPS=function(o){var t=o||{};t.character='<i class="fa fa-crosshairs fa-lg"></i>',t.cls="gps";var r=this;t.render=function(e){var o=e.frameState;if(o){var t=this.getMap();if(t.geolocation){var i=t.geolocation.getTracking(),n=r.element.classList.contains("receiving_gps");n&&!i?this.element.classList.remove("receiving_gps"):!n&&i&&this.element.classList.add("receiving_gps")}}},t.callback=function(){var e=r.element.classList.contains("receiving_gps"),o=r.getMap();o.handleGPS(!e),e?r.element.classList.remove("receiving_gps"):r.element.classList.add("receiving_gps")},e.control.CustomControl.call(this,t)},e.inherits(e.control.SetGPS,e.control.CustomControl),e.control.CompassRotate=function(o){var t=o||{};t.autoHide=!1;var r=document.createElement("span");r.innerHTML='<i class="fa fa-compass fa-lg ol-compass-fa"></i>',t.label=r,t.render=function(e){var o=e.frameState;if(o){var t=this.getMap().getView(),r=o.viewState.rotation,i=t.getCenter(),n=t.getDecimalZoom();if(r!=this.rotation_||i[0]!=this.center_[0]||i[1]!=this.center_[1]||n!=this.zoom_){var l=this.element.classList.contains("no_rotation");l||0!==r?l&&0!==r&&this.element.classList.remove("no_rotation"):this.element.classList.add("no_rotation");var a=this,s=this.getMap().getLayers().item(0).getSource();s.size2MercsAsync().then(function(e){var o=s.mercs2MercRotation(e),t="rotate("+o+"rad)";a.label_.style.msTransform=t,a.label_.style.webkitTransform=t,a.label_.style.transform=t})}this.rotation_=r,this.center_=i,this.zoom_=n}},e.control.Rotate.call(this,t),this.center_=[],this.zoom_=void 0},e.inherits(e.control.CompassRotate,e.control.Rotate),e.const=e.const?e.const:{},e.const.MERC_MAX=20037508.342789244,e.const.MERC_CROSSMATRIX=[[0,0],[0,1],[1,0],[0,-1],[-1,0]];var o=new e.style.Style({image:new e.style.Icon({anchor:[.5,.5],anchorXUnits:"fraction",anchorYUnits:"fraction",src:"parts/bluedot.png"})}),t=new e.style.Style({fill:new e.style.Fill({color:[128,128,256,.2]}),stroke:new e.style.Stroke({color:[128,128,256,1],width:3})}),r=new e.style.Style({image:new e.style.Icon({anchor:[.5,1],anchorXUnits:"fraction",anchorYUnits:"fraction",src:"parts/defaultpin.png"})});return e.source.setCustomFunction=function(o){o.prototype.getMap=function(){return this._map?this._map:(this._map=new e.MaplatMap({source:this,div:this.map_option.div,center:this.map_option.default_center,zoom:this.map_option.default_zoom,rotation:this.map_option.default_rotation}),this._map)},o.prototype.goHome=function(){var o=e.proj.transform(this.home_position,"EPSG:4326","EPSG:3857"),t=this._map,r=t.getView(),i=this.mercsFromGivenZoom(o,this.merc_zoom);this.mercs2SizeAsync(i).then(function(e){r.setCenter(e[0]),r.setZoom(e[1]),r.setRotation(0)})},o.prototype.setGPSMarkerAsync=function(e,o){var t=this,r=t.getMap(),i=r.getView(),n=e?t.mercsFromGPSValue(e.lnglat,e.acc):["dummy"];return Promise.all(n.map(function(e,o){return 5==o||"dummy"==e?e:t.merc2XyAsync(e)})).then(function(e){var n=null;if("dummy"!=e[0]){if(n={xy:e[0]},!t.insideCheckHistMapCoords(e[0]))return r.handleGPS(!1,!0),!1;var l=e.slice(1);n.rad=l.reduce(function(e,o,t){var r=e+Math.sqrt(Math.pow(o[0]-n.xy[0],2)+Math.pow(o[1]-n.xy[1],2));return 3==t?r/4:r},0),o||i.setCenter(n.xy)}return r.setGPSPosition(n),!0})},o.prototype.setGPSMarker=function(e,o){this.setGPSMarkerAsync(e,o).then(function(){})},o.prototype.getRadius=function(o){var t=Math.floor(Math.min(o[0],o[1])/4),r=this._map.getView().getDecimalZoom();return t*e.const.MERC_MAX/128/Math.pow(2,r)},o.prototype.mercsFromGivenZoom=function(o,t){var r=this._map.getSize(),i=Math.floor(Math.min(r[0],r[1])/4),n=i*e.const.MERC_MAX/128/Math.pow(2,t);return e.const.MERC_CROSSMATRIX.map(function(e){return[e[0]*n+o[0],e[1]*n+o[1]]})},o.prototype.mercsFromGPSValue=function(o,t){var r=e.proj.transform(o,"EPSG:4326","EPSG:3857"),i=o[1]*Math.PI/180,n=t/Math.cos(i);return e.const.MERC_CROSSMATRIX.map(function(e){return[e[0]*n+r[0],e[1]*n+r[1]]})},o.prototype.rotateMatrix=function(e){for(var o=1*this._map.getView().getRotation(),t=[],r=0;r<e.length;r++){var i=e[r],n=i[0]*Math.cos(o)-i[1]*Math.sin(o),l=i[0]*Math.sin(o)+i[1]*Math.cos(o);t.push([n,l])}return t},o.prototype.size2Xys=function(o){o||(o=this._map.getView().getCenter());var t=this._map.getSize(),r=this.getRadius(t),i=this.rotateMatrix(e.const.MERC_CROSSMATRIX),n=i.map(function(e){return[e[0]*r+o[0],e[1]*r+o[1]]});return n.push(t),n},o.prototype.size2MercsAsync=function(e){var o=this.size2Xys(e),t=this,r=o.map(function(e,o){return 5==o?e:t.xy2MercAsync(e)});return Promise.all(r)},o.prototype.mercs2SizeAsync=function(e){var o=this,t=e.map(function(e,t){return 5==t?e:o.merc2XyAsync(e)});return Promise.all(t).then(function(e){return o.xys2Size(e)})},o.prototype.xys2Size=function(o){for(var t=o[0],r=o[5],i=o.slice(1,5),n=i.map(function(e){return[e[0]-t[0],e[1]-t[1]]}),l=[[0,1],[1,0],[0,-1],[-1,0]],a=0,s=0,p=0,g=0;g<4;g++){var c=n[g],u=l[g],y=Math.sqrt(Math.pow(c[0],2)+Math.pow(c[1],2));a+=y;var h=c[0]*u[1]-c[1]*u[0],d=Math.acos((c[0]*u[0]+c[1]*u[1])/y),m=h>0?-1*d:d;s+=Math.cos(m),p+=Math.sin(m)}var f=a/4,_=Math.atan2(p,s);r||(r=this._map.getSize());var E=Math.floor(Math.min(r[0],r[1])/4),v=Math.log(E*e.const.MERC_MAX/128/f)/Math.log(2);return[t,v,_]},o.prototype.mercs2MercRotation=function(e){for(var o=e[0],t=e.slice(1,5),r=t.map(function(e){return[e[0]-o[0],e[1]-o[1]]}),i=[[0,1],[1,0],[0,-1],[-1,0]],n=0,l=0,a=0;a<4;a++){var s=r[a],p=i[a],g=Math.sqrt(Math.pow(s[0],2)+Math.pow(s[1],2)),c=s[0]*p[1]-s[1]*p[0],u=Math.acos((s[0]*p[0]+s[1]*p[1])/g),y=c>0?-1*u:u;n+=Math.cos(y),l+=Math.sin(y)}return Math.atan2(l,n)}},e.source.setCustomInitialize=function(e,o){e.sourceID=o.sourceID,e.map_option=o.map_option||{},e.home_position=o.home_position,e.merc_zoom=o.merc_zoom,e.fake_gps=o.fake_gps||!1},e.source.NowMap=function(o){var t=o||{};e.source.OSM.call(this,t),e.source.setCustomInitialize(this,t)},e.inherits(e.source.NowMap,e.source.OSM),e.source.NowMap.createAsync=function(o){return new Promise(function(t,r){var i=new e.source.NowMap(o);t(i)})},e.source.setCustomFunction(e.source.NowMap),e.source.NowMap.prototype.xy2MercAsync=function(e){return new Promise(function(o,t){o(e)})},e.source.NowMap.prototype.merc2XyAsync=function(e){return new Promise(function(o,t){o(e)})},e.source.NowMap.prototype.insideCheckXy=function(e){return!0},e.source.NowMap.prototype.insideCheckHistMapCoords=function(e){return!0},e.source.TmsMap=function(o){var t=o||{};e.source.NowMap.call(this,t)},e.inherits(e.source.TmsMap,e.source.NowMap),e.source.TmsMap.createAsync=function(o){var t=new Promise(function(t,r){var i=new e.source.TmsMap(o);t(i)});return t},e.MaplatMap=function(o){o=o||{},this._gps_source=new e.source.Vector({wrapX:!1});var t=new e.layer.Vector({source:this._gps_source});this._marker_source=new e.source.Vector({wrapX:!1});var r=new e.layer.Vector({source:this._marker_source}),i=this._overlay_group=new e.layer.Group,n=o.off_control?[]:[new e.control.Attribution,new e.control.CompassRotate,new e.control.Zoom,new e.control.SetGPS,new e.control.GoHome],l={controls:n,layers:[new e.layer.Tile({source:o.source}),i,t,r],target:o.div,view:new e.View({center:o.default_center||[0,0],zoom:o.default_zoom||2,rotation:o.default_rotation||0})};e.Map.call(this,l);var a=this.getView(),s=this;s.__AvoidFirstMoveStart=!0;var p=function(){s.__AvoidFirstMoveStart||s.dispatchEvent("movestart"),s.__AvoidFirstMoveStart=!1,a.un("propertychange",p)};a.on("propertychange",p),s.on("moveend",function(){a.on("propertychange",p)})},e.inherits(e.MaplatMap,e.Map),e.MaplatMap.prototype.setGPSPosition=function(r){var i=this._gps_source;if(i.clear(),r){var n=new e.Feature({geometry:new e.geom.Point(r.xy)});n.setStyle(o);var l=new e.Feature({geometry:new e.geom.Circle(r.xy,r.rad)});l.setStyle(t),i.addFeature(n),i.addFeature(l)}},e.MaplatMap.prototype.resetMarker=function(){var e=this._marker_source;e.clear()},e.MaplatMap.prototype.setMarker=function(o,t,i){var n=this._marker_source;t.geometry=new e.geom.Point(o);var l=new e.Feature(t);i||(i=r),l.setStyle(i),n.addFeature(l)},e.MaplatMap.prototype.exchangeSource=function(e){var o=this.getLayers(),t=o.item(0);t.setSource(e),e&&(e._map=this)},e.MaplatMap.prototype.setLayer=function(o){var t=this._overlay_group.getLayers();if(t.clear(),o){var r=new e.layer.Tile({source:o});t.push(r)}},e.MaplatMap.prototype.getSource=function(){return this.getLayers().item(0).getSource()},e.MaplatMap.prototype.setOpacity=function(o){var t=(100-o)/100,r=this.getSource();r instanceof e.source.NowMap?(this.getLayers().item(0).setOpacity(1),this.getLayers().item(1).setOpacity(t)):this.getLayers().item(0).setOpacity(t)},e.MaplatMap.prototype.handleGPS=function(o,t){if(o)if(this.dispatchEvent("gps_request"),this._first_gps_request=!0,this.geolocation)this.geolocation.setTracking(!0);else{var r=this.geolocation=new e.Geolocation({tracking:!0}),i=this;r.on("change",function(o){var t=i.getLayers().item(0).getSource(),n=r.getPosition(),l=r.getAccuracy();t.fake_gps&&e.MathEx.getDistance(t.home_position,n)>t.fake_gps&&(n=[e.MathEx.randomFromCenter(t.home_position[0],.001),e.MathEx.randomFromCenter(t.home_position[1],.001)],l=e.MathEx.randomFromCenter(15,10));var a={lnglat:n,acc:l};t.setGPSMarkerAsync(a,!i._first_gps_request).then(function(o){o||(a={error:"gps_out"}),i._first_gps_request=!1,i.dispatchEvent(new e.MapEvent("gps_result",i,a))})}),r.on("error",function(o){var t=i.getLayers().item(0).getSource(),r=null;if(t.fake_gps){var n=[e.MathEx.randomFromCenter(t.home_position[0],.001),e.MathEx.randomFromCenter(t.home_position[1],.001)],l=e.MathEx.randomFromCenter(15,10);r={lnglat:n,acc:l}}t.setGPSMarkerAsync(r,!i._first_gps_request).then(function(o){o||(r={error:"gps_out"}),i._first_gps_request=!1,i.dispatchEvent(new e.MapEvent("gps_result",i,r))})})}else{this.geolocation&&this.geolocation.setTracking(!1);var n=this.getLayers().item(0).getSource();n.setGPSMarker(),t||this.dispatchEvent(new e.MapEvent("gps_result",i,{error:"gps_off"}))}},e.MathEx={},e.MathEx.randomFromCenter=function(e,o){return e+(Math.random()-.5)*o},e.MathEx.recursiveRound=function(o,t){if(o instanceof Array)return o.map(function(o){return e.MathEx.recursiveRound(o,t)});var r=Math.pow(10,t);return Math.round(o*r)/r},e.MathEx.getDistance=function(e,o){function t(e){return e*Math.PI/180}return 6378.14*Math.acos(Math.cos(t(e[1]))*Math.cos(t(o[1]))*Math.cos(t(o[0])-t(e[0]))+Math.sin(t(e[1]))*Math.sin(t(o[1])))},e}),define("histmap",["ol-custom"],function(e){var o="data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAQAAAAEACAMAAABrrFhUAAAAB3RJTUUH3QgIBToaSbAjlwAAABd0RVh0U29mdHdhcmUAR0xEUE5HIHZlciAzLjRxhaThAAAACHRwTkdHTEQzAAAAAEqAKR8AAAAEZ0FNQQAAsY8L/GEFAAAAA1BMVEX///+nxBvIAAAAAXRSTlMAQObYZgAAAFRJREFUeNrtwQEBAAAAgJD+r+4ICgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAABgBDwABHHIJwwAAAABJRU5ErkJggg==",t=256,r='<canvas width="'+t+'" height="'+t+'" src="'+o+'"></canvas>';return e.source.HistMap=function(i){var n=this,l=i||{};l.wrapX=!1,l.mapID&&(this.mapID=l.mapID,l.url="tiles/"+l.mapID+"/{z}/{x}/{y}.jpg"),l.urls?this._tileUrlFunction=e.TileUrlFunction.createFromTemplates(l.urls):l.url&&(this._tileUrlFunction=e.TileUrlFunction.createFromTemplates(e.TileUrlFunction.expandUrl(l.url))),this.width=l.width,this.height=l.height;var a=Math.log2(this.width/t),s=Math.log2(this.height/t);this.maxZoom=l.maxZoom=Math.ceil(Math.max(a,s)),this._maxxy=Math.pow(2,this.maxZoom)*t,l.tileUrlFunction=function(e){var r=e[0],i=e[1],n=-1*e[2]-1;return i*t*Math.pow(2,this.maxZoom-r)>=this.width||n*t*Math.pow(2,this.maxZoom-r)>=this.height||i<0||n<0?o:this._tileUrlFunction(e)},e.source.XYZ.call(this,l),e.source.setCustomInitialize(this,l),this.setTileLoadFunction(function(){var e=0,o=n.getTileLoadFunction();return function(i,n){++e;var l=i.getImage(),a=i.tImage;a||(a=$("<img>").get(0),a.crossOrigin="Anonymous",i.tImage=a),a.onload=a.onerror=function(){if(a.width&&a.height)if(a.width!=t||a.height!=t){var s=$(r).get(0),p=s.getContext("2d");p.drawImage(a,0,0);var g=s.toDataURL();l.crossOrigin=null,o(i,g),s=a=p=null}else l.crossOrigin="Anonymous",o(i,n);--e},a.src=n}}())},e.inherits(e.source.HistMap,e.source.XYZ),e.source.HistMap.createAsync=function(o){var t=o.algorythm||"tin";return e.source["HistMap_"+t].createAsync(o)},e.source.setCustomFunction(e.source.HistMap),e.source.HistMap.prototype.xy2MercAsync=function(e){return this.xy2MercAsync_(e)},e.source.HistMap.prototype.merc2XyAsync=function(e){return this.merc2XyAsync_(e)},e.source.HistMap.prototype.histMapCoords2Xy=function(o){var t=(o[0]+e.const.MERC_MAX)*this._maxxy/(2*e.const.MERC_MAX),r=(-o[1]+e.const.MERC_MAX)*this._maxxy/(2*e.const.MERC_MAX);return[t,r]},e.source.HistMap.prototype.xy2HistMapCoords=function(o){var t=o[0]*(2*e.const.MERC_MAX)/this._maxxy-e.const.MERC_MAX,r=-1*(o[1]*(2*e.const.MERC_MAX)/this._maxxy-e.const.MERC_MAX);return[t,r]},e.source.HistMap.prototype.insideCheckXy=function(e){return!(e[0]<0||e[0]>this.width||e[1]<0||e[1]>this.height)},e.source.HistMap.prototype.insideCheckHistMapCoords=function(e){return this.insideCheckXy(this.histMapCoords2Xy(e))},e}),define("histmap_tps",["histmap","tps"],function(e,o){return e.source.HistMap_tps=function(t){var r=t||{};if(e.source.HistMap.call(this,r),r.tps_points||r.tps_serial){var i={use_worker:!0,transform_callback:r.transform_callback,error_callback:r.error_callback,web_fallback:r.web_fallback,on_solved:r.on_solved,on_serialized:r.on_serialized};this.tps=new o(i),r.tps_points?this.tps.load_points(r.tps_points):this.tps.load_serial(r.tps_serial)}},e.inherits(e.source.HistMap_tps,e.source.HistMap),e.source.HistMap_tps.createAsync=function(o){return new Promise(function(t,r){var i;o.on_serialized=function(){if(o.tps_points){var e=document.createElement("a");document.body.appendChild(e),e.style="display: none";var r=new Blob([i.tps.serialize()]),n=window.URL.createObjectURL(r);e.href=n,e.download=o.tps_serial,e.click(),window.URL.revokeObjectURL(n)}t(i)},o.transform_callback=function(e,o,t){t.callback&&t.callback(e)},i=new e.source.HistMap_tps(o)})},e.source.HistMap_tps.prototype.xy2MercAsync_=function(e){var o=this;return new Promise(function(t,r){var i=o.histMapCoords2Xy(e);o.tps.transform(i,!1,{callback:function(e){t(e)}})})},e.source.HistMap_tps.prototype.merc2XyAsync_=function(e){var o=this;return new Promise(function(t,r){o.tps.transform(e,!0,{callback:function(e){var r=o.xy2HistMapCoords(e);t(r)}})})},e}),function(e){e(function(e){function o(o,t){var r=t.geometry.coordinates;return[0,1,2,3].map(function(i){var n=(i+1)%4,l=o[i],a=o[n],s=l.geometry.coordinates,p=Math.atan2(s[0]-r[0],s[1]-r[1]),g=e.tin(e.featureCollection([t,l,a]),"target");return[p,g]}).reduce(function(e,o){return e[0].push(o[0]),e[1].push(o[1]),e},[[],[]])}function t(e,o){for(var t=o?function(e){return!(e>=0&&e<2*Math.PI)}:function(e){return!(e>-1*Math.PI&&e<=Math.PI)};t(e);)e+=2*Math.PI*(e>0?-1:1);return e}function r(e,o){for(var r,i=t(e-o[0]),n=2*Math.PI,l=0;l<o.length;l++){var a=(l+1)%o.length,s=t(e-o[a]),p=Math.min(Math.abs(i),Math.abs(s));i*s<=0&&p<n&&(n=p,r=l),i=s}return r}function i(o,t){return e.point(o,{target:t})}function n(e,o){var t=o.geometry.coordinates[0][0],r=o.geometry.coordinates[0][1],i=o.geometry.coordinates[0][2],n=e.geometry.coordinates,l=o.properties.a,a=o.properties.b,s=o.properties.c,p=[r[0]-t[0],r[1]-t[1]],g=[i[0]-t[0],i[1]-t[1]],c=[n[0]-t[0],n[1]-t[1]],u=[a[0]-l[0],a[1]-l[1]],y=[s[0]-l[0],s[1]-l[1]],h=(g[1]*c[0]-g[0]*c[1])/(p[0]*g[1]-p[1]*g[0]),d=(p[0]*c[1]-p[1]*c[0])/(p[0]*g[1]-p[1]*g[0]),m=[h*u[0]+d*y[0]+l[0],h*u[1]+d*y[1]+l[1]];return m}function l(e,o,t){var i=e.geometry.coordinates,l=t.geometry.coordinates,a=Math.atan2(i[0]-l[0],i[1]-l[1]),s=r(a,o[0]),p=o[1][s];return n(e,p.features[0])}function a(o,t){for(var r=0;r<t.features.length;r++){var i=e.inside(o,t.features[r]);if(i)return t.features[r]}}function s(e,o,t,r){var i=a(e,o);return i?n(e,i):l(e,t,r)}var p=function(o){this.points=o.points,this.wh=o.wh,e.inside=function(o,t){var r=e.getCoord(o),i=t.geometry.coordinates;"Polygon"===t.geometry.type&&(i=[i]);for(var n=0,l=!1;n<i.length&&!l;n++)if(e.inRing(r,i[n][0])){for(var a=!1,s=1;s<i[n].length&&!a;)e.inRing(r,i[n][s],!0)&&(a=!0),s++;a||(l=!0)}return l},e.inRing=function(e,o,t){var r=!1;o[0][0]==o[o.length-1][0]&&o[0][1]==o[o.length-1][1]&&(o=o.slice(0,o.length-1));for(var i=0,n=o.length-1;i<o.length;n=i++){var l=o[i][0],a=o[i][1],s=o[n][0],p=o[n][1],g=e[1]*(l-s)+a*(s-e[0])+p*(e[0]-l)==0&&(l-e[0])*(s-e[0])<=0&&(a-e[1])*(p-e[1])<=0;if(g)return!t;var c=a>e[1]!=p>e[1]&&e[0]<(s-l)*(e[1]-a)/(p-a)+l;c&&(r=!r)}return r}};return p.setTurf=function(o){e=o},p.prototype.setPoints=function(e){this.points=e,this.for_tins=void 0,this.bak_tins=void 0},p.prototype.setWh=function(e){this.wh=e,this.for_tins=void 0,this.bak_tins=void 0},p.prototype.updateTin=function(){var t=[],r=this.points;this.wh&&(t=[[0,0],[this.wh[0],0],[0,this.wh[1]],[this.wh[0],this.wh[1]]]);for(var n=[],l=[],a=0;a<r.length;a++){var p=r[a][0],g=r[a][1];n.push(i(p,g)),l.push(i(g,p))}var c=e.featureCollection(n),u=e.featureCollection(l),y=e.tin(c,"target"),h=e.centroid(c),d=h.geometry.coordinates,m=s(h,y);this.for_centroid=i(d,m),this.bak_centroid=i(m,d);var f=e.convex(c).geometry.coordinates[0],_=f.reduce(function(o,t){var r=s(e.point(t),y),i=[t[0]-d[0],t[1]-d[1]],n=[r[0]-m[0],m[1]-r[1]];if(0==i[0]||0==i[1])return o;var l=0;return i[0]>0&&(l+=1),i[1]>0&&(l+=2),o[l].push([i,n]),o},[[],[],[],[]]).map(function(e){return e.reduce(function(e,o,t,r){e||(e=[0,0,0]);var i=e[0]+Math.sqrt(Math.pow(o[0][0],2)+Math.pow(o[0][1],2))/Math.sqrt(Math.pow(o[1][0],2)+Math.pow(o[1][1],2)),n=Math.atan2(o[0][0],o[0][1])-Math.atan2(o[1][0],o[1][1]),l=e[1]+Math.cos(n),a=e[2]+Math.sin(n);return t==r.length-1?[i/r.length,Math.atan2(a,l)]:[i,l,a]},null)}),E=_.map(function(e,o){var r=t[o],i=[r[0]-d[0],r[1]-d[1]],n=Math.sqrt(Math.pow(i[0],2)+Math.pow(i[1],2)),l=n/e[0],a=Math.atan2(i[0],i[1]),s=a-e[1],p=[m[0]+l*Math.sin(s),m[1]-l*Math.cos(s)];return[r,p]}),v=E[2];E[2]=E[3],E[3]=v;for(var S=[],P=[],a=0;a<E.length;a++){var x=(a+1)%E.length,T=E[a][0],R=E[a][1],M=[(T[0]+E[x][0][0])/2,(T[1]+E[x][0][1])/2],L=[(R[0]+E[x][1][0])/2,(R[1]+E[x][1][1])/2];c.features.push(i(T,R)),u.features.push(i(R,T)),S.push(i(T,R)),P.push(i(R,T)),c.features.push(i(M,L)),u.features.push(i(L,M))}this.for_points=c,this.bak_points=u,this.for_tins=e.tin(c,"target"),this.bak_tins=e.tin(u,"target");var C=o(S,this.for_centroid);this.for_vertices_params=C;var I=o(P,this.bak_centroid);this.bak_vertices_params=I},p.prototype.transform=function(o,t){this.bak_tins&&this.for_tins||this.updateTin();var r=e.point(o),i=t?this.bak_tins:this.for_tins,n=t?this.bak_vertices_params:this.for_vertices_params,l=t?this.bak_centroid:this.for_centroid;return s(r,i,n,l)},p})}("function"==typeof turf?function(e){this.Tin=e(this.turf)}:"function"==typeof define&&define.amd?function(e){define("tin",["turf"],function(o){return e(o)})}:"undefined"!=typeof module&&module.exports&&"function"==typeof require?function(e){var o;try{o=require("turf")}catch(e){}module.exports=e(o)}:function(e){if(!this.turf)throw new Error('"turf" not found');this.Tin=e(this.turf)}),define("histmap_tin",["histmap","tin"],function(e,o){e.source.HistMap_tin=function(t){var r=t||{};e.source.HistMap.call(this,r),this.tin=new o({wh:[this.width,this.height]})},e.inherits(e.source.HistMap_tin,e.source.HistMap),e.source.HistMap_tin.createAsync=function(o){return new Promise(function(t,r){var i,n=o.tin_points_url,l=new XMLHttpRequest;l.open("GET",n,!0),l.responseType="json",l.onload=function(e){if(200==this.status){var o=this.response;i.tin.setPoints(o),i.tin.updateTin(),t(i)}},l.send(),i=new e.source.HistMap_tin(o)})},e.source.HistMap_tin.prototype.xy2MercAsync_=function(e){var o=this;return new Promise(function(t,r){var i=o.histMapCoords2Xy(e),n=o.tin.transform(i,!1);t(n)})},e.source.HistMap_tin.prototype.merc2XyAsync_=function(e){var o=this;return new Promise(function(t,r){var i=o.tin.transform(e,!0),n=o.xy2HistMapCoords(i);t(n)})}}),function(e,o,t){"function"==typeof define&&"undefined"!=typeof define.amd?define(o,function(){return t(e)}):"object"==typeof module&&"object"==typeof module.exports?module.exports=t(e):e[o]=t(e)}(this,"sprintf",function(){var e=function(){if(!arguments||arguments.length<1)throw new Error("sprintf: ERROR: not enough arguments");for(var e,o=0,t="",r=arguments[o++];e=/^([^%]*)%(?:(\d+)\$|\((.*?)\))?([#0 +'-]+)?(\*|\d+)?(\.\*|\.\d+)?([%diouxXfFeEcs])((?:.|[\r\n])*)$/.exec(r);){var i,n=e[1],l=e[2],a=e[3],s=e[4],p=e[5],g=e[6],c=e[7],u=e[8];if("%"===c)i="%";else{var y=" ",h=!0;s?(s.indexOf("0")>=0&&(y="0"),s.indexOf("-")>=0&&(y=" ",h=!1)):s="";var d,m=-1;if(p)if("*"===p){if(d=o++,d>=arguments.length)throw new Error("sprintf: ERROR: not enough arguments");m=arguments[d]}else m=parseInt(p,10);var f=-1;if(g)if(".*"===g){if(d=o++,d>=arguments.length)throw new Error("sprintf: ERROR: not enough arguments");f=arguments[d]}else f=parseInt(g.substring(1),10);if(d=o++,l){if(d=parseInt(l,10),d>=arguments.length)throw new Error("sprintf: ERROR: not enough arguments");i=arguments[d]}else if(a){if("object"!=typeof arguments[1])throw new Error("sprintf: ERROR: invalid non-object arguments for named argument");if(i=arguments[1][a],"undefined"==typeof i)throw new Error("sprintf: ERROR: invalid undefined value for named argument")}else{if(d>=arguments.length)throw new Error("sprintf: ERROR: not enough arguments");i=arguments[d]}var _="";switch(c){case"d":case"i":"number"!=typeof i&&(i=0),i=i.toString(10),s.indexOf("#")>=0&&i>=0&&(i="+"+i),s.indexOf(" ")>=0&&i>=0&&(i=" "+i);break;case"b":"number"!=typeof i&&(i=0),i=i.toString(2);break;case"o":"number"!=typeof i&&(i=0),i=i.toString(8);break;case"u":"number"!=typeof i&&(i=0),i=Math.abs(i),i=i.toString(10);break;case"x":"number"!=typeof i&&(i=0),i=i.toString(16).toLowerCase(),s.indexOf("#")>=0&&(_="0x");break;case"X":"number"!=typeof i&&(i=0),i=i.toString(16).toUpperCase(),s.indexOf("#")>=0&&(_="0X");break;case"f":case"F":if("number"!=typeof i&&(i=0),i=0+i,f>-1&&(i.toFixed?i=i.toFixed(f):(i=Math.round(i*Math.pow(10,f))/Math.pow(10,f),i+="0000000000",i=i.substr(0,i.indexOf(".")+f+1))),i=""+i,s.indexOf("'")>=0)for(var E=0,v=i.length-1-3;v>=0;v-=3)i=i.substring(0,v)+(0===E?".":",")+i.substring(v),E=(E+1)%2;break;case"e":case"E":if("number"!=typeof i&&(i=0),i=0+i,f>-1){if(!i.toExponential)throw new Error("sprintf: ERROR: toExponential() method not supported");i=i.toExponential(f)}i=""+i,"E"===c&&(i=i.replace(/e\+/,"E+"));break;case"c":"number"!=typeof i&&(i=0),i=String.fromCharCode(i);break;case"s":"string"!=typeof i&&(i=String(i)),f>-1&&(i=i.substr(0,f));break;default:throw new Error('sprintf: ERROR: invalid conversion character "'+c+'"')}var S=m-i.toString().length-_.toString().length;if(S>0){var P=new Array(S+1);h?i=P.join(y)+i:i+=P.join(y)}i=_+i}t=t+n+i,r=u}return t+r};return e.version={text:"1.2.3",major:1,minor:2,micro:3},e}),define("sprintf",function(){}),!function(e,o){"object"==typeof exports&&"undefined"!=typeof module?module.exports=o():"function"==typeof define&&define.amd?define("i18nxhr",o):e.i18nextXHRBackend=o()}(this,function(){"use strict";function e(e){return i.call(n.call(arguments,1),function(o){if(o)for(var t in o)void 0===e[t]&&(e[t]=o[t])}),e}function o(e,o,t,r,i){if(r&&"object"===("undefined"==typeof r?"undefined":l(r))){
+var n="",a=encodeURIComponent;for(var s in r)n+="&"+a(s)+"="+a(r[s]);r=n.slice(1)+(i?"":"&_t="+new Date)}try{var p=new(XMLHttpRequest||ActiveXObject)("MSXML2.XMLHTTP.3.0");p.open(r?"POST":"GET",e,1),o.crossDomain||p.setRequestHeader("X-Requested-With","XMLHttpRequest"),p.withCredentials=!!o.withCredentials,r&&p.setRequestHeader("Content-type","application/x-www-form-urlencoded"),p.onreadystatechange=function(){p.readyState>3&&t&&t(p.responseText,p)},p.send(r)}catch(e){window.console&&console.log(e)}}function t(){return{loadPath:"/locales/{{lng}}/{{ns}}.json",addPath:"locales/add/{{lng}}/{{ns}}",allowMultiLoading:!1,parse:JSON.parse,crossDomain:!1,ajax:o}}var r=[],i=r.forEach,n=r.slice,l="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e},a=(function(){function e(e){this.value=e}function o(o){function t(e,o){return new Promise(function(t,i){var a={key:e,arg:o,resolve:t,reject:i,next:null};l?l=l.next=a:(n=l=a,r(e,o))})}function r(t,n){try{var l=o[t](n),a=l.value;a instanceof e?Promise.resolve(a.value).then(function(e){r("next",e)},function(e){r("throw",e)}):i(l.done?"return":"normal",l.value)}catch(e){i("throw",e)}}function i(e,o){switch(e){case"return":n.resolve({value:o,done:!0});break;case"throw":n.reject(o);break;default:n.resolve({value:o,done:!1})}n=n.next,n?r(n.key,n.arg):l=null}var n,l;this._invoke=t,"function"!=typeof o.return&&(this.return=void 0)}return"function"==typeof Symbol&&Symbol.asyncIterator&&(o.prototype[Symbol.asyncIterator]=function(){return this}),o.prototype.next=function(e){return this._invoke("next",e)},o.prototype.throw=function(e){return this._invoke("throw",e)},o.prototype.return=function(e){return this._invoke("return",e)},{wrap:function(e){return function(){return new o(e.apply(this,arguments))}},await:function(o){return new e(o)}}}(),function(e,o){if(!(e instanceof o))throw new TypeError("Cannot call a class as a function")}),s=function(){function e(e,o){for(var t=0;t<o.length;t++){var r=o[t];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}return function(o,t,r){return t&&e(o.prototype,t),r&&e(o,r),o}}(),p=function(){function o(e){var t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{};a(this,o),this.init(e,t),this.type="backend"}return s(o,[{key:"init",value:function(o){var r=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{};this.services=o,this.options=e(r,this.options||{},t())}},{key:"readMulti",value:function(e,o,t){var r=this.options.loadPath;"function"==typeof this.options.loadPath&&(r=this.options.loadPath(e,o));var i=this.services.interpolator.interpolate(r,{lng:e.join("+"),ns:o.join("+")});this.loadUrl(i,t)}},{key:"read",value:function(e,o,t){var r=this.options.loadPath;"function"==typeof this.options.loadPath&&(r=this.options.loadPath([e],[o]));var i=this.services.interpolator.interpolate(r,{lng:e,ns:o});this.loadUrl(i,t)}},{key:"loadUrl",value:function(e,o){var t=this;this.options.ajax(e,this.options,function(r,i){if(i.status>=500&&i.status<600)return o("failed loading "+e,!0);if(i.status>=400&&i.status<500)return o("failed loading "+e,!1);var n=void 0,l=void 0;try{n=t.options.parse(r,e)}catch(o){l="failed parsing "+e+" to json"}return l?o(l,!1):void o(null,n)})}},{key:"create",value:function(e,o,t,r){var i=this;"string"==typeof e&&(e=[e]);var n={};n[t]=r||"",e.forEach(function(e){var t=i.services.interpolator.interpolate(i.options.addPath,{lng:e,ns:o});i.options.ajax(t,i.options,function(e,o){},n)})}}]),o}();return p.type="backend",p}),define("app",["jquery","ol-custom","sprintf","i18n","i18nxhr","ji18n","bootstrap","slick"],function(e,o,t,r,i,n){return e.fn.nodoubletapzoom=function(){e(this).bind("touchstart",function(o){var t=o.timeStamp,r=e(this).data("lastTouch")||t,i=t-r,n=o.originalEvent.touches.length;e(this).data("lastTouch",t),!i||i>500||n>1||(o.preventDefault(),e(o.target).trigger("click"))})},e("body").nodoubletapzoom(),function(l){var a=l.appid||"sample",s="json/"+a+".json",p=l.debug?function(e){console.log(e)}:function(){},g=l.lang||"ja",c=l.overlay||!1;c&&(e("body").addClass("with-opacity"),e(".opacity-slider").removeClass("hide"),e(".opacity-slider input").prop("disabled",!0)),r.use(i).init({lng:g,backend:{loadPath:"locales/{{lng}}/{{ns}}.json"}},function(i,a){n.init(r,e),e("body").localize(),e.get(s,function(r){e("#all").show(),e("#loadWait").modal(),e(".slick-class").slick({prevArrow:"",nextArrow:"",centerMode:!0,focusOnSelect:!0,slidesToScroll:3,centerPadding:"40px"});var i,n=null,s=r.home_position,g=r.default_zoom,u=r.app_name,y=!!l.fake&&r.fake_gps,h=!!l.fake&&r.fake_center,d=!!l.fake&&r.fake_radius,m=r.make_binary,f=null,_=null,E=null,v="map_div",S=null;c&&(e('<div id="map_div_back" class="map h100p w100p" style="top:0px; left:0px; width: 100%; position:absolute;"></div>').insertBefore("#center_circle"),e('<div id="map_div_front" class="map h100p w100p" style="top:0px; left:0px; width: 100%; position:absolute;"></div>').insertBefore("#center_circle"),v="map_div_front",S="map_div_back",E=new o.MaplatMap({off_control:!0,div:S})),y?e("#gps_etc").append(t(a("app.fake_explanation"),h,d)):e("#gps_etc").append(a("app.acquiring_gps_desc"));var P=r.pois;e("title").html(u);for(var x=r.sources,T={},R=[],M={map_option:{div:v},home_position:s,merc_zoom:g,fake_gps:!!y&&d},L=0;L<=x.length;L++)if(L==x.length)R.push(o.source.NowMap.createAsync(Object.assign({sourceID:"osm"},M))),e(".slick-class").slick("slickAdd",'<div class="slick-item" data="osm"><img src="./tmbs/osm_menu.jpg"><div>'+a("app.osm_now")+"</div></div>"),e(".slick-class").slick("slickGoTo",x.length);else{var C=x[L];C.maptype||(C.maptype="maplat"),C.algorythm||(C.algorythm=l.algorythm||"tin"),function(t){if("base"==t.maptype||"overlay"==t.maptype){t.sourceID=t.mapID;var r="base"==t.maptype?o.source.NowMap:o.source.TmsMap;R.push(r.createAsync(Object.assign({attributions:[new o.Attribution({html:t.attr})],url:t.url,sourceID:t.sourceID},M)))}else t.sourceID=t.mapID+":"+t.maptype+":"+t.algorythm,R.push(new Promise(function(e,r){var i=function(){T[t.sourceID]=t;var r=Object.assign({attributions:[new o.Attribution({html:t.attr})],mapID:t.mapID,width:t.width,height:t.height,maptype:t.maptype,algorythm:t.algorythm,sourceID:t.sourceID},M);"tin"==t.algorythm?r.tin_points_url="json/"+t.mapID+"_points.json":m?(r.tps_serial=t.mapID+".bin",r.tps_points="../json/"+t.mapID+"_points.json"):r.tps_serial="../bin/"+t.mapID+".bin",e(o.source.HistMap.createAsync(r))};"maplat"==t.maptype?require(["histmap_"+t.algorythm],i):i()}));e(".slick-class").slick("slickAdd",'<div class="slick-item" data="'+t.sourceID+'"><img src="./tmbs/'+t.mapID+'_menu.jpg"><div>'+(t.label||t.year)+"</div></div>")}(C)}Promise.all(R).then(function(t){function r(e,t){var r=_.getView(),l=i.size2MercsAsync();if(n&&n.mercs&&n.buffer[i.sourceID]){var a=n.buffer[i.sourceID],s=o.MathEx.recursiveRound([r.getCenter(),r.getZoom(),r.getRotation()],10);a[0][0]==s[0][0]&&a[0][1]==s[0][1]&&a[1]==s[1]&&a[2]==s[2]?(p("From: Use buffer"),l=new Promise(function(e,o){e(n.mercs)})):n={buffer:{}}}else n={buffer:{}};l.then(function(r){n.mercs=r;var l=_.getView();p("From: Center: "+l.getCenter()+" Zoom: "+l.getZoom()+" Rotation: "+l.getRotation()),n.buffer[i.sourceID]=o.MathEx.recursiveRound([l.getCenter(),l.getZoom(),l.getRotation()],10),p("Mercs: "+r);var a=e.mercs2SizeAsync(r),s=e.sourceID;n.buffer[s]&&(p("To: Use buffer"),a=new Promise(function(e,o){e(n.buffer[s])})),a.then(function(r){p("To: Center: "+[r[0][0],r[0][1]]+" Zoom: "+r[1]+" Rotation: "+r[2]),n.buffer[e.sourceID]=o.MathEx.recursiveRound(r,10),t(r)})})}function l(t,n){var l=c.osm,s=c[n];s==i&&s!=l||s!=i&&r(s,function(n){var p=null,g=null;E&&(p=E.getSource(),s instanceof o.source.NowMap?s instanceof o.source.NowMap&&E.exchangeSource():p?g=p:(g=l,i instanceof o.source.NowMap&&(g=i instanceof o.source.TmsMap?_.getSource():i),E.exchangeSource(g)),!(s instanceof o.source.NowMap)||s instanceof o.source.TmsMap?e(".opacity-slider input").removeProp("disabled"):e(".opacity-slider input").prop("disabled",!0)),s instanceof o.source.TmsMap?(_.setLayer(s),i instanceof o.source.NowMap||_.exchangeSource(p||l),e(".opacity-slider input").removeProp("disabled")):(_.setLayer(),_.exchangeSource(s)),_.setOpacity(e(".opacity-slider input").val());var c=_.getView();s.insideCheckHistMapCoords(n[0])?(c.setCenter(n[0]),c.setZoom(n[1]),c.setRotation(n[2])):(e("#gpsDialogTitle").text(a("app.out_of_map")),e("#gpsDialogBody").text(a("app.out_of_map_area")),e("#gpsDialog").modal(),s.goHome()),s.setGPSMarker(f,!0),_.resetMarker();for(var u=0;u<P.length;u++)!function(e){var t=[e.lng,e.lat],r=o.proj.transform(t,"EPSG:4326","EPSG:3857");s.merc2XyAsync(r).then(function(o){s.insideCheckHistMapCoords(o)&&_.setMarker(o,{datum:e})})}(P[u]);_.updateSize(),_.renderSync(),i=s,1==t&&s.goHome(),E&&g&&r(g,function(e){var o=E.getView();o.setCenter(e[0]),o.setZoom(e[1]),o.setRotation(e[2]),E.updateSize(),E.renderSync()})})}function s(o){e("#poi_name").text(o.name),e("#poi_img").attr("src","img/"+o.image),e("#poi_address").text(o.address),e("#poi_desc").html(o.desc.replace(/\n/g,"<br>")),e("#poi_info").modal()}e("#loadWait").modal("hide");for(var g=[],c={},u=!1,y=0;y<t.length;y++){var h=t[y];_||h instanceof o.source.TmsMap||(_=h.getMap(),_.on("gps_request",function(){e("#gpsWait").modal()}),_.on("gps_result",function(o){var t=(e("#gpsWait").data("bs.modal")||{isShown:!1}).isShown,r=o.frameState;r&&r.error?(f=null,"gps_out"==r.error&&t&&(e("#gpsWait").modal("hide"),e("#gpsDialogTitle").text(a("app.out_of_map")),e("#gpsDialogBody").text(a("app.out_of_map_desc")),e("#gpsDialog").modal())):f=r,t&&e("#gpsWait").modal("hide")})),h._map=_,g.push(h),c[h.sourceID]=h}e(".slick-item").on("click",function(){u||l(!1,e(this).attr("data"))}),e(".slick-class").on("beforeChange",function(e,o,t,r){u=t!=r}),e(".slick-class").on("afterChange",function(e,o,t){u=!1}),i=g.reduce(function(e,t){return e?e:t instanceof o.source.HistMap?t:e},null),l(!0,"osm");var d=function(e){return function(o){var t=e.forEachFeatureAtPixel(o.pixel,function(e){if(e.get("datum"))return e});t&&s(t.get("datum"))}}(_);_.on("click",d);var m=function(o){return function(t){var r=o.getEventPixel(t.originalEvent),i=o.hasFeatureAtPixel(r),n=o.getTarget();if(i){var l=o.forEachFeatureAtPixel(t.pixel,function(e){if(e.get("datum"))return e});return void e("#"+n).css("cursor",l?"pointer":"")}e("#"+n).css("cursor","")}}(_);_.on("pointermove",m);var v=function(e){return function(o){var t=o.frameState.viewState.center,r=e.getSource();if(!r.insideCheckHistMapCoords(t)){var i=r.histMapCoords2Xy(t);p(i);var n=i[0]/(r.width/2)-1,l=i[1]/(r.height/2)-1,a=Math.max(Math.abs(n),Math.abs(l));i=[(n/a+1)*r.width/2,(l/a+1)*r.height/2],p(i),t=r.xy2HistMapCoords(i),e.getView().setCenter(t)}}}(_);_.on("moveend",v);var S=function(e){return function(o){if(E){if(e._backMapMoving)return void p("Backmap moving skipped");var t=E.getSource();t&&(e._backMapMoving=!0,p("Backmap moving started"),r(t,function(o){var t=E.getView();t.setCenter(o[0]),t.setZoom(o[1]),t.setRotation(o[2]),p("Backmap moving ended"),e._backMapMoving=!1}))}}}(_);_.on("postrender",S);var x=function(){_.setOpacity(e(".opacity-slider input").val())};e(".opacity-slider input").on("input",x),e(".opacity-slider input").on("change",x)})},"json")})}});
