+++ js/histmap_tin.js
@@ -1,4 +1,4 @@
-define(['histmap', 'tin2'], function(ol, Tin) {
+define(['histmap', 'tin'], function(ol, Tin) {
     ol.source.HistMap_tin = function(optOptions) {
         var options = optOptions || {};
