--- js/the_result
@@ -1,1 +1,0 @@
-false
