--- js/tinTest.js
@@ -1,52 +1,0 @@
-var turf = require('./turf.min');
-var Tin = require('./tin2');
-Tin.setTurf(turf);
-var points = require('../json/morioka_points.json');
-
-// for turf inside patch
-
-turf.inside = function input(point, polygon) {
-    var pt = turf.getCoord(point);
-    var polys = polygon.geometry.coordinates;
-    // normalize to multipolygon
-    if (polygon.geometry.type === 'Polygon') polys = [polys];
-
-    for (var i = 0, insidePoly = false; i < polys.length && !insidePoly; i++) {
-        // check if it is in the outer ring first
-        if (turf.inRing(pt, polys[i][0])) {
-            var inHole = false;
-            var k = 1;
-            // check for the point in any of the holes
-            while (k < polys[i].length && !inHole) {
-                if (turf.inRing(pt, polys[i][k], true)) {
-                    inHole = true;
-                }
-                k++;
-            }
-            if (!inHole) insidePoly = true;
-        }
-    }
-    return insidePoly;
-};
-
-// pt is [x,y] and ring is [[x,y], [x,y],..]
-turf.inRing = function(pt, ring, ignoreBoundary) {
-    var isInside = false;
-    if (ring[0][0] == ring[ring.length-1][0] && ring[0][1] == ring[ring.length-1][1]) ring = ring.slice(0, ring.length-1);
-
-    for (var i = 0, j = ring.length - 1; i < ring.length; j = i++) {
-        var xi = ring[i][0], yi = ring[i][1];
-        var xj = ring[j][0], yj = ring[j][1];
-        var onBoundary = (pt[1] * (xi - xj) + yi * (xj - pt[0]) + yj * (pt[0] - xi) == 0) &&
-            ((xi - pt[0]) * (xj - pt[0]) <= 0) && ((yi - pt[1]) * (yj - pt[1]) <= 0);
-        if (onBoundary) return !ignoreBoundary;
-        var intersect = ((yi > pt[1]) !== (yj > pt[1])) &&
-            (pt[0] < (xj - xi) * (pt[1] - yi) / (yj - yi) + xi);
-        if (intersect) isInside = !isInside;
-    }
-    return isInside;
-}
-
-var tin = new Tin({points: points, wh: [6144, 4096]});
-tin.updateTin2();
-
